Migrated old icons to visyn icons
