+++ package.json
@@ -121,6 +121,7 @@
     "jszip": "^3.10.1",
     "lineupjs": "4.14.1",
     "lodash": "~4.17.21",
+    "lucide-react": "^0.526.0",
     "plotly.js-dist-min": "~2.12.1",
     "react": "~18.3.1",
     "react-dom": "~18.3.1",
