--- src/icons/EntityIcons.tsx
@@ -1,70 +1,0 @@
-import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
-
-export const dvGene: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'gene' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M228.204 80.9721C228.213 80.9585 228.222 80.9445 228.229 80.9301C246.828 45.4255 282.306 34.3055 301.536 31.4207C312.186 29.8229 322.116 37.1615 323.713 47.8119C325.311 58.4622 317.972 68.3913 307.322 69.9891C293.296 72.0932 272.569 79.6979 262.45 99.6603C262.086 100.379 261.678 101.073 261.228 101.741C252.441 114.779 249.762 130.761 256.009 163.692C256.01 163.699 256.011 163.705 256.013 163.712L275.316 247.158L275.316 247.159L275.338 247.255L275.338 247.256L295.521 336.542C295.569 336.755 295.614 336.969 295.655 337.183C302.871 374.964 300.048 405.037 282.969 430.38C258.09 467.3 221.197 476.095 198.581 480.444C188.006 482.477 177.783 475.553 175.75 464.977C173.716 454.401 180.641 444.179 191.216 442.145C212.085 438.132 235.032 431.728 250.628 408.586C259.744 395.057 263.488 376.906 257.409 344.823C257.408 344.818 257.407 344.812 257.405 344.806L237.32 255.951L237.32 255.949L217.928 172.119L217.441 172.232L217.928 172.119C217.871 171.874 217.819 171.629 217.772 171.382C210.821 134.988 211.461 106.537 228.204 80.9721ZM432.267 232.301C432.275 232.293 432.282 232.285 432.289 232.277C441.328 221.871 446.911 208.226 450.041 194.802C453.171 181.376 453.858 168.131 453.075 158.487C452.205 147.752 460.201 138.345 470.935 137.474C481.669 136.603 491.077 144.6 491.948 155.334C493.916 179.607 489.579 226.317 461.206 258.453C460.998 258.688 460.785 258.919 460.566 259.144C428.043 292.577 381.291 297.775 318.702 298.334C307.933 298.43 299.125 289.778 299.029 279.008C298.933 268.239 307.585 259.431 318.354 259.335C381.941 258.768 412.369 252.471 432.267 232.301ZM84.3441 275.776C84.3255 275.795 84.3083 275.815 84.2928 275.836C67.0992 299.383 59.4295 320.525 58.6115 334.669C57.9897 345.421 48.7697 353.633 38.0181 353.011C27.2666 352.389 19.0547 343.169 19.6766 332.418C21.0938 307.912 33.3035 279.145 53.7779 251.502C54.2969 250.801 54.8623 250.136 55.4703 249.511C87.8997 216.173 131.814 210.793 194.725 210.232C205.494 210.136 214.302 218.788 214.398 229.557C214.494 240.326 205.842 249.134 195.073 249.23C132.411 249.789 104.273 255.967 84.3441 275.776ZM286.443 100.704C293.386 92.4707 305.688 91.4244 313.921 98.3668L422.956 190.308C431.189 197.25 432.235 209.552 425.293 217.786C418.35 226.019 406.048 227.065 397.815 220.123L288.78 128.182C280.547 121.239 279.501 108.937 286.443 100.704ZM280.095 173.91C287.037 165.677 299.34 164.631 307.573 171.573L354.309 210.983C362.543 217.925 363.589 230.227 356.647 238.461C349.704 246.694 337.402 247.74 329.169 240.798L282.432 201.388C274.199 194.446 273.152 182.143 280.095 173.91ZM156.781 270.105C163.723 261.872 176.025 260.825 184.259 267.768L230.995 307.177C239.228 314.12 240.275 326.422 233.332 334.655C226.39 342.889 214.088 343.935 205.854 336.993L159.118 297.583C150.885 290.641 149.838 278.338 156.781 270.105ZM88.1342 290.78C95.0767 282.547 107.379 281.5 115.612 288.443L224.647 380.384C232.88 387.326 233.926 399.628 226.984 407.861C220.041 416.095 207.739 417.141 199.506 410.199L90.4714 318.258C82.2381 311.315 81.1918 299.013 88.1342 290.78Z',
-  ],
-};
-
-export const dvDrug: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'drug' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M303.388 440.368C369.935 478.789 454.954 456.008 493.375 389.461C505.935 367.706 511.917 343.958 512 320.519C512.044 308.098 499.777 300.313 487.903 303.495L265.143 363.183C253.226 366.44 246.613 379.284 252.787 389.975C264.578 410.233 281.632 427.807 303.388 440.368Z
-    M252.481 250.381C239.92 272.137 233.938 295.884 233.856 319.324C233.811 331.744 246.078 339.529 257.953 336.348L480.712 276.659C492.63 273.402 499.242 260.558 493.069 249.868C481.278 229.61 464.223 212.035 442.468 199.474C375.921 161.054 290.902 183.834 252.481 250.381Z
-    M0 356.716C0 347.781 1.15243 338.697 3.5652 329.652L57.2404 128.43C72.0729 72.8249 128.881 39.8786 184.237 54.7781C239.593 69.6775 272.391 126.742 257.559 182.348C242.726 237.953 203.884 383.569 203.884 383.569C189.051 439.175 132.243 472.121 76.8868 457.222C30.5356 444.746 0 402.707 0 356.716ZM168.901 112.27C145.203 105.891 120.824 120.03 114.474 143.835L87.6366 244.446L173.487 267.554L200.325 166.943C206.675 143.137 192.6 118.649 168.901 112.27Z
-    `,
-  ],
-};
-
-export const dvCellLine: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'cellline' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M469.975 290.738V289.646L469.148 290.36C457.706 300.24 444.632 308.85 430.816 316.134C385.003 340.288 323.174 354.595 256 354.595C188.826 354.595 126.997 340.288 81.1842 316.134C67.3678 308.85 54.2937 300.24 42.8517 290.36L42.0249 289.646V290.738V304.452C42.0249 324.297 58.0639 349.388 98.8468 370.891C137.992 391.529 193.532 404.927 256 404.927C318.468 404.927 374.008 391.529 413.153 370.891C453.936 349.388 469.975 324.297 469.975 304.452V290.738ZM511.5 207.548V304.452C511.5 352.956 474.996 389.746 430.816 413.039C385.003 437.193 323.174 451.5 256 451.5C188.826 451.5 126.997 437.193 81.1842 413.039C37.0036 389.746 0.5 352.956 0.5 304.452V207.548C0.5 159.044 37.0036 122.254 81.1842 98.9609C126.997 74.807 188.826 60.5 256 60.5C323.174 60.5 385.003 74.807 430.816 98.9609C474.996 122.254 511.5 159.044 511.5 207.548ZM98.8468 141.109C58.0639 162.612 42.0249 187.703 42.0249 207.548C42.0249 227.392 58.0639 252.484 98.8468 273.986C137.992 294.624 193.532 308.022 256 308.022C318.468 308.022 374.008 294.624 413.153 273.986C453.936 252.484 469.975 227.392 469.975 207.548C469.975 187.703 453.936 162.612 413.153 141.109C374.008 120.471 318.468 107.073 256 107.073C193.532 107.073 137.992 120.471 98.8468 141.109ZM256 176.529C191.033 176.529 132.457 196.762 91.7205 220.164C83.6487 224.801 73.7058 221.268 69.5222 212.154C65.3273 203.015 68.5683 191.835 76.6782 187.176C121.514 161.419 185.21 139.471 256 139.471C326.79 139.471 390.486 161.419 435.322 187.176C443.432 191.835 446.673 203.015 442.478 212.154C438.294 221.268 428.351 224.801 420.28 220.164C379.543 196.762 320.967 176.529 256 176.529ZM186.61 234.117C186.61 237.305 185.564 241.255 183.783 245.408C182.006 249.554 179.512 253.868 176.651 257.777C173.79 261.687 170.573 265.175 167.358 267.681C164.13 270.197 160.964 271.675 158.193 271.675C152.586 271.675 147.402 267.299 143.583 261.139C139.778 255.001 137.43 247.241 137.43 240.777C137.43 234.318 139.774 227.314 143.567 221.931C147.365 216.54 152.548 212.862 158.193 212.862C163.96 212.862 171.094 214.984 176.78 218.731C182.47 222.48 186.61 227.785 186.61 234.117ZM262.304 272.175C262.304 278.801 258.341 284.857 252.86 289.278C247.38 293.699 240.471 296.413 234.738 296.413C229.01 296.413 222.531 293.704 217.48 289.291C212.432 284.88 208.872 278.822 208.872 272.175C208.872 265.501 212.456 257.761 217.535 251.673C222.626 245.57 229.09 241.277 234.738 241.277C237.553 241.277 240.687 242.343 243.823 244.192C246.954 246.039 250.057 248.648 252.803 251.688C258.311 257.783 262.304 265.518 262.304 272.175ZM363.513 234.117C363.513 240.472 359.559 246.288 354.077 250.541C348.598 254.792 341.686 257.403 335.947 257.403C330.198 257.403 322.426 254.785 316.088 250.521C309.735 246.248 304.978 240.432 304.978 234.117C304.978 231.04 306.78 228.165 309.799 225.553C312.812 222.947 316.972 220.663 321.529 218.775C330.647 214.997 341.212 212.862 347.003 212.862C352.737 212.862 356.83 214.957 359.509 218.609C362.206 222.286 363.513 227.603 363.513 234.117ZM469.975 290.738V289.646L469.148 290.36C457.706 300.24 444.632 308.85 430.816 316.134C385.003 340.288 323.174 354.595 256 354.595C188.826 354.595 126.997 340.288 81.1842 316.134C67.3678 308.85 54.2937 300.24 42.8517 290.36L42.0249 289.646V290.738V304.452C42.0249 324.297 58.0639 349.388 98.8468 370.891C137.992 391.529 193.532 404.927 256 404.927C318.468 404.927 374.008 391.529 413.153 370.891C453.936 349.388 469.975 324.297 469.975 304.452V290.738ZM511.5 207.548V304.452C511.5 352.956 474.996 389.746 430.816 413.039C385.003 437.193 323.174 451.5 256 451.5C188.826 451.5 126.997 437.193 81.1842 413.039C37.0036 389.746 0.5 352.956 0.5 304.452V207.548C0.5 159.044 37.0036 122.254 81.1842 98.9609C126.997 74.807 188.826 60.5 256 60.5C323.174 60.5 385.003 74.807 430.816 98.9609C474.996 122.254 511.5 159.044 511.5 207.548ZM98.8468 141.109C58.0639 162.612 42.0249 187.703 42.0249 207.548C42.0249 227.392 58.0639 252.484 98.8468 273.986C137.992 294.624 193.532 308.022 256 308.022C318.468 308.022 374.008 294.624 413.153 273.986C453.936 252.484 469.975 227.392 469.975 207.548C469.975 187.703 453.936 162.612 413.153 141.109C374.008 120.471 318.468 107.073 256 107.073C193.532 107.073 137.992 120.471 98.8468 141.109ZM256 176.529C191.033 176.529 132.457 196.762 91.7205 220.164C83.6487 224.801 73.7058 221.268 69.5222 212.154C65.3273 203.015 68.5683 191.835 76.6782 187.176C121.514 161.419 185.21 139.471 256 139.471C326.79 139.471 390.486 161.419 435.322 187.176C443.432 191.835 446.673 203.015 442.478 212.154C438.294 221.268 428.351 224.801 420.28 220.164C379.543 196.762 320.967 176.529 256 176.529ZM186.61 234.117C186.61 237.305 185.564 241.255 183.783 245.408C182.006 249.554 179.512 253.868 176.651 257.777C173.79 261.687 170.573 265.175 167.358 267.681C164.13 270.197 160.964 271.675 158.193 271.675C152.586 271.675 147.402 267.299 143.583 261.139C139.778 255.001 137.43 247.241 137.43 240.777C137.43 234.318 139.774 227.314 143.567 221.931C147.365 216.54 152.548 212.862 158.193 212.862C163.96 212.862 171.094 214.984 176.78 218.731C182.47 222.48 186.61 227.785 186.61 234.117ZM262.304 272.175C262.304 278.801 258.341 284.857 252.86 289.278C247.38 293.699 240.471 296.413 234.738 296.413C229.01 296.413 222.531 293.704 217.48 289.291C212.432 284.88 208.872 278.822 208.872 272.175C208.872 265.501 212.456 257.761 217.535 251.673C222.626 245.57 229.09 241.277 234.738 241.277C237.553 241.277 240.687 242.343 243.823 244.192C246.954 246.039 250.057 248.648 252.803 251.688C258.311 257.783 262.304 265.518 262.304 272.175ZM363.513 234.117C363.513 240.472 359.559 246.288 354.077 250.541C348.598 254.792 341.686 257.403 335.947 257.403C330.198 257.403 322.426 254.785 316.088 250.521C309.735 246.248 304.978 240.432 304.978 234.117C304.978 231.04 306.78 228.165 309.799 225.553C312.812 222.947 316.972 220.663 321.529 218.775C330.647 214.997 341.212 212.862 347.003 212.862C352.737 212.862 356.83 214.957 359.509 218.609C362.206 222.286 363.513 227.603 363.513 234.117Z`,
-  ],
-};
-
-export const dvTissue: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'tissue' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M236.3 61.4L226.9 75.5C205.6 107.5 167.5 124 129.6 117.6L70.0002 107.7C33.3002 101.6 -0.099776 129.9 0.000223981 167.1C0.000223981 183 6.40022 198.3 17.6002 209.6L46.8002 238.8C57.8002 249.8 64.0002 264.7 64.0002 280.3C64.0002 296.1 57.6002 311.2 46.3002 322.3L33.2002 335.1C22.1002 345.9 15.9002 360.7 15.9002 376.2C15.9002 413 50.0002 440.4 86.0002 432.4L148.3 418.6C156 416.9 164 416 171.9 416H181.9C209.1 416 235.6 425.3 256.9 442.3L287.7 467C298.2 475.4 311.3 480 324.7 480C357.4 480 384 453.5 384 420.7V395.5C384 360.6 405.4 329.3 437.9 316.7L474.8 302.4C497.2 293.7 512 272.1 512 248.1C512 220 491.9 195.8 464.2 190.8L436.2 185.7C399.7 179 370.8 151.2 362.6 115L355.5 83.5C348.8 53.4 322 32 291.2 32C269.2 32 248.6 43 236.3 61.4ZM159.9 192C168.387 192 176.526 195.371 182.528 201.373C188.529 207.374 191.9 215.513 191.9 224C191.9 232.487 188.529 240.626 182.528 246.627C176.526 252.628 168.387 256 159.9 256C151.413 256 143.274 252.628 137.273 246.627C131.272 240.626 127.9 232.487 127.9 224C127.9 215.513 131.272 207.374 137.273 201.373C143.274 195.371 151.413 192 159.9 192ZM287.9 208C287.9 203.756 289.586 199.687 292.586 196.686C295.587 193.686 299.657 192 303.9 192C308.144 192 312.213 193.686 315.214 196.686C318.214 199.687 319.9 203.756 319.9 208C319.9 212.243 318.214 216.313 315.214 219.314C312.213 222.314 308.144 224 303.9 224C299.657 224 295.587 222.314 292.586 219.314C289.586 216.313 287.9 212.243 287.9 208ZM287.9 288C296.387 288 304.526 291.371 310.528 297.372C316.529 303.374 319.9 311.513 319.9 320C319.9 328.487 316.529 336.626 310.528 342.627C304.526 348.628 296.387 352 287.9 352C279.413 352 271.274 348.628 265.273 342.627C259.272 336.626 255.9 328.487 255.9 320C255.9 311.513 259.272 303.374 265.273 297.372C271.274 291.371 279.413 288 287.9 288Z
-    `,
-  ],
-};
-
-export const dvDisease: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'disease' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M107.055 28.9758C112.485 45.2667 103.661 62.8183 87.4667 68.2486L62.0606 76.6849V193.921C62.0606 245.315 103.758 287.012 155.152 287.012C206.545 287.012 248.242 245.315 248.242 193.921V76.6849L222.933 68.2486C206.642 62.8183 197.915 45.2667 203.345 28.9758C208.776 12.6849 226.327 3.95766 242.618 9.38796L267.927 17.8243C293.236 26.2607 310.303 49.9213 310.303 76.6849V193.921C310.303 268.782 257.358 331.23 186.861 345.873C192.97 400.079 238.933 442.164 294.788 442.164C354.812 442.164 403.394 393.582 403.394 333.558V265C375.952 253.073 356.848 225.727 356.848 193.921C356.848 151.061 391.564 116.346 434.424 116.346C477.285 116.346 512 151.061 512 193.921C512 225.727 492.897 253.073 465.455 265V333.558C465.455 427.812 389.042 504.224 294.788 504.224C204.703 504.224 131.006 434.503 124.606 346.067C53.5273 331.812 0 269.17 0 193.921V76.6849C0 50.0183 17.0667 26.2607 42.4727 17.8243L67.7818 9.38796C84.0727 3.95766 101.624 12.7819 107.055 28.9758ZM434.424 224.952C442.654 224.952 450.547 221.682 456.366 215.863C462.185 210.044 465.455 202.151 465.455 193.921C465.455 185.692 462.185 177.799 456.366 171.98C450.547 166.16 442.654 162.891 434.424 162.891C426.195 162.891 418.302 166.16 412.483 171.98C406.663 177.799 403.394 185.692 403.394 193.921C403.394 202.151 406.663 210.044 412.483 215.863C418.302 221.682 426.195 224.952 434.424 224.952Z
-    `,
-  ],
-};
