--- src/icons/Icons.tsx
@@ -1,457 +1,0 @@
-import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
-
-/**
- * Visualization icons
- */
-
-// TODO: Split this into multiple files to allow code splitting
-export const dvAddVisualization: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'addVisualization' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M335.181 113.681H335.681V113.181V54.3191C335.681 34.5367 351.718 18.5 371.5 18.5C391.282 18.5 407.319 34.5367 407.319 54.3191V113.181V113.681H407.819H466.681C486.463 113.681 502.5 129.718 502.5 149.5C502.5 169.282 486.463 185.319 466.681 185.319H407.819H407.319V185.819V244.681C407.319 264.463 391.282 280.5 371.5 280.5C351.718 280.5 335.681 264.463 335.681 244.681V185.819V185.319H335.181H276.319C256.537 185.319 240.5 169.282 240.5 149.5C240.5 129.718 256.537 113.681 276.319 113.681H335.181ZM10.5 39.125C10.5 24.4031 22.0038 12.5 36.1566 12.5C50.3093 12.5 61.8132 24.4031 61.8132 39.125V436.435L61.8128 436.441C61.8122 436.455 61.8113 436.474 61.8103 436.499C61.8083 436.549 61.806 436.621 61.8043 436.714C61.8011 436.899 61.8007 437.165 61.8117 437.494C61.8337 438.152 61.9013 439.068 62.0845 440.1C62.4487 442.152 63.2802 444.727 65.179 446.634C67.0801 448.543 69.6474 449.375 71.694 449.736C72.7231 449.918 73.6357 449.984 74.2919 450.004C74.6204 450.015 74.8855 450.013 75.0699 450.01C75.1621 450.008 75.2342 450.005 75.284 450.003L75.3113 450.002L75.3419 450L75.3473 450H472.971C487.637 450 499.5 461.542 499.5 475.75C499.5 489.958 487.637 501.5 472.971 501.5H61.446H61.4413L61.432 501.5L61.3748 501.498C61.3238 501.497 61.2476 501.495 61.1473 501.492C60.9466 501.486 60.6498 501.475 60.2672 501.457C59.502 501.421 58.3941 501.356 57.027 501.242C54.292 501.013 50.523 500.585 46.387 499.791C38.087 498.197 28.4193 495.146 22.6028 489.355C16.6702 483.448 13.6281 473.531 12.0805 464.999C11.3093 460.747 10.914 456.869 10.7116 454.053C10.6104 452.646 10.5575 451.505 10.5299 450.716C10.5161 450.322 10.5086 450.016 10.5046 449.81C10.5026 449.706 10.5014 449.628 10.5008 449.575L10.5001 449.516L10.5 449.502L10.5 449.5V449.499V39.125ZM130.072 75.5L217.182 75.5C209.257 91.3114 203.759 108.543 201.176 126.71L130.072 126.71C115.379 126.71 103.5 115.229 103.5 101.105C103.5 86.9808 115.379 75.5 130.072 75.5ZM130.072 171.445H200.797C203.058 189.569 208.214 206.797 215.789 222.655L130.072 222.655C115.379 222.655 103.5 211.174 103.5 197.05C103.5 182.926 115.379 171.445 130.072 171.445ZM130.072 267.39H246.973C261.912 282.763 279.772 295.299 299.662 304.109C295.37 312.678 286.266 318.6 275.717 318.6H130.072C115.379 318.6 103.5 307.119 103.5 292.995C103.5 278.87 115.379 267.39 130.072 267.39ZM439.691 388.94C439.691 403.064 427.812 414.545 413.119 414.545H130.072C115.379 414.545 103.5 403.064 103.5 388.94C103.5 374.815 115.379 363.335 130.072 363.335H413.119C427.812 363.335 439.691 374.815 439.691 388.94Z',
-  ],
-};
-
-export const dvDownloadVisualization: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'downloadVisualization' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M9.49999 449.5L9.50001 449.503L10 449.5C9.50001 449.503 9.50001 449.503 9.50001 449.503L9.50002 449.504L9.50005 449.509L9.50018 449.525L9.50084 449.587C9.50152 449.642 9.5027 449.723 9.50476 449.829C9.50888 450.041 9.5165 450.352 9.53049 450.751C9.55848 451.55 9.612 452.704 9.71416 454.125C9.91841 456.966 10.3173 460.882 11.0965 465.177C12.6495 473.739 15.7324 483.926 21.8972 490.063C27.9395 496.079 37.8675 499.173 46.1984 500.773C50.3778 501.576 54.1832 502.007 56.9436 502.238C58.3242 502.354 59.4442 502.419 60.2201 502.456C60.6081 502.474 60.9101 502.485 61.1156 502.491C61.2184 502.495 61.297 502.497 61.3503 502.498L61.4109 502.499L61.4266 502.5L61.4308 502.5L61.4319 502.5C61.4322 502.5 61.4324 502.5 61.4413 502L61.4324 502.5H61.4413H472.971C488.161 502.5 500.5 490.537 500.5 475.75C500.5 460.963 488.161 449 472.971 449H75.3333H75.3176L75.3023 449.001L75.2999 449.001L75.289 449.002L75.2415 449.004C75.1986 449.006 75.1339 449.008 75.0494 449.01C74.8804 449.013 74.6325 449.014 74.323 449.005C73.7032 448.986 72.8397 448.923 71.868 448.752C69.913 448.406 67.5803 447.628 65.8876 445.928C64.196 444.229 63.4174 441.888 63.0691 439.925C62.896 438.95 62.8319 438.083 62.8111 437.461C62.8007 437.15 62.8012 436.901 62.8042 436.732C62.8057 436.647 62.8078 436.582 62.8095 436.539L62.8116 436.491L62.8122 436.48L62.8123 436.478L62.8123 436.478L62.8132 436.463V436.448V39.125C62.8132 23.8854 50.8956 11.5 36.1566 11.5C21.4176 11.5 9.5 23.8854 9.5 39.125L9.49999 449.5ZM218.44 75.2262L218.808 74.5H217.994L130.072 74.5C114.861 74.5 102.5 86.3946 102.5 101.105C102.5 115.816 114.861 127.71 130.072 127.71L201.611 127.71H202.046L202.106 127.279C204.686 108.787 210.304 91.2636 218.44 75.2262ZM201.735 170.884L201.682 170.445H201.239H130.072C114.861 170.445 102.5 182.339 102.5 197.05C102.5 211.76 114.861 223.655 130.072 223.655L216.584 223.655H217.382L217.035 222.937C209.256 206.851 203.987 189.332 201.735 170.884ZM247.543 266.542L247.396 266.39H247.184H130.072C114.861 266.39 102.5 278.284 102.5 292.995C102.5 307.705 114.861 319.6 130.072 319.6H275.717C286.839 319.6 296.433 313.243 300.79 304.076L301.012 303.61L300.54 303.403C280.514 294.608 262.543 282.016 247.543 266.542ZM413.119 415.545C428.329 415.545 440.691 403.65 440.691 388.94C440.691 374.229 428.329 362.335 413.119 362.335H130.072C114.861 362.335 102.5 374.229 102.5 388.94C102.5 403.65 114.861 415.545 130.072 415.545H413.119ZM427.181 138.396L407.5 158.077V53C407.5 33.3939 391.606 17.5 372 17.5C352.394 17.5 336.5 33.3939 336.5 53V161.454L313.319 138.273C299.507 124.461 277.113 124.461 263.302 138.273C249.49 152.085 249.49 174.479 263.302 188.291L340.56 265.549C341.82 267.42 343.277 269.197 344.931 270.851C352.391 278.311 362.344 281.756 372.108 281.188C380.742 280.888 389.288 277.442 395.879 270.851C397.353 269.378 398.669 267.807 399.828 266.159L477.386 188.601C491.249 174.737 491.249 152.26 477.386 138.396C463.522 124.533 441.045 124.533 427.181 138.396Z',
-  ],
-};
-
-export const dvViolin: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'violin' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M61.6512 502V501H474.861C489.325 501 501 489.62 501 475.643C501 461.666 489.325 450.286 474.861 450.286H75.6274C75.613 450.286 75.5944 450.287 75.5717 450.288C75.5182 450.291 75.4421 450.293 75.3456 450.295C75.1528 450.299 74.878 450.3 74.5389 450.29C73.8619 450.269 72.921 450.201 71.8591 450.013C69.7586 449.642 67.0641 448.78 65.0514 446.759C63.0416 444.741 62.1804 442.04 61.8068 439.934C61.6179 438.87 61.5481 437.926 61.5255 437.247C61.5141 436.907 61.5145 436.632 61.5179 436.439C61.5196 436.342 61.522 436.266 61.5242 436.212C61.5251 436.189 61.526 436.169 61.5267 436.155V37.2357C61.5267 22.7116 50.1819 11 36.2633 11C22.3447 11 11 22.7116 11 37.2357V449.283L11 449.284L11.0001 449.297L11.0007 449.355C11.0014 449.406 11.0025 449.484 11.0045 449.587C11.0085 449.792 11.0159 450.096 11.0297 450.49C11.0572 451.276 11.1101 452.415 11.2112 453.822C11.4135 456.636 11.8086 460.512 12.5789 464.759C14.1302 473.311 17.1644 483.135 23.0055 488.95C28.7333 494.653 38.3109 497.695 46.6294 499.293C50.7608 500.086 54.5269 500.514 57.2604 500.742C58.6265 500.857 59.7329 500.921 60.4958 500.957C60.8773 500.975 61.1727 500.986 61.3718 500.992C61.4713 500.995 61.5467 500.997 61.5967 500.999L61.6524 501L61.6656 501L61.6685 501H61.6691C61.6691 501 61.669 501 61.6512 502ZM61.6512 502H474.861C489.85 502 502 490.199 502 475.643C502 461.086 489.85 449.286 474.861 449.286H75.6L61.6512 502Z
-    M389 24L385.539 30.3361L381.565 38.3509C377.703 46.7222 363.025 67.0526 335.214 81.4035C318.476 90.9708 285 114.411 285 131.632C285 140.003 288.863 159.257 304.313 169.303C321.051 181.86 354.527 212.355 354.527 233.882V262.583C353.883 267.367 351.823 279.446 348.733 289.491C346.158 299.656 341.008 321.781 341.008 328.956C341.652 340.317 345.257 365.551 354.527 375.596C363.54 385.762 381.565 409.321 381.565 422.237L384.481 427.401L389 433L393.523 427.401L396.439 422.237C396.439 409.321 414.464 385.762 423.476 375.596C432.746 365.551 436.351 340.317 436.995 328.956C436.995 321.781 431.845 299.656 429.27 289.491C426.18 279.446 424.12 267.367 423.476 262.583V233.882C423.476 212.355 456.951 181.86 473.688 169.303C489.138 159.257 493 140.003 493 131.632C493 114.411 459.526 90.9708 442.788 81.4035C414.979 67.0526 400.302 46.7222 396.439 38.3509L392.465 30.3361L389 24Z
-    M174.2 77H175.572H176.428H177.8C177.8 77 178.139 81.6353 181.916 86.6102C185.692 91.585 191.829 111.603 199.529 116.965C210.652 123.589 240.599 142.426 240.599 161.211V175.485C240.599 182.668 244.877 191.23 248.299 198.322C250.866 204.507 256 218.875 256 226.868C256 236.383 251.926 256.896 240.599 268.26C229.476 279.419 206.969 298.255 204.663 311.079C202.096 325.352 189.262 366.752 187.551 373.601C185.84 380.451 184.128 382.164 181.916 385.898C178.717 391.296 176.428 401 176.428 401H175.572C175.572 401 173.283 391.296 170.084 385.898C167.872 382.164 166.16 380.451 164.449 373.601C162.738 366.752 149.904 325.352 147.337 311.079C145.031 298.255 122.524 279.419 111.401 268.26C100.074 256.896 96 236.383 96 226.868C96 218.875 101.134 204.507 103.701 198.322C107.123 191.23 111.401 182.668 111.401 175.485V161.211C111.401 142.426 141.348 123.589 152.471 116.965C160.171 111.603 166.308 91.585 170.084 86.6102C173.861 81.6353 174.2 77 174.2 77Z
-    `,
-  ],
-};
-
-export const dvBoxplot: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'boxplot' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M75.5717 450.288C75.5944 450.287 75.613 450.286 75.6274 450.286H474.861C489.325 450.286 501 461.666 501 475.643C501 489.62 489.325 501 474.861 501H61.6512V502C61.669 501 61.6691 501 61.6691 501H61.6685L61.6656 501L61.6524 501L61.5967 500.999C61.5467 500.997 61.4713 500.995 61.3718 500.992C61.1727 500.986 60.8773 500.975 60.4958 500.957C59.7329 500.921 58.6265 500.857 57.2604 500.742C54.5269 500.514 50.7608 500.086 46.6294 499.293C38.3109 497.695 28.7333 494.653 23.0055 488.95C17.1644 483.135 14.1302 473.311 12.5789 464.759C11.8086 460.512 11.4135 456.636 11.2112 453.822C11.1101 452.415 11.0572 451.276 11.0297 450.49C11.0159 450.096 11.0085 449.792 11.0045 449.587C11.0025 449.484 11.0014 449.406 11.0007 449.355L11.0001 449.297L11 449.284L11 449.283V37.2357C11 22.7116 22.3447 11 36.2633 11C50.1819 11 61.5267 22.7116 61.5267 37.2357V436.155C61.526 436.169 61.5251 436.189 61.5242 436.212C61.5221 436.266 61.5196 436.342 61.5179 436.439C61.5145 436.632 61.5141 436.907 61.5255 437.247C61.5481 437.926 61.6179 438.87 61.8068 439.934C62.1804 442.04 63.0416 444.741 65.0514 446.759C67.0641 448.78 69.7586 449.642 71.8591 450.013C72.921 450.201 73.8619 450.269 74.5389 450.29C74.878 450.3 75.1528 450.299 75.3456 450.295C75.4421 450.293 75.5182 450.291 75.5717 450.288Z
-    M275 336.162V118.4C275 113.443 270.971 109.424 266 109.424H208.4V75.3143L231.8 75.3143C236.771 75.3143 240.8 71.2955 240.8 66.3381V51.9762C240.8 47.0188 236.771 43 231.8 43L150.8 43C145.829 43 141.8 47.0188 141.8 51.9762V66.3381C141.8 71.2955 145.829 75.3143 150.8 75.3143H176V109.424H113C108.029 109.424 104 113.443 104 118.4V336.162C104 341.12 108.029 345.139 113 345.139H176V387.686H152.6C147.629 387.686 143.6 391.704 143.6 396.662V411.024C143.6 415.981 147.629 420 152.6 420H233.6C238.571 420 242.6 415.981 242.6 411.024V396.662C242.6 391.704 238.571 387.686 233.6 387.686H208.4V345.139H266C270.971 345.139 275 341.12 275 336.162Z
-    M490 248.367V129.837C490 124.877 485.994 120.857 481.053 120.857H423.789V46.3265L447.053 46.3265C451.994 46.3265 456 42.3062 456 37.3469V22.9796C456 18.0203 451.994 14 447.053 14L366.526 14C361.585 14 357.579 18.0203 357.579 22.9796V37.3469C357.579 42.3062 361.585 46.3265 366.526 46.3265H391.579V120.857H328.947C324.006 120.857 320 124.877 320 129.837V248.367C320 253.327 324.006 257.347 328.947 257.347H391.579V333.673H368.316C363.374 333.673 359.368 337.694 359.368 342.653V357.02C359.368 361.98 363.374 366 368.316 366H448.842C453.784 366 457.789 361.98 457.789 357.02V342.653C457.789 337.694 453.784 333.673 448.842 333.673H423.789V257.347H481.053C485.994 257.347 490 253.327 490 248.367Z
-    `,
-  ],
-};
-
-export const dvScatterplot: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'scatterplot' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M480.18 88C480.18 110.443 462.048 128.5 439.68 128.5C417.312 128.5 399.18 110.443 399.18 88C399.18 65.5573 417.312 47.5 439.68 47.5C462.048 47.5 480.18 65.5573 480.18 88ZM333 129C333 151.445 314.866 169.637 292.5 169.637C270.134 169.637 252 151.445 252 129C252 106.555 270.134 88.3629 292.5 88.3629C314.866 88.3629 333 106.555 333 129ZM206.3 362.134C206.3 384.579 188.166 402.771 165.8 402.771C143.434 402.771 125.3 384.579 125.3 362.134C125.3 339.689 143.434 321.497 165.8 321.497C188.166 321.497 206.3 339.689 206.3 362.134ZM232 212.137C232 234.582 213.866 252.774 191.5 252.774C169.134 252.774 151 234.582 151 212.137C151 189.692 169.134 171.5 191.5 171.5C213.866 171.5 232 189.692 232 212.137ZM351.5 320.997C351.5 343.442 333.366 361.634 311 361.634C288.634 361.634 270.5 343.442 270.5 320.997C270.5 298.552 288.634 280.36 311 280.36C333.366 280.36 351.5 298.552 351.5 320.997ZM439.18 212.137C439.18 234.582 421.046 252.774 398.68 252.774C376.314 252.774 358.18 234.582 358.18 212.137C358.18 189.692 376.314 171.5 398.68 171.5C421.046 171.5 439.18 189.692 439.18 212.137Z
-    M75.5717 450.288C75.5943 450.287 75.613 450.286 75.6274 450.286H474.861C489.325 450.286 501 461.666 501 475.643C501 489.62 489.325 501 474.861 501H61.6512V502C61.6689 501 61.669 501 61.669 501H61.6685L61.6656 501L61.6524 501L61.5966 500.999C61.5467 500.997 61.4713 500.995 61.3717 500.992C61.1727 500.986 60.8772 500.975 60.4958 500.957C59.7328 500.921 58.6265 500.857 57.2604 500.742C54.5269 500.514 50.7607 500.086 46.6294 499.293C38.3109 497.695 28.7333 494.653 23.0055 488.95C17.1644 483.135 14.1301 473.311 12.5789 464.759C11.8086 460.512 11.4135 456.636 11.2111 453.822C11.11 452.415 11.0572 451.276 11.0297 450.49C11.0159 450.096 11.0085 449.792 11.0045 449.587C11.0025 449.484 11.0013 449.406 11.0007 449.355L11.0001 449.297L11 449.284L11 449.283V37.2357C11 22.7116 22.3447 11 36.2633 11C50.1819 11 61.5267 22.7116 61.5267 37.2357V436.155C61.5259 436.169 61.5251 436.189 61.5241 436.212C61.522 436.266 61.5196 436.342 61.5179 436.439C61.5145 436.632 61.5141 436.907 61.5254 437.247C61.5481 437.926 61.6178 438.87 61.8068 439.934C62.1804 442.04 63.0416 444.741 65.0514 446.759C67.064 448.78 69.7586 449.642 71.8591 450.013C72.921 450.201 73.8619 450.269 74.5388 450.29C74.878 450.3 75.1528 450.299 75.3456 450.295C75.4421 450.293 75.5182 450.291 75.5717 450.288Z
-    `,
-  ],
-};
-
-export const dvHexbinplot: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'hexbinplot' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M61.6512 502V501H474.861C489.325 501 501 489.62 501 475.643C501 461.666 489.325 450.286 474.861 450.286H75.6274C75.613 450.286 75.5944 450.287 75.5717 450.288C75.5182 450.291 75.4421 450.293 75.3456 450.295C75.1528 450.299 74.878 450.3 74.5389 450.29C73.8619 450.269 72.921 450.201 71.8591 450.013C69.7586 449.642 67.0641 448.78 65.0514 446.759C63.0416 444.741 62.1804 442.04 61.8068 439.934C61.6179 438.87 61.5481 437.926 61.5255 437.247C61.5141 436.907 61.5145 436.632 61.5179 436.439C61.5196 436.342 61.522 436.266 61.5242 436.212C61.5251 436.189 61.526 436.169 61.5267 436.155V37.2357C61.5267 22.7116 50.1819 11 36.2633 11C22.3447 11 11 22.7116 11 37.2357V449.283L11 449.284L11.0001 449.297L11.0007 449.355C11.0014 449.406 11.0025 449.484 11.0045 449.587C11.0085 449.792 11.0159 450.096 11.0297 450.49C11.0572 451.276 11.1101 452.415 11.2112 453.822C11.4135 456.636 11.8086 460.512 12.5789 464.759C14.1302 473.311 17.1644 483.135 23.0055 488.95C28.7333 494.653 38.3109 497.695 46.6294 499.293C50.7608 500.086 54.5269 500.514 57.2604 500.742C58.6265 500.857 59.7329 500.921 60.4958 500.957C60.8773 500.975 61.1727 500.986 61.3718 500.992C61.4713 500.995 61.5467 500.997 61.5967 500.999L61.6524 501L61.6656 501L61.6685 501H61.6691C61.6691 501 61.669 501 61.6512 502ZM61.6512 502H474.861C489.85 502 502 490.199 502 475.643C502 461.086 489.85 449.286 474.861 449.286H75.6L61.6512 502Z
-    M153.169 243.901H244.832L290.664 323.336L244.832 402.771H153.169L107.337 323.336L153.169 243.901ZM319.648 148.697H411.311L457.143 228.132L411.311 307.567H319.648L273.816 228.132L319.648 148.697ZM153.169 56.5736L244.832 56.5736L290.664 136.009L244.832 215.444H153.169L107.337 136.009L153.169 56.5736Z
-    `,
-  ],
-};
-
-export const dvSankey: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'sankey' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M63.7633 473.613L63.7633 348.159C107.446 348.584 140.255 367.276 188.941 395.049C190.408 395.886 191.889 396.731 193.385 397.583C286.27 450.528 367.626 444.279 419.463 440.297L419.481 440.296C430.475 439.451 440.102 438.712 448.237 438.655V473.613C448.237 489.172 460.409 501.849 475.5 501.849C490.591 501.849 502.763 489.172 502.763 473.613V38.3871C502.763 22.8275 490.591 10.1514 475.5 10.1514C460.409 10.1514 448.237 22.8275 448.237 38.3871V63.3311C440.102 63.2738 430.475 62.5349 419.481 61.6905L419.463 61.6891C367.626 57.7073 286.27 51.4581 193.385 104.403C191.889 105.255 190.408 106.1 188.941 106.937C140.255 134.71 107.446 153.402 63.7633 153.827L63.7633 38.3871C63.7633 22.8275 51.591 10.1514 36.4999 10.1514C21.4089 10.1514 9.23659 22.8275 9.23659 38.3871L9.23657 473.613C9.23657 489.172 21.4088 501.849 36.4999 501.849C51.591 501.849 63.7633 489.172 63.7633 473.613ZM63.7633 277.827L63.7633 215.829C124.375 215.392 170.413 189.013 215.511 163.173L215.588 163.129C218.423 161.504 221.255 159.881 224.088 158.267C298.865 115.644 358.136 119.814 409.687 123.445L409.707 123.447C423.006 124.384 435.817 125.286 448.237 125.333V187.331C440.102 187.274 430.475 186.535 419.481 185.69L419.463 185.689C367.626 181.707 286.27 175.458 193.385 228.403C191.889 229.255 190.408 230.1 188.941 230.937C140.255 258.71 107.446 277.402 63.7633 277.827ZM215.511 338.813C199.972 329.909 184.321 320.941 168 312.993C184.321 305.045 199.972 296.077 215.511 287.173L215.588 287.129L215.628 287.105C218.451 285.488 221.269 283.873 224.088 282.267C298.865 239.644 358.136 243.814 409.687 247.445L409.709 247.447C423.007 248.384 435.818 249.286 448.237 249.333L448.237 376.653C435.818 376.7 423.007 377.602 409.709 378.539L409.687 378.541C358.136 382.172 298.865 386.342 224.088 343.719C221.269 342.112 218.451 340.498 215.628 338.881L215.588 338.857L215.511 338.813Z',
-  ],
-};
-
-export const dvHeatmap: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'heatmap' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M61.6424 502.5H61.6512H474.861C490.112 502.5 502.5 490.489 502.5 475.643C502.5 460.796 490.112 448.786 474.861 448.786H75.6H75.5843L75.5689 448.787L75.5665 448.787L75.5555 448.787L75.5078 448.79C75.4648 448.792 75.3998 448.794 75.3149 448.796C75.1452 448.799 74.8963 448.8 74.5854 448.791C73.963 448.771 73.0959 448.709 72.1202 448.536C70.157 448.189 67.8143 447.408 66.1143 445.701C64.4153 443.994 63.6334 441.643 63.2837 439.672C63.1099 438.693 63.0455 437.822 63.0246 437.197C63.0142 436.885 63.0146 436.636 63.0176 436.465C63.0191 436.38 63.0213 436.315 63.023 436.271L63.0251 436.223L63.0257 436.212L63.0258 436.21L63.0258 436.21L63.0267 436.195V436.181V37.2357C63.0267 21.935 51.0613 9.5 36.2633 9.5C21.4654 9.5 9.49997 21.935 9.49997 37.2357L9.49996 449.286L9.49998 449.288L9.99997 449.286C9.49998 449.288 9.49998 449.289 9.49998 449.289L9.49999 449.29L9.50002 449.294L9.50015 449.311L9.50082 449.373C9.50149 449.428 9.50268 449.51 9.50475 449.616C9.50889 449.829 9.51653 450.141 9.53059 450.542C9.55869 451.344 9.61243 452.502 9.71499 453.929C9.92008 456.782 10.3206 460.714 11.103 465.027C12.6623 473.623 15.7577 483.851 21.9472 490.013C28.0137 496.053 37.9816 499.16 46.3465 500.766C50.5429 501.572 54.3637 502.005 57.1353 502.237C58.5215 502.353 59.6461 502.419 60.4252 502.455C60.8147 502.474 61.1179 502.485 61.3243 502.491C61.4274 502.495 61.5064 502.497 61.5599 502.498L61.6207 502.499L61.6365 502.5L61.6407 502.5L61.6418 502.5C61.6419 502.5 61.642 502.5 61.6421 502.494C61.6427 502.474 61.6444 502.386 61.6512 502L61.6424 502.5ZM189 29.5H99C95.9624 29.5 93.5 31.9624 93.5 35V125C93.5 128.038 95.9624 130.5 99 130.5H189C192.038 130.5 194.5 128.038 194.5 125V35C194.5 31.9624 192.038 29.5 189 29.5ZM449 29.5H359C355.962 29.5 353.5 31.9624 353.5 35V125C353.5 128.038 355.962 130.5 359 130.5H449C452.038 130.5 454.5 128.038 454.5 125V35C454.5 31.9624 452.038 29.5 449 29.5ZM189 174.5H99C95.9624 174.5 93.5 176.962 93.5 180V270C93.5 273.038 95.9624 275.5 99 275.5H189C192.038 275.5 194.5 273.038 194.5 270V180C194.5 176.962 192.038 174.5 189 174.5ZM319 174.5H229C225.962 174.5 223.5 176.962 223.5 180V270C223.5 273.038 225.962 275.5 229 275.5H319C322.038 275.5 324.5 273.038 324.5 270V180C324.5 176.962 322.038 174.5 319 174.5ZM319 319.5H229C225.962 319.5 223.5 321.962 223.5 325V415C223.5 418.038 225.962 420.5 229 420.5H319C322.038 420.5 324.5 418.038 324.5 415V325C324.5 321.962 322.038 319.5 319 319.5ZM449 319.5H359C355.962 319.5 353.5 321.962 353.5 325V415C353.5 418.038 355.962 420.5 359 420.5H449C452.038 420.5 454.5 418.038 454.5 415V325C454.5 321.962 452.038 319.5 449 319.5Z',
-  ],
-};
-
-export const dvOncoprint: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'oncoprint' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M52.6424 493.5H52.6512H465.861C481.112 493.5 493.5 481.489 493.5 466.643C493.5 451.796 481.112 439.786 465.861 439.786H66.6H66.5843L66.5689 439.787L66.5665 439.787L66.5555 439.787L66.5078 439.79C66.4648 439.792 66.3998 439.794 66.3149 439.796C66.1452 439.799 65.8963 439.8 65.5855 439.791C64.9631 439.771 64.0959 439.709 63.1202 439.536C61.157 439.189 58.8143 438.408 57.1143 436.701C55.4153 434.994 54.6334 432.643 54.2837 430.672C54.1099 429.693 54.0455 428.822 54.0246 428.197C54.0142 427.885 54.0147 427.636 54.0177 427.465C54.0192 427.38 54.0213 427.315 54.023 427.271L54.0251 427.223L54.0257 427.212L54.0258 427.21L54.0258 427.21L54.0267 427.195V427.181V28.2357C54.0267 12.935 42.0613 0.5 27.2633 0.5C12.4654 0.5 0.499985 12.935 0.499985 28.2357L0.499977 440.286L0.499992 440.288L0.999985 440.286C0.499992 440.288 0.499993 440.289 0.499995 440.289L0.500002 440.29L0.500031 440.294L0.500161 440.311L0.500833 440.373C0.501508 440.428 0.502698 440.51 0.504766 440.616C0.508902 440.829 0.516548 441.141 0.5306 441.542C0.558704 442.344 0.612442 443.502 0.71501 444.929C0.920092 447.782 1.32065 451.714 2.10298 456.027C3.66234 464.623 6.75771 474.851 12.9472 481.013C19.0137 487.053 28.9817 490.16 37.3465 491.766C41.5429 492.572 45.3637 493.005 48.1354 493.237C49.5215 493.353 50.6462 493.419 51.4252 493.455C51.8147 493.474 52.1179 493.485 52.3243 493.491C52.4275 493.495 52.5064 493.497 52.5599 493.498L52.6208 493.499L52.6365 493.5L52.6407 493.5L52.6418 493.5C52.6419 493.5 52.642 493.5 52.6422 493.494C52.6427 493.474 52.6444 493.386 52.6512 493L52.6424 493.5ZM133 20.5H90C86.9624 20.5 84.5 22.9624 84.5 26V116C84.5 119.038 86.9624 121.5 90 121.5H133C136.038 121.5 138.5 119.038 138.5 116V26C138.5 22.9624 136.038 20.5 133 20.5ZM216 20.5H173C169.962 20.5 167.5 22.9624 167.5 26V116C167.5 119.038 169.962 121.5 173 121.5H216C219.038 121.5 221.5 119.038 221.5 116V26C221.5 22.9624 219.038 20.5 216 20.5ZM299 20.5H256C252.962 20.5 250.5 22.9624 250.5 26V116C250.5 119.038 252.962 121.5 256 121.5H299C302.038 121.5 304.5 119.038 304.5 116V26C304.5 22.9624 302.038 20.5 299 20.5ZM216 163.5H173C169.962 163.5 167.5 165.962 167.5 169V259C167.5 262.038 169.962 264.5 173 264.5H216C219.038 264.5 221.5 262.038 221.5 259V169C221.5 165.962 219.038 163.5 216 163.5ZM299 163.5H256C252.962 163.5 250.5 165.962 250.5 169V259C250.5 262.038 252.962 264.5 256 264.5H299C302.038 264.5 304.5 262.038 304.5 259V169C304.5 165.962 302.038 163.5 299 163.5ZM382 163.5H339C335.962 163.5 333.5 165.962 333.5 169V259C333.5 262.038 335.962 264.5 339 264.5H382C385.038 264.5 387.5 262.038 387.5 259V169C387.5 165.962 385.038 163.5 382 163.5ZM299 306.5H256C252.962 306.5 250.5 308.962 250.5 312V402C250.5 405.038 252.962 407.5 256 407.5H299C302.038 407.5 304.5 405.038 304.5 402V312C304.5 308.962 302.038 306.5 299 306.5ZM382 306.5H339C335.962 306.5 333.5 308.962 333.5 312V402C333.5 405.038 335.962 407.5 339 407.5H382C385.038 407.5 387.5 405.038 387.5 402V312C387.5 308.962 385.038 306.5 382 306.5ZM465 306.5H422C418.962 306.5 416.5 308.962 416.5 312V402C416.5 405.038 418.962 407.5 422 407.5H465C468.038 407.5 470.5 405.038 470.5 402V312C470.5 308.962 468.038 306.5 465 306.5Z',
-  ],
-};
-
-export const dvCorrelationplot: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'correlationplot' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M61.6424 502.5H61.6512H474.861C490.112 502.5 502.5 490.489 502.5 475.643C502.5 460.796 490.112 448.786 474.861 448.786H75.6H75.5843L75.5689 448.787L75.5665 448.787L75.5555 448.787L75.5078 448.79C75.4648 448.792 75.3998 448.794 75.3149 448.796C75.1452 448.799 74.8963 448.8 74.5854 448.791C73.963 448.771 73.0959 448.709 72.1202 448.536C70.157 448.189 67.8143 447.408 66.1143 445.701C64.4153 443.994 63.6334 441.643 63.2837 439.672C63.1099 438.693 63.0455 437.822 63.0246 437.197C63.0142 436.885 63.0146 436.636 63.0176 436.465C63.0191 436.38 63.0213 436.315 63.023 436.271L63.0251 436.223L63.0257 436.212L63.0258 436.21L63.0258 436.21L63.0267 436.195V436.181V37.2357C63.0267 21.935 51.0613 9.5 36.2633 9.5C21.4654 9.5 9.49997 21.935 9.49997 37.2357L9.49996 449.286L9.49998 449.288L9.99997 449.286C9.49998 449.288 9.49998 449.289 9.49998 449.289L9.49999 449.29L9.50002 449.294L9.50015 449.311L9.50082 449.373C9.50149 449.428 9.50268 449.51 9.50475 449.616C9.50889 449.829 9.51653 450.141 9.53059 450.542C9.55869 451.344 9.61243 452.502 9.71499 453.929C9.92008 456.782 10.3206 460.714 11.103 465.027C12.6623 473.623 15.7577 483.851 21.9472 490.013C28.0137 496.053 37.9816 499.16 46.3465 500.766C50.5429 501.572 54.3637 502.005 57.1353 502.237C58.5215 502.353 59.6461 502.419 60.4252 502.455C60.8147 502.474 61.1179 502.485 61.3243 502.491C61.4274 502.495 61.5064 502.497 61.5599 502.498L61.6207 502.499L61.6365 502.5L61.6407 502.5L61.6418 502.5C61.6419 502.5 61.642 502.5 61.6421 502.494C61.6427 502.474 61.6444 502.386 61.6512 502L61.6424 502.5ZM153 166.5C186.413 166.5 213.5 139.413 213.5 106C213.5 72.5868 186.413 45.5 153 45.5C119.587 45.5 92.5 72.5868 92.5 106C92.5 139.413 119.587 166.5 153 166.5ZM153 277.5C178.129 277.5 198.5 257.129 198.5 232C198.5 206.871 178.129 186.5 153 186.5C127.871 186.5 107.5 206.871 107.5 232C107.5 257.129 127.871 277.5 153 277.5ZM153 390.5C169.845 390.5 183.5 376.845 183.5 360C183.5 343.155 169.845 329.5 153 329.5C136.155 329.5 122.5 343.155 122.5 360C122.5 376.845 136.155 390.5 153 390.5ZM279 151.5C304.129 151.5 324.5 131.129 324.5 106C324.5 80.871 304.129 60.5 279 60.5C253.871 60.5 233.5 80.871 233.5 106C233.5 131.129 253.871 151.5 279 151.5ZM279 292.5C312.413 292.5 339.5 265.413 339.5 232C339.5 198.587 312.413 171.5 279 171.5C245.587 171.5 218.5 198.587 218.5 232C218.5 265.413 245.587 292.5 279 292.5ZM279 405.5C304.129 405.5 324.5 385.129 324.5 360C324.5 334.871 304.129 314.5 279 314.5C253.871 314.5 233.5 334.871 233.5 360C233.5 385.129 253.871 405.5 279 405.5ZM407 136.5C423.845 136.5 437.5 122.845 437.5 106C437.5 89.1553 423.845 75.5 407 75.5C390.155 75.5 376.5 89.1553 376.5 106C376.5 122.845 390.155 136.5 407 136.5ZM407 277.5C432.129 277.5 452.5 257.129 452.5 232C452.5 206.871 432.129 186.5 407 186.5C381.871 186.5 361.5 206.871 361.5 232C361.5 257.129 381.871 277.5 407 277.5ZM407 420.5C440.413 420.5 467.5 393.413 467.5 360C467.5 326.587 440.413 299.5 407 299.5C373.587 299.5 346.5 326.587 346.5 360C346.5 393.413 373.587 420.5 407 420.5Z',
-  ],
-};
-
-/**
- * Other custom icons
- */
-
-export const dvChemicalStructure: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'chemicalStructure' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M253.549 213.538L298.645 142.445C304.45 133.293 314.697 127.934 325.525 128.385L350.626 129.431L350.782 129.438L407.553 131.804L443.531 75.8179C443.153 74.9596 442.816 74.0922 442.521 73.2182L442.512 73.1936C437.697 58.9103 443.843 42.8479 457.622 35.6405C472.88 27.6596 491.719 33.5587 499.7 48.8166C507.681 64.0745 501.782 82.9133 486.524 90.8943C481.76 93.3861 476.647 94.5249 471.625 94.4482L434.864 151.652C429.039 160.715 418.845 166.01 408.08 165.561L326.065 162.143L279.642 235.328L311.104 295.476C313.337 299.744 313.122 304.879 310.542 308.947L265.56 379.863C262.98 383.93 258.426 386.313 253.613 386.112L181.818 383.119L147.026 437.967C152.377 452.455 146.262 469.011 132.213 476.36C116.955 484.341 98.1161 478.441 90.1351 463.184C82.1541 447.926 88.0533 429.087 103.311 421.106C108.645 418.316 114.417 417.222 120.01 417.632L153.116 365.441L123.717 309.237L66.2436 306.842C63.4104 311.308 59.4013 315.113 54.3776 317.741C39.1197 325.722 20.2809 319.823 12.3 304.565C4.319 289.307 10.2181 270.468 25.476 262.487C39.8473 254.97 57.3953 259.766 66.0552 273.099L127.475 275.658L164.942 216.591C167.522 212.523 172.076 210.141 176.888 210.342L253.549 213.538ZM247.725 247.03L187.15 244.505L154.676 295.701L182.776 349.423L243.351 351.949L275.825 300.752L247.725 247.03Z
-    M253.549 213.538L298.645 142.445C304.45 133.293 314.697 127.934 325.525 128.385L350.626 129.431L350.782 129.438L407.553 131.804L443.531 75.8179C443.153 74.9596 442.816 74.0922 442.521 73.2182L442.512 73.1936C437.697 58.9103 443.843 42.8479 457.622 35.6405C472.88 27.6596 491.719 33.5587 499.7 48.8166C507.681 64.0745 501.782 82.9133 486.524 90.8943C481.76 93.3861 476.647 94.5249 471.625 94.4482L434.864 151.652C429.039 160.715 418.845 166.01 408.08 165.561L326.065 162.143L279.642 235.328L311.104 295.476C313.337 299.744 313.122 304.879 310.542 308.947L265.56 379.863C262.98 383.93 258.426 386.313 253.613 386.112L181.818 383.119L147.026 437.967C152.377 452.455 146.262 469.011 132.213 476.36C116.955 484.341 98.1161 478.441 90.1351 463.184C82.1541 447.926 88.0533 429.087 103.311 421.106C108.645 418.316 114.417 417.222 120.01 417.632L153.116 365.441L123.717 309.237L66.2436 306.842C63.4104 311.308 59.4013 315.113 54.3776 317.741C39.1197 325.722 20.2809 319.823 12.3 304.565C4.319 289.307 10.2181 270.468 25.476 262.487C39.8473 254.97 57.3953 259.766 66.0552 273.099L127.475 275.658L164.942 216.591C167.522 212.523 172.076 210.141 176.888 210.342L253.549 213.538ZM247.725 247.03L187.15 244.505L154.676 295.701L182.776 349.423L243.351 351.949L275.825 300.752L247.725 247.03Z
-    M298.645 142.445L294.423 139.767L294.423 139.767L298.645 142.445ZM253.549 213.538L253.341 218.534L256.225 218.654L257.771 216.216L253.549 213.538ZM325.525 128.385L325.733 123.389L325.733 123.389L325.525 128.385ZM350.626 129.431L350.418 134.427L350.418 134.427L350.626 129.431ZM350.782 129.438L350.99 124.442L350.99 124.442L350.782 129.438ZM407.553 131.804L407.345 136.799L410.209 136.919L411.759 134.507L407.553 131.804ZM443.531 75.8179L447.738 78.521L449.191 76.26L448.106 73.8007L443.531 75.8179ZM442.521 73.2182L447.257 71.6168L447.257 71.6147L442.521 73.2182ZM442.512 73.1936L437.774 74.7911L437.777 74.7971L442.512 73.1936ZM457.622 35.6405L459.94 40.071L459.94 40.071L457.622 35.6405ZM499.7 48.8166L504.13 46.4992L504.13 46.4991L499.7 48.8166ZM486.524 90.8943L484.206 86.4638L484.206 86.4638L486.524 90.8943ZM471.625 94.4482L471.701 89.4488L468.921 89.4064L467.418 91.7451L471.625 94.4482ZM434.864 151.652L430.657 148.949L430.657 148.949L434.864 151.652ZM408.08 165.561L408.289 160.565L408.289 160.565L408.08 165.561ZM326.065 162.143L326.274 157.147L323.39 157.027L321.843 159.465L326.065 162.143ZM279.642 235.328L275.42 232.65L273.874 235.087L275.212 237.645L279.642 235.328ZM311.104 295.476L306.673 297.793L306.673 297.793L311.104 295.476ZM310.542 308.947L306.32 306.268L306.32 306.268L310.542 308.947ZM265.56 379.863L261.337 377.185L261.337 377.185L265.56 379.863ZM253.613 386.112L253.821 381.116L253.821 381.116L253.613 386.112ZM181.818 383.119L182.026 378.123L179.142 378.003L177.595 380.44L181.818 383.119ZM147.026 437.967L142.804 435.289L141.478 437.378L142.336 439.699L147.026 437.967ZM132.213 476.36L129.895 471.929L129.895 471.929L132.213 476.36ZM103.311 421.106L105.629 425.536L105.629 425.536L103.311 421.106ZM120.01 417.632L119.645 422.618L122.629 422.837L124.232 420.31L120.01 417.632ZM153.116 365.441L157.338 368.119L158.884 365.681L157.546 363.123L153.116 365.441ZM123.717 309.237L128.148 306.92L126.81 304.362L123.926 304.241L123.717 309.237ZM66.2436 306.842L66.4518 301.846L63.5676 301.726L62.0214 304.164L66.2436 306.842ZM54.3776 317.741L56.6951 322.171L56.6951 322.171L54.3776 317.741ZM12.3 304.565L7.86946 306.882L7.86946 306.882L12.3 304.565ZM25.476 262.487L27.7935 266.917L27.7935 266.917L25.476 262.487ZM66.0552 273.099L61.8621 275.822L63.2681 277.987L65.847 278.094L66.0552 273.099ZM127.475 275.658L127.266 280.654L130.151 280.774L131.697 278.337L127.475 275.658ZM164.942 216.591L169.164 219.269L169.164 219.269L164.942 216.591ZM176.888 210.342L176.68 215.338L176.68 215.338L176.888 210.342ZM187.15 244.505L187.358 239.509L184.474 239.389L182.928 241.827L187.15 244.505ZM247.725 247.03L252.155 244.713L250.817 242.155L247.933 242.035L247.725 247.03ZM154.676 295.701L150.454 293.023L148.907 295.461L150.245 298.019L154.676 295.701ZM182.776 349.423L178.346 351.741L179.684 354.299L182.568 354.419L182.776 349.423ZM243.351 351.949L243.143 356.944L246.027 357.065L247.573 354.627L243.351 351.949ZM275.825 300.752L280.047 303.43L281.594 300.993L280.256 298.435L275.825 300.752ZM294.423 139.767L249.327 210.86L257.771 216.216L302.867 145.123L294.423 139.767ZM325.733 123.389C313.12 122.864 301.185 129.107 294.423 139.767L302.867 145.123C307.716 137.48 316.273 133.004 325.317 133.381L325.733 123.389ZM350.834 124.435L325.733 123.389L325.317 133.381L350.418 134.427L350.834 124.435ZM350.99 124.442L350.834 124.435L350.418 134.427L350.574 134.433L350.99 124.442ZM407.761 126.808L350.99 124.442L350.574 134.433L407.345 136.799L407.761 126.808ZM439.325 73.1148L403.347 129.101L411.759 134.507L447.738 78.521L439.325 73.1148ZM437.784 74.8195C438.127 75.8341 438.518 76.8403 438.956 77.8351L448.106 73.8007C447.788 73.079 447.505 72.3504 447.257 71.6168L437.784 74.8195ZM437.777 74.7971L437.785 74.8216L447.257 71.6147L447.248 71.5902L437.777 74.7971ZM455.305 31.21C439.312 39.5754 432.185 58.2131 437.774 74.7911L447.25 71.5962C443.208 59.6074 448.375 46.1204 459.94 40.071L455.305 31.21ZM504.13 46.4991C494.869 28.7944 473.01 21.9492 455.305 31.21L459.94 40.071C472.751 33.37 488.568 38.3231 495.269 51.1341L504.13 46.4991ZM488.841 95.3248C506.546 86.0639 513.391 64.204 504.13 46.4992L495.269 51.1341C501.97 63.9451 497.017 79.7627 484.206 86.4638L488.841 95.3248ZM471.548 99.4476C477.383 99.5367 483.321 98.2122 488.841 95.3248L484.206 86.4638C480.199 88.56 475.911 89.513 471.701 89.4488L471.548 99.4476ZM439.07 154.355L475.831 97.1513L467.418 91.7451L430.657 148.949L439.07 154.355ZM407.872 170.557C420.411 171.079 432.286 164.912 439.07 154.355L430.657 148.949C425.793 156.518 417.279 160.94 408.289 160.565L407.872 170.557ZM325.857 167.138L407.872 170.557L408.289 160.565L326.274 157.147L325.857 167.138ZM283.864 238.006L330.288 164.821L321.843 159.465L275.42 232.65L283.864 238.006ZM315.534 293.159L284.073 233.01L275.212 237.645L306.673 297.793L315.534 293.159ZM314.765 311.625C318.302 306.049 318.595 299.01 315.534 293.158L306.673 297.793C308.078 300.479 307.943 303.709 306.32 306.268L314.765 311.625ZM269.782 382.541L314.765 311.625L306.32 306.268L261.337 377.185L269.782 382.541ZM253.405 391.108C260.002 391.383 266.245 388.117 269.782 382.541L261.337 377.185C259.714 379.744 256.849 381.242 253.821 381.116L253.405 391.108ZM181.609 388.114L253.405 391.108L253.821 381.116L182.026 378.123L181.609 388.114ZM151.248 440.645L186.04 385.797L177.595 380.44L142.804 435.289L151.248 440.645ZM134.53 480.79C150.837 472.261 157.927 453.051 151.716 436.235L142.336 439.699C146.827 451.859 141.688 465.761 129.895 471.929L134.53 480.79ZM85.7046 465.501C94.9654 483.206 116.825 490.051 134.53 480.79L129.895 471.929C117.084 478.63 101.267 473.677 94.5656 460.866L85.7046 465.501ZM100.994 416.675C83.2889 425.936 76.4437 447.796 85.7046 465.501L94.5656 460.866C87.8645 448.055 92.8176 432.237 105.629 425.536L100.994 416.675ZM120.375 412.645C113.881 412.169 107.176 413.441 100.994 416.675L105.629 425.536C110.114 423.19 114.954 422.275 119.645 422.618L120.375 412.645ZM148.894 362.763L115.788 414.953L124.232 420.31L157.338 368.119L148.894 362.763ZM119.287 311.555L148.685 367.758L157.546 363.123L128.148 306.92L119.287 311.555ZM66.0354 311.838L123.509 314.233L123.926 304.241L66.4518 301.846L66.0354 311.838ZM56.6951 322.171C62.517 319.126 67.175 314.708 70.4658 309.52L62.0214 304.164C59.6457 307.909 56.2855 311.1 52.0602 313.31L56.6951 322.171ZM7.86946 306.882C17.1303 324.587 38.9903 331.432 56.6951 322.171L52.0602 313.31C39.2492 320.011 23.4315 315.058 16.7305 302.247L7.86946 306.882ZM23.1586 258.056C5.45377 267.317 -1.3914 289.177 7.86946 306.882L16.7305 302.247C10.0294 289.436 14.9825 273.619 27.7935 266.917L23.1586 258.056ZM70.2483 270.375C60.1967 254.9 39.837 249.332 23.1586 258.056L27.7935 266.917C39.8575 260.607 54.5939 264.633 61.8621 275.822L70.2483 270.375ZM127.683 270.663L66.2634 268.103L65.847 278.094L127.266 280.654L127.683 270.663ZM160.719 213.913L123.252 272.98L131.697 278.337L169.164 219.269L160.719 213.913ZM177.097 205.346C170.499 205.071 164.256 208.337 160.719 213.913L169.164 219.269C170.787 216.71 173.652 215.211 176.68 215.338L177.097 205.346ZM253.757 208.542L177.097 205.346L176.68 215.338L253.341 218.534L253.757 208.542ZM186.942 249.501L247.517 252.026L247.933 242.035L187.358 239.509L186.942 249.501ZM158.898 298.38L191.372 247.183L182.928 241.827L150.454 293.023L158.898 298.38ZM187.207 347.106L159.106 293.384L150.245 298.019L178.346 351.741L187.207 347.106ZM243.559 346.953L182.984 344.428L182.568 354.419L243.143 356.944L243.559 346.953ZM271.603 298.074L239.129 349.27L247.573 354.627L280.047 303.43L271.603 298.074ZM243.294 249.348L271.395 303.07L280.256 298.435L252.155 244.713L243.294 249.348Z
-    `,
-  ],
-};
-
-export const dvFilterOutline: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'filterOutline' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M182.47 320.171C182.964 320.782 183.234 321.545 183.234 322.332V403.913C184.447 415.915 187.528 421.126 196.799 428.028L279.143 476.062C279.32 476.193 279.515 476.31 279.715 476.403C306.683 489.012 325.695 474.564 327.285 451.34V322.342C327.285 321.549 327.559 320.781 328.061 320.167L504.611 93.6265C522.166 68.7071 507.005 32.5339 474.29 30.9262L45.0065 31.73C5.1103 30.1223 -9.25235 71.1187 5.90822 91.2149L182.47 320.171ZM215.869 293.162L47.7754 74.6779L465.263 73.8962L294.802 292.986C288.031 301.271 284.332 311.641 284.332 322.342V426.526L226.187 396.33V322.332C226.187 311.714 222.545 301.417 215.869 293.162ZM41.2861 74.7644C41.2282 74.7875 41.1954 74.7938 41.1954 74.7938C41.1954 74.7938 41.2231 74.7805 41.2861 74.7644Z
-    `,
-  ],
-};
-
-export const dvFilterClear: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'filterClear' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M182.47 320.171C182.964 320.782 183.234 321.545 183.234 322.332V403.913C184.447 415.915 187.528 421.126 196.799 428.028L279.143 476.062C279.32 476.193 279.515 476.31 279.715 476.403C298.697 485.278 313.738 480.748 321.536 469.188C272.975 449.051 238.819 401.186 238.819 345.342C238.819 271.329 298.819 211.329 372.832 211.329C385.375 211.329 397.514 213.052 409.028 216.274L504.611 93.6265C522.166 68.7071 507.005 32.5339 474.29 30.9262L45.0065 31.73C5.1103 30.1223 -9.25235 71.1187 5.90822 91.2149L182.47 320.171Z
-    M374.55 237.101C434.345 237.101 482.792 285.547 482.792 345.342C482.792 405.137 434.345 453.584 374.55 453.584C314.756 453.584 266.309 405.137 266.309 345.342C266.309 285.547 314.756 237.101 374.55 237.101ZM427.362 374.148L398.992 345.342L427.362 316.972C429.544 315.227 429.544 311.735 427.362 309.553L410.34 292.531C408.158 290.349 404.666 290.349 402.92 292.531L374.55 320.901L346.181 292.531C343.998 290.349 340.507 290.349 338.761 292.531L321.302 309.553C319.12 311.735 319.12 315.227 321.302 316.972L350.109 345.342L321.302 373.712C319.12 375.894 319.12 379.386 321.302 381.132L338.324 398.59C340.507 400.772 343.998 400.772 345.744 398.59L374.55 369.784L402.92 398.59C404.666 400.772 408.158 400.772 410.34 398.59L427.362 381.568C429.544 379.386 429.544 375.894 427.362 374.148Z
-    `,
-  ],
-};
-
-export const dvStringCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'stringCol' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-M403.178 293.855L403.182 293.863C407.692 303.971 413.72 311.585 421.28 316.669C428.829 321.746 437.139 324.286 446.199 324.286C453.925 324.286 461.095 322.633 467.7 319.323C474.297 316.108 480.444 311.85 486.141 306.553L511.353 346.295C501.963 354.857 492.219 361.388 482.12 365.899C471.924 370.235 460.513 372.407 447.879 372.407C435.071 372.407 422.822 369.803 411.127 364.593C399.621 359.166 389.41 351.459 380.494 341.467C371.767 331.479 364.797 319.31 359.589 304.952C354.572 290.606 352.058 274.291 352.058 256C352.058 237.708 354.572 221.394 359.589 207.048C364.797 192.69 371.767 180.522 380.493 170.533C389.41 160.539 399.528 152.943 410.847 147.734L410.852 147.732C422.359 142.306 434.512 139.593 447.319 139.593C459.398 139.593 470.442 141.549 480.458 145.453L480.458 145.453L480.47 145.458C490.56 149.105 500.386 154.664 509.949 162.141L486.097 203.183C480.404 198.736 474.265 195.122 467.681 192.34C460.895 189.255 453.639 187.714 445.919 187.714C436.86 187.714 428.548 190.254 421 195.331L420.995 195.334C413.623 200.419 407.69 208.032 403.182 218.137L403.179 218.144C398.851 228.268 396.698 240.893 396.698 256C396.698 271.107 398.851 283.732 403.178 293.855ZM113.033 143.523L183.004 368.477H136.46L123.941 324.305L123.838 323.941H123.46H59.1023H58.7244L58.6213 324.305L46.1025 368.477H0.679148L70.6494 143.523H113.033ZM91.7623 210.767L91.2812 209.07L90.8002 210.767L70.4088 282.716L70.2284 283.353H70.8898H111.673H112.334L112.154 282.716L91.7623 210.767ZM194.052 368.477V143.523H265.513C277.576 143.523 288.51 146.019 298.328 151.001L298.336 151.004C308.149 155.768 315.823 162.582 321.371 171.451L321.375 171.457C327.1 180.313 329.974 191.02 329.974 203.605C329.974 216.19 327.1 226.786 321.378 235.42L321.377 235.422C317.606 241.155 312.853 245.939 307.114 249.774L306.42 250.238L307.157 250.631C316.891 255.822 324.516 263.258 330.04 272.948L330.042 272.953C335.59 282.469 338.374 293.731 338.374 306.758C338.374 319.778 335.406 330.925 329.492 340.219L329.491 340.22C323.756 349.308 315.894 356.343 305.893 361.324C296.077 366.089 284.954 368.477 272.514 368.477H194.052ZM307.86 249.873C307.783 249.831 307.705 249.79 307.628 249.748L307.392 250.189L307.67 250.605C307.743 250.556 307.817 250.507 307.89 250.458C317.47 255.681 324.998 263.095 330.474 272.701C336.074 282.306 338.874 293.659 338.874 306.758C338.874 319.856 335.888 331.1 329.914 340.487L307.86 249.873ZM238.353 186.404H237.853V186.904V234.059V234.559H238.353H262.713C269.903 234.559 275.726 232.565 280.131 228.534L280.141 228.525C284.558 224.257 286.734 218.218 286.734 210.481C286.734 202.749 284.56 196.813 280.136 192.762C275.733 188.51 269.908 186.404 262.713 186.404H238.353ZM238.353 274.166H237.853V274.666V325.096V325.596H238.353H268.313C275.877 325.596 281.982 323.493 286.573 319.241L286.575 319.239C291.371 314.753 293.734 308.271 293.734 299.881C293.734 291.496 291.375 285.12 286.574 280.849C281.984 276.38 275.879 274.166 268.313 274.166H238.353Z
-    `,
-  ],
-};
-
-export const dvNumericCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'numericalCol' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    `
-    M3.91283 329.338V368.452H149.34V329.338H104.016V143.548H69.1267L0 171.905L15.6513 210.041L51.845 195.24V329.338H3.91283Z
-    M190.623 156.586C181.928 161.801 173.885 167.777 166.494 174.513L194.862 212.323C200.079 207.325 205.296 202.979 210.513 199.285C215.948 195.591 221.491 192.766 227.143 190.81C232.795 188.637 238.555 187.551 244.425 187.551C251.163 187.551 256.706 188.746 261.054 191.136C265.402 193.527 268.662 196.786 270.836 200.915C273.227 204.826 274.423 209.281 274.423 214.279C274.423 218.625 273.336 223.188 271.162 227.968C268.988 232.749 265.619 237.964 261.054 243.614C256.489 249.264 250.511 255.674 243.12 262.845L170.081 334.228V368.452H331.159V322.819H244.398L274.749 296.092C287.357 284.792 297.574 274.47 305.399 265.127C313.225 255.565 318.877 246.222 322.355 237.095C326.051 227.751 327.898 217.973 327.898 207.76C327.898 194.722 324.529 183.096 317.79 172.883C311.051 162.67 301.813 154.63 290.074 148.763C278.336 142.679 265.076 139.636 250.294 139.636C239.207 139.636 228.665 141.157 218.665 144.2C208.666 147.242 199.318 151.37 190.623 156.586Z
-    M369.834 303.914L342.77 341.073C353.639 351.068 365.921 358.782 379.616 364.215C393.311 369.647 408.745 372.364 425.918 372.364C441.352 372.364 455.59 369.647 468.633 364.215C481.676 358.565 492.11 350.199 499.935 339.117C507.978 328.035 512 314.453 512 298.373C512 283.597 508.413 271.32 501.24 261.541C494.066 251.545 484.175 244.049 471.567 239.051C464.495 236.074 456.856 233.947 448.65 232.669L503.848 177.772V143.548H351.574V189.181H440.178L389.398 241.332V269.364H418.418C428.2 269.364 436.135 270.668 442.221 273.275C448.308 275.883 452.764 279.36 455.59 283.706C458.633 288.052 460.155 292.941 460.155 298.373C460.155 307.5 456.46 314.562 449.069 319.56C441.678 324.341 431.896 326.731 419.722 326.731C409.723 326.731 400.593 324.775 392.333 320.864C384.072 316.735 376.573 311.085 369.834 303.914Z    `,
-  ],
-};
-
-export const dvNumberArrayCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'numberArrayCol' as IconName,
-  icon: [
-    512,
-    312,
-    [],
-    '',
-    `M0 312V0H83.4854V33.28H41.0103V278.72H83.4854V312H0Z
-    M80.0617 250.187V216.907H123.123V102.923L90.6072 115.403L76.5465 82.9547L138.648 58.8267H169.992V216.907H210.709V250.187H80.0617Z
-    M232.857 301.493L246.332 252.128C241.45 250.649 237.251 247.876 233.736 243.808C230.221 239.556 228.463 234.286 228.463 228C228.463 220.604 231.002 214.503 236.08 209.696C241.352 204.889 247.992 202.485 255.999 202.485C264.006 202.485 270.548 204.981 275.625 209.973C280.898 214.78 283.534 221.067 283.534 228.832C283.534 234.564 282.753 239.925 281.191 244.917C279.629 249.909 277.285 256.196 274.161 263.776L258.635 301.493H232.857Z
-    M293.002 85.1733C299.642 79.4418 306.867 74.3573 314.679 69.92C322.49 65.4827 330.888 61.9698 339.871 59.3813C348.854 56.7929 358.326 55.4987 368.285 55.4987C381.565 55.4987 393.477 58.0871 404.023 63.264C414.568 68.256 422.868 75.0969 428.922 83.7867C434.976 92.4764 438.003 102.368 438.003 113.461C438.003 122.151 436.343 130.471 433.023 138.421C429.898 146.187 424.821 154.137 417.791 162.272C410.76 170.222 401.582 179.004 390.255 188.619L363.013 211.36H440.932V250.187H296.224V221.067L361.841 160.331C368.481 154.229 373.851 148.775 377.952 143.968C382.053 139.161 385.08 134.724 387.033 130.656C388.986 126.588 389.962 122.706 389.962 119.008C389.962 114.756 388.888 110.965 386.74 107.637C384.787 104.124 381.858 101.351 377.952 99.3173C374.046 97.2836 369.066 96.2667 363.013 96.2667C357.74 96.2667 352.565 97.1911 347.487 99.04C342.41 100.704 337.43 103.108 332.548 106.251C327.861 109.394 323.174 113.092 318.487 117.344L293.002 85.1733Z
-    M428.515 312V278.72H470.99V33.28H428.515V0H512V312H428.515Z`,
-  ],
-};
-
-export const dvStringArrayCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'stringArrayCol' as IconName,
-  icon: [
-    512,
-    307,
-    [],
-    '',
-    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
-    M51.0333 246.998L120.156 58.7044H162.566L231.689 246.998H184.873L168.9 199.515L192.309 209.885H89.3125L111.895 201.698L96.7481 246.998H51.0333ZM114.373 193.784L100.879 175.228H180.742L167.248 193.784L134.477 96.636H147.144L114.373 193.784Z
-    M234.512 297.482L247.18 248.908C242.59 247.453 238.643 244.724 235.338 240.721C232.034 236.537 230.381 231.352 230.381 225.167C230.381 217.89 232.768 211.886 237.541 207.156C242.498 202.426 248.741 200.061 256.268 200.061C263.795 200.061 269.946 202.517 274.719 207.429C279.676 212.159 282.155 218.344 282.155 225.985C282.155 231.625 281.42 236.901 279.951 241.813C278.483 246.725 276.28 252.91 273.342 260.369L258.746 297.482H234.512Z
-    M304.274 246.998V58.7044H375.049C387.166 58.7044 397.999 60.7966 407.545 64.9809C417.276 68.9833 424.895 74.7139 430.403 82.1729C436.094 89.6319 438.94 98.6372 438.94 109.189C438.94 118.103 436.92 125.835 432.881 132.384C428.842 138.752 423.334 143.937 416.358 147.939C426.088 152.305 433.616 158.582 438.94 166.768C444.448 174.773 447.202 184.233 447.202 195.149C447.202 205.883 444.264 215.161 438.389 222.984C432.698 230.806 424.895 236.81 414.981 240.994C405.25 244.997 394.235 246.998 381.934 246.998H304.274ZM348.336 210.431H377.803C385.147 210.431 391.022 208.702 395.428 205.246C400.018 201.607 402.313 196.331 402.313 189.418C402.313 182.505 400.018 177.32 395.428 173.864C391.022 170.225 385.147 168.406 377.803 168.406H348.336V210.431ZM348.336 134.568H372.295C379.272 134.568 384.872 132.93 389.094 129.656C393.317 126.199 395.428 121.287 395.428 114.92C395.428 108.552 393.317 103.731 389.094 100.456C384.872 96.9998 379.272 95.2715 372.295 95.2715H348.336V134.568Z
-    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z`,
-  ],
-};
-
-export const dvBooleanArrayCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'booleanArrayCol' as IconName,
-  icon: [
-    512,
-    308,
-    [],
-    '',
-    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
-    M234.512 297.482L247.18 248.908C242.59 247.453 238.643 244.724 235.338 240.721C232.034 236.537 230.381 231.352 230.381 225.167C230.381 217.89 232.768 211.886 237.541 207.156C242.498 202.426 248.741 200.061 256.268 200.061C263.795 200.061 269.946 202.517 274.719 207.429C279.676 212.159 282.155 218.344 282.155 225.985C282.155 231.625 281.42 236.901 279.951 241.813C278.483 246.725 276.28 252.91 273.342 260.369L258.746 297.482H234.512Z
-    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z
-    M150 68C196.956 68 235 106.044 235 153C235 199.956 196.956 238 150 238C103.044 238 65 199.956 65 153C65 106.044 103.044 68 150 68Z
-    M452 153C452 103.283 411.717 63 362 63C312.283 63 272 103.283 272 153C272 202.717 312.283 243 362 243C411.717 243 452 202.717 452 153ZM425.548 153C425.548 188.267 396.956 216.548 362 216.548C326.717 216.548 298.452 188.283 298.452 153C298.452 118.044 326.733 89.4516 362 89.4516C396.94 89.4516 425.548 118.06 425.548 153Z`,
-  ],
-};
-
-export const dvCategoricalArrayCol: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'categoricalArrayCol' as IconName,
-  icon: [
-    512,
-    307,
-    [],
-    '',
-    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
-    M246.512 297.482L259.18 248.908C254.59 247.453 250.643 244.724 247.338 240.721C244.034 236.537 242.381 231.352 242.381 225.167C242.381 217.89 244.768 211.886 249.541 207.156C254.498 202.426 260.741 200.061 268.268 200.061C275.795 200.061 281.946 202.517 286.719 207.429C291.676 212.159 294.155 218.344 294.155 225.985C294.155 231.625 293.42 236.901 291.951 241.813C290.483 246.725 288.28 252.91 285.342 260.369L270.746 297.482H246.512Z
-    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z
-    M65 103.786C65 89.0536 76.3839 77 91.7857 77H188.214C202.946 77 215 89.0536 215 103.786V200.214C215 215.616 202.946 227 188.214 227H91.7857C76.3839 227 65 215.616 65 200.214V103.786Z
-    M436.286 65.4136C429.589 61.2008 421.554 61.2008 414.857 65.4136L307.714 132.818C301.018 137.031 297 144.052 297 151.776C297 160.202 301.018 167.223 307.714 171.436L414.857 238.84C421.554 243.053 429.589 243.053 436.286 238.84C442.982 235.33 447 227.606 447 219.181L447 84.3712C447 76.6477 442.982 68.9242 436.286 65.4136Z`,
-  ],
-};
-
-export const dvSortAsc: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'sortAsc' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M250.622 110.314C261.557 121.248 261.557 138.977 250.622 149.912C239.688 160.846 221.959 160.846 211.024 149.912L157.421 96.308V443.711C157.421 459.175 144.886 471.711 129.421 471.711C113.957 471.711 101.421 459.175 101.421 443.711V96.2895L47.7988 149.912C36.8643 160.846 19.1357 160.846 8.20117 149.912C-2.7334 138.977 -2.7334 121.248 8.20117 110.314L105.577 12.9375C106.61 11.3647 107.818 9.87315 109.201 8.49028C114.773 2.91771 122.111 0.185044 129.414 0.292283C136.716 0.186143 144.051 2.91881 149.622 8.49034C151.005 9.87322 152.213 11.3647 153.246 12.9374L250.622 110.314ZM482 451C498.568 451 512 437.569 512 421C512 404.431 498.568 391 482 391H222C205.432 391 192 404.431 192 421C192 437.569 205.432 451 222 451H482ZM482 265.586C498.568 265.586 512 252.154 512 235.585C512 219.017 498.568 205.585 482 205.585H222C205.432 205.585 192 219.017 192 235.585C192 252.154 205.432 265.586 222 265.586H482ZM482 81C498.568 81 512 67.5685 512 51C512 34.4314 498.568 21 482 21H282C265.432 21 252 34.4314 252 51C252 67.5685 265.432 81 282 81H482Z',
-  ],
-};
-
-export const dvSortDesc: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'sortDesc' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M157.421 28.2893C157.421 12.8253 144.885 0.289307 129.421 0.289307C113.957 0.289307 101.421 12.8253 101.421 28.2893V375.711L47.7988 322.088C36.8641 311.154 19.1355 311.154 8.20087 322.088C-2.73383 333.023 -2.73383 350.752 8.20087 361.686L105.576 459.062C106.61 460.635 107.818 462.127 109.201 463.51C114.773 469.082 122.111 471.815 129.414 471.708C136.715 471.814 144.051 469.081 149.622 463.51C151.005 462.127 152.213 460.635 153.246 459.062L250.622 361.686C261.557 350.752 261.557 333.023 250.622 322.088C239.688 311.154 221.959 311.154 211.024 322.088L157.421 375.691V28.2893ZM482 21C498.569 21 512 34.4315 512 51C512 67.5685 498.569 81 482 81H222C205.431 81 192 67.5685 192 51C192 34.4315 205.431 21 222 21H482ZM482 206.415C498.569 206.415 512 219.846 512 236.414C512 252.983 498.569 266.414 482 266.414H222C205.431 266.414 192 252.983 192 236.414C192 219.846 205.431 206.415 222 206.415H482ZM482 391C498.569 391 512 404.431 512 421C512 437.569 498.569 451 482 451H282C265.431 451 252 437.569 252 421C252 404.431 265.431 391 282 391H482Z',
-  ],
-};
-
-export const dvSort: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'sort' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    '',
-    'M8.20093 110.314C-2.73364 121.248 -2.73364 138.977 8.20093 149.912C19.1355 160.846 36.864 160.846 47.7986 149.912L101.402 96.308V443.711C101.402 459.175 113.938 471.711 129.402 471.711C144.866 471.711 157.402 459.175 157.402 443.711V96.2896L211.024 149.912C221.959 160.846 239.687 160.846 250.622 149.912C261.556 138.977 261.556 121.248 250.622 110.314L153.246 12.9376C152.213 11.3649 151.005 9.87317 149.622 8.49023C144.05 2.91797 136.713 0.185425 129.41 0.292358C122.108 0.185913 114.772 2.91858 109.201 8.49036C107.818 9.87317 106.61 11.3646 105.577 12.9373L8.20093 110.314ZM354.579 375.692V28.2893C354.579 12.8253 367.114 0.289307 382.579 0.289307C398.043 0.289307 410.579 12.8253 410.579 28.2893V375.71L464.201 322.088C475.135 311.154 492.864 311.154 503.799 322.088C514.733 333.023 514.733 350.752 503.799 361.686L406.423 459.062C405.389 460.635 404.181 462.127 402.799 463.51C397.226 469.082 389.889 471.815 382.587 471.708C375.285 471.814 367.949 469.081 362.378 463.51C360.995 462.127 359.787 460.635 358.754 459.063L261.378 361.686C250.443 350.752 250.443 333.023 261.378 322.088C272.312 311.154 290.041 311.154 300.975 322.088L354.579 375.692Z',
-  ],
-};
-
-/**
- * Select / deselect icons
- */
-
-export const dvBrushSelection: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'select' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    null,
-    `M89.6547 68.6723H67.8031V90.524C67.8031 109.727 52.2356 125.295 33.0323 125.295C17.6314 125.295 4.56916 115.282 0 101.411V30.3841C4.48165 15.0438 17.1934 3.21497 33.033 0H97.4152C112.883 3.52605 124.426 17.3658 124.426 33.9015C124.426 53.1049 108.858 68.6723 89.6547 68.6723Z
-      M199.875 0C184.408 3.52605 172.865 17.3658 172.865 33.9015C172.865 53.1049 188.432 68.6723 207.636 68.6723H303.495C322.699 68.6723 338.266 53.1049 338.266 33.9015C338.266 17.3658 326.723 3.52605 311.256 0H199.875Z
-      M413.715 0C398.248 3.52605 386.705 17.3658 386.705 33.9015C386.705 53.1049 402.273 68.6723 421.476 68.6723H443.328V90.524C443.328 109.727 458.895 125.295 478.098 125.295C494.634 125.295 508.474 113.752 512 98.2845V33.9023C508.543 16.8715 495.128 3.45674 478.098 0H413.715Z
-      M512 200.744C508.474 185.277 494.634 173.734 478.098 173.734C458.895 173.734 443.328 189.302 443.328 208.505V304.364C443.328 323.568 458.895 339.135 478.098 339.135C494.634 339.135 508.474 327.593 512 312.125V200.744Z
-      M512 414.585C508.474 399.117 494.634 387.574 478.098 387.574C458.895 387.574 443.328 403.142 443.328 422.345V444.197H421.476C402.273 444.197 386.705 459.764 386.705 478.968C386.705 494.369 396.718 507.431 410.589 512H481.616C496.956 507.518 508.785 494.807 512 478.967V414.585Z
-      M314.382 512C328.253 507.431 338.266 494.369 338.266 478.968C338.266 459.764 322.699 444.197 303.495 444.197H207.636C188.432 444.197 172.865 459.764 172.865 478.968C172.865 494.369 182.877 507.431 196.748 512H314.382Z
-      M100.542 512C114.413 507.431 124.426 494.369 124.426 478.968C124.426 459.764 108.858 444.197 89.6547 444.197H67.8031V422.345C67.8031 403.142 52.2356 387.574 33.0323 387.574C17.6314 387.574 4.56916 397.587 0 411.458V482.485C4.14462 496.672 15.3281 507.855 29.5148 512H100.542Z
-      M0 315.252C4.56916 329.123 17.6314 339.135 33.0323 339.135C52.2356 339.135 67.8031 323.568 67.8031 304.364V208.505C67.8031 189.301 52.2356 173.734 33.0323 173.734C17.6314 173.734 4.56916 183.747 0 197.618V315.252Z
-      `,
-  ],
-};
-
-export const dvClearSelection: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'select' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    null,
-    `M199.875 0C184.407 3.52605 172.865 17.3658 172.865 33.9015C172.865 53.1049 188.432 68.6723 207.636 68.6723H303.495C322.699 68.6723 338.266 53.1049 338.266 33.9015C338.266 17.3658 326.723 3.52605 311.256 0H199.875Z
-      M413.715 0C398.248 3.52605 386.705 17.3658 386.705 33.9015C386.705 53.1049 402.273 68.6723 421.476 68.6723H443.328V90.524C443.328 109.727 458.895 125.295 478.098 125.295C494.634 125.295 508.474 113.752 512 98.2845V33.9023C508.543 16.8715 495.128 3.45674 478.098 0H413.715Z
-      M512 200.744C508.474 185.277 494.634 173.734 478.098 173.734C458.895 173.734 443.328 189.302 443.328 208.505V304.364C443.328 323.568 458.895 339.135 478.098 339.135C494.634 339.135 508.474 327.593 512 312.125V200.744Z
-      M314.382 512C328.253 507.431 338.266 494.369 338.266 478.968C338.266 459.764 322.699 444.197 303.495 444.197H207.636C188.432 444.197 172.865 459.764 172.865 478.968C172.865 494.369 182.877 507.431 196.748 512H314.382Z
-      M100.542 512C114.413 507.431 124.426 494.369 124.426 478.968C124.426 459.764 108.858 444.197 89.6547 444.197H67.8031V422.345C67.8031 403.142 52.2356 387.574 33.0323 387.574C17.6314 387.574 4.56916 397.587 0 411.458V482.485C4.14462 496.672 15.3281 507.855 29.5148 512H100.542Z
-      M0 315.252C4.56916 329.123 17.6314 339.135 33.0323 339.135C52.2356 339.135 67.8031 323.568 67.8031 304.364V208.505C67.8031 189.301 52.2356 173.734 33.0323 173.734C17.6314 173.734 4.56916 183.747 0 197.618V315.252Z
-      M472 512C482.237 512 492.474 508.095 500.284 500.284C515.905 484.663 515.905 459.337 500.284 443.716L69.2843 12.7157C53.6633 -2.90524 28.3367 -2.90524 12.7157 12.7157C-2.90524 28.3367 -2.90524 53.6633 12.7157 69.2843L443.716 500.284C451.526 508.095 461.763 512 472 512Z
-      `,
-  ],
-};
-
-/*
- * Species
- */
-export const dvMouse: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'mouse' as IconName,
-  icon: [
-    512,
-    512,
-    [],
-    null,
-    `M433.2 93.5c-19.7 13.3-21.6 41.5-4.9 74.9 4.3 8.7 11.7 20.3 22.8 35.8 17 23.6 23.1 35.3 27.3 52.3 5.1 20.5 1.6 41.3-10 60.6-4.6 7.6-19.1 24.3-18 20.7 3.1-10.2 3.9-32.8 1.7-45.8-6-34.1-27.2-63.3-58.3-80-22.5-12-47.1-15.4-84.5-11.5-23.3 2.5-53.6 9.1-69.8 15.4-4.5 1.8-5 2.2-4.7 4.8.2 1.5.4 9.5.5 17.8.1 14.4 0 15.4-2.9 23.2-7.4 19.9-22 35.8-40.8 44.3-13.5 6.2-14.5 6.3-59.7 6.9l-41.7.6-9.3 10c-10.7 11.6-24.8 28.4-33.8 40.3-8.1 10.7-11.7 13.1-21.8 14.3-9.2 1.2-15.8 4.1-17.7 7.8-2 3.8-6.5 18.9-7.2 24.3-.8 5.7.9 10.6 4 11.3 1.1.3 91.9.4 201.7.3l199.6-.3 7.9-4c21-10.6 45.9-28.3 61.3-43.3 36.4-35.8 45.7-75.1 28.6-121-7.8-20.7-17.9-37.7-43.5-73.2-18-25.1-22.8-36.7-23.7-58-.5-11.7.8-24.2 3.2-30.3 1-2.5-.4-2.1-6.3 1.8zM109.4 343.9c9.9 6.1 10.2 22.4.5 28.8-10.8 7.1-25.3 1.1-26.6-11.1-.3-2.5-.3-6 .1-7.9.8-4.2 6.5-10.3 10.9-11.6 4.4-1.4 11.3-.5 15.1 1.8M154 189.1c-17.6 2.5-30.7 13.3-40.6 33.4-6.5 13.2-8.5 22.1-9.1 40.5-.4 10-.1 18.1.7 22.8l1.2 7.2h29.1c33.2 0 38.7-.8 50.6-6.6 8.8-4.3 19.1-14.1 23.4-22.2 9.1-16.9 8.9-37.4-.6-51.1-11.6-16.9-34.6-27-54.7-24z`,
-  ],
-};
-
-/*
- * Protein structure representation
- */
-export const dvProteinBallAndStick: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'proteinBallAndStick' as IconName,
-  icon: [
-    23,
-    24,
-    [],
-    null,
-    `M6.85714 2.28571C6.85714 2.64748 6.94119 2.98961 7.09081 3.29365L4.40099 4.84662C4.05931 4.01453 3.24094 3.42857 2.28571 3.42857C1.02335 3.42857 0 4.45192 0 5.71429C0 6.90101 0.904379 7.8765 2.06152 7.98914V12.5823C0.904379 12.6949 0 13.6704 0 14.8571C0 16.1195 1.02335 17.1429 2.28571 17.1429C3.24094 17.1429 4.05932 16.5569 4.40099 15.7248L8 17.8027V19.7344C7.3168 20.1296 6.85714 20.8683 6.85714 21.7143C6.85714 22.9767 7.88049 24 9.14286 24C10.4052 24 11.4286 22.9767 11.4286 21.7143C11.4286 20.8683 10.9689 20.1296 10.2857 19.7344V17.8027L13.8847 15.7248C14.2264 16.5569 15.0448 17.1429 16 17.1429C17.2624 17.1429 18.2857 16.1195 18.2857 14.8571C18.2857 13.6704 17.3813 12.6949 16.2242 12.5823V7.38756L19.5439 5.47088C19.8528 5.62659 20.2019 5.71429 20.5714 5.71429C21.8338 5.71429 22.8571 4.69094 22.8571 3.42857C22.8571 2.16621 21.8338 1.14286 20.5714 1.14286C19.3091 1.14286 18.2857 2.16621 18.2857 3.42857C18.2857 3.47133 18.2869 3.51381 18.2892 3.55598L14.9692 5.47278L11.1949 3.29367C11.3445 2.98962 11.4286 2.64748 11.4286 2.28571C11.4286 1.02335 10.4052 0 9.14286 0C7.88049 0 6.85714 1.02335 6.85714 2.28571ZM13.7143 13.1839L9.14284 15.8232L4.34724 13.0545V7.51697L9.14284 4.74823L13.7143 7.38756V13.1839Z`,
-  ],
-};
-
-export const dvProteinCartoon: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'proteinCartoon' as IconName,
-  icon: [
-    20,
-    18,
-    [],
-    null,
-    `M18.1132 9.15125C18.3735 9.42009 18.5179 9.62296 18.5878 9.76977C18.6066 9.80933 18.6164 9.83736 18.6215 9.85566C18.6079 9.87081 18.5836 9.89448 18.5424 9.92614C18.4351 10.0086 18.2779 10.0958 18.0565 10.2019C17.9892 10.2341 17.9071 10.2721 17.819 10.3127C17.6639 10.3843 17.49 10.4646 17.3467 10.5372C17.1061 10.659 16.8254 10.8182 16.591 11.0342C16.3476 11.2586 16.1128 11.5835 16.0601 12.0276C15.9926 12.5962 16.1226 13.1341 16.2587 13.5726C16.3145 13.7524 16.3779 13.9348 16.4362 14.1024L16.4753 14.215C16.546 14.4197 16.6074 14.6042 16.6549 14.782C16.7508 15.1406 16.7712 15.406 16.7184 15.6304C16.6701 15.8361 16.5387 16.0978 16.1531 16.4064C15.8189 16.6739 15.76 17.1677 16.0216 17.5095C16.2831 17.8513 16.7661 17.9115 17.1003 17.644C17.7039 17.161 18.0682 16.6122 18.2127 15.998C18.3527 15.4025 18.2633 14.8385 18.1374 14.3676C18.074 14.1307 17.9965 13.9003 17.9243 13.6913L17.8806 13.5653C17.823 13.3996 17.7706 13.2488 17.7235 13.0972C17.6089 12.7278 17.5623 12.4598 17.5834 12.2396C17.5897 12.2316 17.6011 12.219 17.62 12.2015C17.6946 12.1327 17.8226 12.0498 18.0285 11.9456C18.157 11.8806 18.2732 11.8272 18.4016 11.7683C18.4945 11.7256 18.5937 11.68 18.7086 11.625C18.9466 11.511 19.2244 11.3678 19.4652 11.1829C19.7063 10.9978 19.975 10.7221 20.0981 10.3196C20.2272 9.89762 20.1556 9.47255 19.9696 9.08183C19.7261 8.57046 19.2381 8.01647 18.4895 7.38197L17.9164 3.55115L17.8614 3.40956C17.825 3.31594 17.7914 3.22927 17.7572 3.14282C17.7128 3.03052 17.6705 2.91712 17.6281 2.80372C17.5317 2.54542 17.4353 2.2871 17.314 2.04172C17.0612 1.53088 16.6595 1.19629 16.0844 1.11798C15.62 1.05475 15.1557 0.990978 14.6914 0.927202L14.0882 0.844383C14.0175 0.85027 13.9469 0.856157 13.8387 0.853573C13.7789 0.841084 13.7566 0.837727 13.7342 0.835027C13.6634 0.841052 13.5927 0.847077 13.4814 0.843359C13.3878 0.826091 13.3348 0.818562 13.2818 0.811035C13.2108 0.817083 13.1397 0.823126 13.0476 0.824246C13.0265 0.819319 13.0265 0.819562 13.0265 0.819562L12.2869 1.00615C12.5836 0.931309 12.805 0.875354 13.0039 0.81197C12.8983 0.817061 12.7936 0.820225 12.6899 0.823358C12.3756 0.832857 12.0702 0.842086 11.7754 0.9039C11.6022 0.940207 11.449 1.07634 11.3041 1.20512C11.2679 1.23732 11.2322 1.26907 11.1968 1.29868C11.2325 1.35098 11.2681 1.4019 11.3032 1.452C11.4254 1.62645 11.5407 1.79113 11.6263 1.97063C11.6934 2.1112 11.7481 2.25854 11.8027 2.40566C11.8774 2.60678 11.9519 2.80754 12.0576 2.98996C12.1813 3.20353 12.3961 3.37901 12.6032 3.52122C12.9681 3.77182 13.3583 3.98397 13.7442 4.18947L13.7488 4.1736C13.824 3.91223 13.8931 3.67171 13.9565 3.40461C13.9099 3.27453 13.8602 3.14721 13.807 3.02267C13.8638 3.15473 13.917 3.29024 13.967 3.42926C13.9908 3.50422 14.0118 3.57366 14.0315 3.63918C14.0706 3.76865 14.1053 3.88366 14.1474 3.99475C14.4808 4.87365 14.8163 5.75173 15.1518 6.6298L15.1522 6.63106L15.1527 6.63232L15.4758 7.47811C15.5011 7.48345 15.5286 7.48616 15.5561 7.48888C15.6051 7.49371 15.6541 7.49855 15.6912 7.51825C15.6334 7.48783 15.5479 7.49616 15.4887 7.50572C15.8564 8.35808 16.3978 8.94294 17.331 9.03114L17.5312 9.06345C17.6354 9.08023 17.6914 9.08924 17.7473 9.09859C17.8121 9.10942 17.8769 9.1207 18.0165 9.14502C18.0489 9.15064 18.0816 9.15263 18.1132 9.15125ZM17.5761 12.2499C17.576 12.2502 17.5761 12.2499 17.5761 12.2499ZM13.7464 4.83233C13.7543 5.33418 13.7623 5.83603 13.7508 6.33742C13.7448 6.59547 13.6983 6.85255 13.6486 7.12676C13.6319 7.21891 13.6149 7.31299 13.599 7.40969C13.5655 7.48637 13.5478 7.5224 13.5302 7.55843C13.4604 7.63638 13.3907 7.71438 13.278 7.81121C13.2263 7.81217 13.1754 7.81505 13.126 7.81786C13.0322 7.82318 12.9435 7.8282 12.8643 7.81911C12.9297 7.82728 13.0007 7.82645 13.0741 7.8256C13.1348 7.82489 13.197 7.82417 13.2589 7.8285C13.1978 7.88074 13.1365 7.93499 13.0746 7.9898C12.8618 8.17828 12.6412 8.37356 12.3924 8.51743C12.0275 8.72838 11.6472 8.91156 11.267 9.09474C11.1057 9.17243 10.9445 9.25012 10.7844 9.32993C10.1533 9.64451 9.53403 9.71192 8.92139 9.2706C8.80896 9.1896 8.68345 9.12758 8.55154 9.06239C8.48282 9.02843 8.41237 8.99361 8.34112 8.9548C8.38045 8.92775 8.43115 8.893 8.48295 8.87672C8.45422 8.88494 8.42491 8.89798 8.3956 8.91103C8.37045 8.92222 8.3453 8.93341 8.32052 8.94155C8.26565 8.92021 8.211 8.89807 8.15636 8.87594C8.0151 8.81871 7.87372 8.76144 7.72851 8.71803C7.68494 8.70501 7.63178 8.72549 7.57864 8.74597C7.55431 8.75534 7.52999 8.76471 7.50658 8.77087C7.51349 8.79432 7.51823 8.81993 7.52296 8.84554C7.53315 8.90064 7.54334 8.95576 7.57521 8.98933C8.18832 9.6351 8.50498 10.4496 8.79717 11.2705C8.86189 11.4523 8.85422 11.661 8.84595 11.8863C8.84348 11.9537 8.84095 12.0225 8.84029 12.0926C8.45943 12.3915 8.16471 12.7213 7.96506 13.1178C8.16467 12.7222 8.45927 12.3966 8.83946 12.1163C8.8373 12.1512 8.83618 12.1879 8.83503 12.2256C8.83064 12.3693 8.8258 12.5276 8.76143 12.6542C8.26891 13.6228 7.75819 14.5819 7.24576 15.5398C7.08715 15.8364 6.82483 15.9446 6.47593 15.8917C7.10161 15.3386 7.43202 14.6029 7.69921 13.8293C7.43165 14.603 7.09859 15.3384 6.45196 15.8874C6.17569 15.8108 5.89919 15.7351 5.6227 15.6594L5.62165 15.6591C4.96991 15.4805 4.31818 15.302 3.66944 15.1128C3.06526 14.9365 2.46346 14.9743 1.80779 15.0722C1.71109 15.0958 1.6171 15.1175 1.52541 15.1386C1.25882 15.2002 1.0112 15.2574 0.773103 15.3415C0.178908 15.5512 0.103148 15.4956 0.197411 14.8509C0.199517 14.8365 0.20145 14.8221 0.203384 14.8076C0.209424 14.7624 0.215474 14.7171 0.226855 14.6733C0.39855 14.013 0.571134 13.3529 0.743718 12.6928L0.868742 12.2146C1.61314 11.4488 2.56716 11.3526 3.54337 11.3635C3.78834 11.3663 4.03283 11.4033 4.27746 11.4403C4.40554 11.4597 4.53368 11.4791 4.66192 11.4936C4.72815 11.5011 4.79536 11.4996 4.86763 11.4979C4.90224 11.4971 4.93865 11.4962 4.97604 11.4964C4.52884 10.8514 4.10642 10.2421 3.6782 9.58314C3.87524 9.30484 4.06433 9.06107 4.28527 8.85259C4.12095 9.00764 3.97427 9.18197 3.82094 9.36443C3.76814 9.42727 3.71446 9.49114 3.65909 9.55537C3.42706 9.29639 3.1938 9.02538 2.9937 8.73081C2.80533 8.45352 2.78233 8.14262 2.96719 7.84763C3.30568 7.3075 3.65152 6.77219 4.008 6.22041L4.11908 6.04846C4.06298 6.25478 3.99595 6.48374 3.9149 6.7074C3.97129 6.55178 4.02087 6.39384 4.07048 6.23565C4.09217 6.16649 4.11386 6.09733 4.13612 6.02839C4.22976 5.90468 4.32302 5.78066 4.41629 5.65665C4.63199 5.36984 4.84768 5.08302 5.06807 4.80003C5.32637 4.46834 5.69067 4.40032 6.06265 4.48279L6.06636 4.48362C6.98673 4.68766 7.90905 4.89213 8.81362 5.157C9.14834 5.25501 9.46681 5.41102 9.80005 5.57427C9.95483 5.65009 10.1128 5.72747 10.277 5.80133C10.2789 5.76349 10.2813 5.72549 10.2838 5.68737C10.2892 5.60418 10.2947 5.52001 10.2936 5.43639C10.2857 4.79755 10.2778 4.15867 10.2633 3.51996C10.2438 2.65627 10.6419 1.96944 11.1799 1.31973C11.3566 1.5335 11.5161 1.73957 11.6263 1.97063C11.6934 2.1112 11.7481 2.25854 11.8027 2.40566C11.8774 2.60678 11.9519 2.80754 12.0576 2.98996C12.1813 3.20353 12.3961 3.37901 12.6032 3.52122C12.8604 3.69787 13.1302 3.85584 13.4 4.01338C13.5129 4.07933 13.6259 4.14528 13.7379 4.21263C13.7398 4.41917 13.7431 4.62576 13.7464 4.83233Z`,
-  ],
-};
-
-export const dvProteinStick: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'proteinStick' as IconName,
-  icon: [
-    22,
-    24,
-    [],
-    null,
-    `M8.64145 25C9.33999 25 9.90636 24.4344 9.90636 23.7368V18.1504L15.2141 15.0903L15.2198 15.0871L15.2373 15.0767L15.8466 14.7255C16.4791 14.1875 16.4791 13.8555 16.4791 13.3359V6.78208C16.4791 6.73489 16.4778 6.68771 16.4752 6.64114L22.1712 3.35714C22.7761 3.00835 22.9835 2.23584 22.6341 1.63171C22.2848 1.02759 21.5112 0.820658 20.9063 1.16944L15.0903 4.52253L9.90636 1.53396C9.12366 1.08279 8.15923 1.08279 7.37653 1.53396L2.06875 4.59407C1.28605 5.04555 0.803833 5.87943 0.803833 6.78208V12.9023C0.803833 13.805 1.28605 14.6388 2.06875 15.0903L7.37653 18.1504V23.7368C7.37653 24.4344 7.9429 25 8.64145 25ZM13.7011 13.0454L8.64145 15.9624L3.33366 12.9023V6.78208L8.64145 3.72196L13.7011 6.63898V13.0454Z`,
-  ],
-};
-
-export const dvProteinSurface: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'proteinSurface' as IconName,
-  icon: [
-    24,
-    22,
-    [],
-    null,
-    `M16.8103 0.468166C14.917 -0.614518 12.4949 0.25773 11.7476 2.29133L11.6499 2.55724C11.1567 3.89944 9.91213 4.82923 8.47256 4.93099L6.04224 5.10277C4.12786 5.23809 2.64464 6.81442 2.64464 8.71365V9.4774C2.64464 10.5233 2.15666 11.5106 1.32237 12.1528C-0.577428 13.615 -0.400827 16.5091 1.66297 17.7345L2.19765 18.0519C2.86182 18.4463 3.35332 19.0718 3.57588 19.8059C4.23128 21.9678 6.93794 22.7185 8.63329 21.2086L8.87829 20.9904C10.0262 19.9681 11.7351 19.8647 13.0002 20.7411L13.2207 20.8938C15.0497 22.1607 17.589 21.5239 18.5874 19.5478L19.5841 17.5753C19.9823 16.7872 20.6706 16.1802 21.5074 15.879L21.8696 15.7487C24.0936 14.9484 24.7248 12.1332 23.0514 10.4773C22.0534 9.4897 21.8178 7.97576 22.4692 6.73673L22.8133 6.0821C24.17 3.50166 21.9811 0.487916 19.0712 0.899278C18.2916 1.0095 17.4922 0.858122 16.8103 0.468166Z`,
-  ],
-};
-
-export const dvProteinBackbone: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'proteinBackbone' as IconName,
-  icon: [
-    24,
-    18,
-    [],
-    null,
-    `M3.47472 11.5486C2.62747 10.7018 2.41571 10.0262 2.44306 9.54479C2.47063 9.05963 2.75193 8.57865 3.29892 8.14683C3.84567 7.7152 4.60047 7.38394 5.40483 7.21366C6.21113 7.04297 7.0041 7.04727 7.62078 7.22213L7.65794 7.23267L7.69597 7.23933C8.24462 7.33544 8.74122 7.38558 9.19858 7.29628C9.69694 7.19899 10.0717 6.9537 10.4099 6.61702L10.525 6.50246L10.5825 6.35054C11.3769 4.24813 11.9875 3.29079 12.4131 2.92133C12.5994 2.75959 12.69 2.76332 12.7078 2.76406C12.7452 2.76561 12.8379 2.78536 13.0066 2.91768C13.1757 3.05024 13.3594 3.24646 13.5761 3.50371C13.6473 3.58832 13.728 3.68688 13.8127 3.79046C13.9562 3.96582 14.1115 4.15576 14.253 4.31661C14.4894 4.58535 14.7715 4.87491 15.097 5.08606C15.4326 5.30372 15.8615 5.46768 16.3586 5.39711C17.0488 5.29912 17.6507 4.97154 18.1571 4.64755C18.3641 4.51516 18.5711 4.3725 18.7659 4.23821L18.8988 4.14675C19.1377 3.98274 19.3608 3.83328 19.5827 3.70369C20.0289 3.44305 20.4137 3.29738 20.7927 3.28763C21.1532 3.27835 21.5998 3.39018 22.175 3.82289C23.3411 4.70013 23.8911 5.48283 24.0885 6.11246C24.2756 6.70958 24.1706 7.2447 23.8392 7.74065C23.1271 8.80659 21.3568 9.67002 19.3561 9.86394L19.2984 9.86954L19.2422 9.88394C18.9717 9.95328 18.6096 10.0819 18.3846 10.4208C18.1748 10.7368 18.1938 11.087 18.2017 11.2321L18.2025 11.2474C18.2368 11.9069 18.4727 12.5978 18.6982 13.203C18.753 13.35 18.8075 13.4929 18.8607 13.6324C19.0417 14.1068 19.2077 14.5421 19.3195 14.9695C19.4632 15.519 19.4753 15.9164 19.3697 16.2109C19.2775 16.4678 19.049 16.7732 18.3989 17.0408C18.2555 17.0998 18.1139 17.1589 17.9739 17.2173C17.5033 17.4137 17.0515 17.6022 16.6169 17.7533C16.0424 17.953 15.6061 18.0463 15.3047 18.0298C15.166 18.0222 15.0912 17.9926 15.0514 17.9683C15.0187 17.9483 14.9746 17.9113 14.9304 17.8142C14.8246 17.5818 14.7493 17.07 14.9172 16.063C14.9853 15.6545 14.7093 15.268 14.3007 15.1999C13.8921 15.1318 13.5057 15.4078 13.4376 15.8164C13.2568 16.9011 13.2686 17.7838 13.5651 18.4355C13.7221 18.7803 13.958 19.0578 14.2688 19.248C14.5727 19.4338 14.9053 19.5102 15.2226 19.5276C15.8332 19.5611 16.5035 19.3808 17.1095 19.1701C17.5915 19.0025 18.1011 18.7898 18.5788 18.5904C18.712 18.5348 18.8429 18.4801 18.9699 18.4279C19.9008 18.0446 20.51 17.4743 20.7816 16.7175C21.0397 15.9981 20.9417 15.2436 20.7707 14.5899C20.6361 14.0753 20.4303 13.5367 20.2446 13.0507C20.1957 12.9228 20.1482 12.7985 20.1038 12.6792C19.9052 12.1462 19.7633 11.7043 19.7156 11.3338C21.8682 11.0785 24.0545 10.1186 25.0865 8.57393C25.6444 7.73888 25.8559 6.73628 25.5198 5.66386C25.1939 4.62394 24.3808 3.60527 23.0767 2.6242C22.2856 2.02908 21.5156 1.76852 20.7541 1.78812C20.011 1.80726 19.3683 2.09177 18.8261 2.40847C18.5538 2.56755 18.2925 2.74354 18.05 2.91002L17.9056 3.00933C17.7125 3.14244 17.5319 3.26685 17.3488 3.38398C16.8899 3.67754 16.5104 3.86052 16.1478 3.912C16.1185 3.91616 16.0527 3.91801 15.9133 3.8276C15.7639 3.73068 15.5907 3.56623 15.3793 3.32591C15.2485 3.17716 15.1322 3.03467 15.0073 2.88151C14.9197 2.77404 14.8277 2.66125 14.7234 2.53744C14.4952 2.26646 14.2294 1.97031 13.9322 1.73727C13.6347 1.50401 13.2443 1.28494 12.7697 1.26534C12.2756 1.24493 11.8253 1.44523 11.4298 1.78858C10.7112 2.41238 10.0122 3.64886 9.24229 5.65488C9.11183 5.76461 9.01491 5.80382 8.91115 5.82408C8.74333 5.85684 8.48463 5.8527 7.99039 5.768C7.09318 5.52254 6.06121 5.54146 5.09417 5.74618C4.11094 5.95432 3.13167 6.36778 2.36947 6.96949C1.60752 7.57101 1.00501 8.41198 0.945478 9.4597C0.885737 10.5112 1.38361 11.5794 2.41432 12.6096C2.73824 12.9333 3.24866 13.2865 3.73219 13.621C3.85865 13.7085 3.98327 13.7947 4.10224 13.8788C4.73391 14.3253 5.30217 14.7701 5.66196 15.2379C6.00378 15.6823 6.09269 16.0547 5.96 16.4468C5.80425 16.907 5.28789 17.5739 3.94155 18.4154C3.5903 18.6349 3.48352 19.0976 3.70306 19.4489C3.9226 19.8001 4.38531 19.9069 4.73656 19.6873C6.17912 18.7857 7.05722 17.8838 7.38083 16.9276C7.72748 15.9034 7.38063 15.012 6.85091 14.3233C6.33917 13.6581 5.60012 13.1007 4.96804 12.6539C4.80139 12.5361 4.64497 12.4276 4.49948 12.3266C4.04264 12.0096 3.69365 11.7675 3.47472 11.5486Z`,
-  ],
-};
-
-/*
- * Aevidence app icons
- */
-export const dvJunction: IconDefinition = {
-  prefix: 'dv' as IconPrefix,
-  iconName: 'junction' as IconName,
-  icon: [
-    24,
-    24,
-    [],
-    null,
-    `M9.90909 19.7647H0V24H9.90909V19.7647Z
-    M24 19.7647H14.0909V24H24V19.7647Z
-    M16.8868 4.85812C16.8868 6.69237 15.8643 8.2892 14.3549 9.11596C15.2396 10.7537 16.5039 12.1457 17.1972 12.7664C17.9371 13.3696 18.364 14.3022 18.6149 15.0594C18.8801 15.86 19.0126 16.6656 19.0552 17.1858C19.1038 17.78 18.6587 18.301 18.0609 18.3493C17.4631 18.3977 16.9391 17.9551 16.8904 17.3609C16.8599 16.9876 16.7565 16.3524 16.5519 15.7348C16.3358 15.0824 16.066 14.632 15.8128 14.4302L15.79 14.4121L15.7682 14.3928C14.926 13.6444 13.2884 11.871 12.2215 9.71133C12.1481 9.71459 12.0742 9.71623 12 9.71623C11.9258 9.71623 11.8519 9.71459 11.7785 9.71133C10.7116 11.871 9.074 13.6444 8.23175 14.3928L8.21 14.4121L8.18723 14.4302C7.93401 14.632 7.66419 15.0824 7.44808 15.7348C7.24346 16.3524 7.14014 16.9876 7.10959 17.3609C7.06094 17.9551 6.53691 18.3977 5.93913 18.3493C5.34134 18.301 4.89617 17.78 4.94481 17.1858C4.98739 16.6656 5.11991 15.86 5.38514 15.0594C5.63596 14.3022 6.06293 13.3696 6.80283 12.7664C7.49615 12.1457 8.76044 10.7537 9.64505 9.11596C8.13572 8.2892 7.11316 6.69237 7.11316 4.85812C7.11316 2.17505 9.30107 0 12 0C14.6989 0 16.8868 2.17505 16.8868 4.85812Z
-    `,
-  ],
-};
