+++ src/icons/factories/visyn-font-awesome-icon.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+import { IconProp } from '@fortawesome/fontawesome-svg-core';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+import { resolveIconSize } from '../size-map';
+import { VisynIconProps } from '../types';
+
+export function VisynFontAwesomeIcon({ size = 'md', icon }: VisynIconProps & { icon: IconProp }) {
+  const pixelSize = resolveIconSize(size);
+
+  return <FontAwesomeIcon icon={icon} style={{ width: pixelSize, height: pixelSize }} />;
+}
