+++ src/icons/factories/visyn-lucide-icon.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+import { LucideIcon } from 'lucide-react';
+
+import { resolveIconSize } from '../size-map';
+import { VisynIconProps } from '../types';
+
+export function VisynLucideIcon({ size = 'md', icon: Icon }: VisynIconProps & { icon: LucideIcon }) {
+  const resolvedSize = resolveIconSize(size);
+
+  return <Icon size={resolvedSize} />;
+}
