+++ src/icons/font-awesome/icon-add-visualization.ts
@@ -1,0 +1,17 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvAddVisualization: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'addVisualization' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M335.181 113.681H335.681V113.181V54.3191C335.681 34.5367 351.718 18.5 371.5 18.5C391.282 18.5 407.319 34.5367 407.319 54.3191V113.181V113.681H407.819H466.681C486.463 113.681 502.5 129.718 502.5 149.5C502.5 169.282 486.463 185.319 466.681 185.319H407.819H407.319V185.819V244.681C407.319 264.463 391.282 280.5 371.5 280.5C351.718 280.5 335.681 264.463 335.681 244.681V185.819V185.319H335.181H276.319C256.537 185.319 240.5 169.282 240.5 149.5C240.5 129.718 256.537 113.681 276.319 113.681H335.181ZM10.5 39.125C10.5 24.4031 22.0038 12.5 36.1566 12.5C50.3093 12.5 61.8132 24.4031 61.8132 39.125V436.435L61.8128 436.441C61.8122 436.455 61.8113 436.474 61.8103 436.499C61.8083 436.549 61.806 436.621 61.8043 436.714C61.8011 436.899 61.8007 437.165 61.8117 437.494C61.8337 438.152 61.9013 439.068 62.0845 440.1C62.4487 442.152 63.2802 444.727 65.179 446.634C67.0801 448.543 69.6474 449.375 71.694 449.736C72.7231 449.918 73.6357 449.984 74.2919 450.004C74.6204 450.015 74.8855 450.013 75.0699 450.01C75.1621 450.008 75.2342 450.005 75.284 450.003L75.3113 450.002L75.3419 450L75.3473 450H472.971C487.637 450 499.5 461.542 499.5 475.75C499.5 489.958 487.637 501.5 472.971 501.5H61.446H61.4413L61.432 501.5L61.3748 501.498C61.3238 501.497 61.2476 501.495 61.1473 501.492C60.9466 501.486 60.6498 501.475 60.2672 501.457C59.502 501.421 58.3941 501.356 57.027 501.242C54.292 501.013 50.523 500.585 46.387 499.791C38.087 498.197 28.4193 495.146 22.6028 489.355C16.6702 483.448 13.6281 473.531 12.0805 464.999C11.3093 460.747 10.914 456.869 10.7116 454.053C10.6104 452.646 10.5575 451.505 10.5299 450.716C10.5161 450.322 10.5086 450.016 10.5046 449.81C10.5026 449.706 10.5014 449.628 10.5008 449.575L10.5001 449.516L10.5 449.502L10.5 449.5V449.499V39.125ZM130.072 75.5L217.182 75.5C209.257 91.3114 203.759 108.543 201.176 126.71L130.072 126.71C115.379 126.71 103.5 115.229 103.5 101.105C103.5 86.9808 115.379 75.5 130.072 75.5ZM130.072 171.445H200.797C203.058 189.569 208.214 206.797 215.789 222.655L130.072 222.655C115.379 222.655 103.5 211.174 103.5 197.05C103.5 182.926 115.379 171.445 130.072 171.445ZM130.072 267.39H246.973C261.912 282.763 279.772 295.299 299.662 304.109C295.37 312.678 286.266 318.6 275.717 318.6H130.072C115.379 318.6 103.5 307.119 103.5 292.995C103.5 278.87 115.379 267.39 130.072 267.39ZM439.691 388.94C439.691 403.064 427.812 414.545 413.119 414.545H130.072C115.379 414.545 103.5 403.064 103.5 388.94C103.5 374.815 115.379 363.335 130.072 363.335H413.119C427.812 363.335 439.691 374.815 439.691 388.94Z',
+  ],
+};
+
+export const IconAddVisualization = createFontAwesomeIcon(dvAddVisualization);
