+++ src/icons/font-awesome/icon-boolean-array-column.ts
@@ -1,0 +1,21 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvBooleanArrayCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'booleanArrayCol' as IconName,
+  icon: [
+    512,
+    308,
+    [],
+    '',
+    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
+    M234.512 297.482L247.18 248.908C242.59 247.453 238.643 244.724 235.338 240.721C232.034 236.537 230.381 231.352 230.381 225.167C230.381 217.89 232.768 211.886 237.541 207.156C242.498 202.426 248.741 200.061 256.268 200.061C263.795 200.061 269.946 202.517 274.719 207.429C279.676 212.159 282.155 218.344 282.155 225.985C282.155 231.625 281.42 236.901 279.951 241.813C278.483 246.725 276.28 252.91 273.342 260.369L258.746 297.482H234.512Z
+    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z
+    M150 68C196.956 68 235 106.044 235 153C235 199.956 196.956 238 150 238C103.044 238 65 199.956 65 153C65 106.044 103.044 68 150 68Z
+    M452 153C452 103.283 411.717 63 362 63C312.283 63 272 103.283 272 153C272 202.717 312.283 243 362 243C411.717 243 452 202.717 452 153ZM425.548 153C425.548 188.267 396.956 216.548 362 216.548C326.717 216.548 298.452 188.283 298.452 153C298.452 118.044 326.733 89.4516 362 89.4516C396.94 89.4516 425.548 118.06 425.548 153Z`,
+  ],
+};
+
+export const IconBooleanArrayColumn = createFontAwesomeIcon(dvBooleanArrayCol);
