+++ src/icons/font-awesome/icon-boxplot.ts
@@ -1,0 +1,21 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvBoxplot: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'boxplot' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M75.5717 450.288C75.5944 450.287 75.613 450.286 75.6274 450.286H474.861C489.325 450.286 501 461.666 501 475.643C501 489.62 489.325 501 474.861 501H61.6512V502C61.669 501 61.6691 501 61.6691 501H61.6685L61.6656 501L61.6524 501L61.5967 500.999C61.5467 500.997 61.4713 500.995 61.3718 500.992C61.1727 500.986 60.8773 500.975 60.4958 500.957C59.7329 500.921 58.6265 500.857 57.2604 500.742C54.5269 500.514 50.7608 500.086 46.6294 499.293C38.3109 497.695 28.7333 494.653 23.0055 488.95C17.1644 483.135 14.1302 473.311 12.5789 464.759C11.8086 460.512 11.4135 456.636 11.2112 453.822C11.1101 452.415 11.0572 451.276 11.0297 450.49C11.0159 450.096 11.0085 449.792 11.0045 449.587C11.0025 449.484 11.0014 449.406 11.0007 449.355L11.0001 449.297L11 449.284L11 449.283V37.2357C11 22.7116 22.3447 11 36.2633 11C50.1819 11 61.5267 22.7116 61.5267 37.2357V436.155C61.526 436.169 61.5251 436.189 61.5242 436.212C61.5221 436.266 61.5196 436.342 61.5179 436.439C61.5145 436.632 61.5141 436.907 61.5255 437.247C61.5481 437.926 61.6179 438.87 61.8068 439.934C62.1804 442.04 63.0416 444.741 65.0514 446.759C67.0641 448.78 69.7586 449.642 71.8591 450.013C72.921 450.201 73.8619 450.269 74.5389 450.29C74.878 450.3 75.1528 450.299 75.3456 450.295C75.4421 450.293 75.5182 450.291 75.5717 450.288Z
+    M275 336.162V118.4C275 113.443 270.971 109.424 266 109.424H208.4V75.3143L231.8 75.3143C236.771 75.3143 240.8 71.2955 240.8 66.3381V51.9762C240.8 47.0188 236.771 43 231.8 43L150.8 43C145.829 43 141.8 47.0188 141.8 51.9762V66.3381C141.8 71.2955 145.829 75.3143 150.8 75.3143H176V109.424H113C108.029 109.424 104 113.443 104 118.4V336.162C104 341.12 108.029 345.139 113 345.139H176V387.686H152.6C147.629 387.686 143.6 391.704 143.6 396.662V411.024C143.6 415.981 147.629 420 152.6 420H233.6C238.571 420 242.6 415.981 242.6 411.024V396.662C242.6 391.704 238.571 387.686 233.6 387.686H208.4V345.139H266C270.971 345.139 275 341.12 275 336.162Z
+    M490 248.367V129.837C490 124.877 485.994 120.857 481.053 120.857H423.789V46.3265L447.053 46.3265C451.994 46.3265 456 42.3062 456 37.3469V22.9796C456 18.0203 451.994 14 447.053 14L366.526 14C361.585 14 357.579 18.0203 357.579 22.9796V37.3469C357.579 42.3062 361.585 46.3265 366.526 46.3265H391.579V120.857H328.947C324.006 120.857 320 124.877 320 129.837V248.367C320 253.327 324.006 257.347 328.947 257.347H391.579V333.673H368.316C363.374 333.673 359.368 337.694 359.368 342.653V357.02C359.368 361.98 363.374 366 368.316 366H448.842C453.784 366 457.789 361.98 457.789 357.02V342.653C457.789 337.694 453.784 333.673 448.842 333.673H423.789V257.347H481.053C485.994 257.347 490 253.327 490 248.367Z
+    `,
+  ],
+};
+
+export const IconBoxplot = createFontAwesomeIcon(dvBoxplot);
