+++ src/icons/font-awesome/icon-categorical-array-column.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvCategoricalArrayCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'categoricalArrayCol' as IconName,
+  icon: [
+    512,
+    307,
+    [],
+    '',
+    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
+    M246.512 297.482L259.18 248.908C254.59 247.453 250.643 244.724 247.338 240.721C244.034 236.537 242.381 231.352 242.381 225.167C242.381 217.89 244.768 211.886 249.541 207.156C254.498 202.426 260.741 200.061 268.268 200.061C275.795 200.061 281.946 202.517 286.719 207.429C291.676 212.159 294.155 218.344 294.155 225.985C294.155 231.625 293.42 236.901 291.951 241.813C290.483 246.725 288.28 252.91 285.342 260.369L270.746 297.482H246.512Z
+    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z
+    M65 103.786C65 89.0536 76.3839 77 91.7857 77H188.214C202.946 77 215 89.0536 215 103.786V200.214C215 215.616 202.946 227 188.214 227H91.7857C76.3839 227 65 215.616 65 200.214V103.786Z
+    M436.286 65.4136C429.589 61.2008 421.554 61.2008 414.857 65.4136L307.714 132.818C301.018 137.031 297 144.052 297 151.776C297 160.202 301.018 167.223 307.714 171.436L414.857 238.84C421.554 243.053 429.589 243.053 436.286 238.84C442.982 235.33 447 227.606 447 219.181L447 84.3712C447 76.6477 442.982 68.9242 436.286 65.4136Z`,
+  ],
+};
+export const IconCategoricalArrayColumn = createFontAwesomeIcon(dvCategoricalArrayCol);
