+++ src/icons/font-awesome/icon-cell-line.ts
@@ -1,0 +1,18 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvCellLine: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'cellline' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M469.975 290.738V289.646L469.148 290.36C457.706 300.24 444.632 308.85 430.816 316.134C385.003 340.288 323.174 354.595 256 354.595C188.826 354.595 126.997 340.288 81.1842 316.134C67.3678 308.85 54.2937 300.24 42.8517 290.36L42.0249 289.646V290.738V304.452C42.0249 324.297 58.0639 349.388 98.8468 370.891C137.992 391.529 193.532 404.927 256 404.927C318.468 404.927 374.008 391.529 413.153 370.891C453.936 349.388 469.975 324.297 469.975 304.452V290.738ZM511.5 207.548V304.452C511.5 352.956 474.996 389.746 430.816 413.039C385.003 437.193 323.174 451.5 256 451.5C188.826 451.5 126.997 437.193 81.1842 413.039C37.0036 389.746 0.5 352.956 0.5 304.452V207.548C0.5 159.044 37.0036 122.254 81.1842 98.9609C126.997 74.807 188.826 60.5 256 60.5C323.174 60.5 385.003 74.807 430.816 98.9609C474.996 122.254 511.5 159.044 511.5 207.548ZM98.8468 141.109C58.0639 162.612 42.0249 187.703 42.0249 207.548C42.0249 227.392 58.0639 252.484 98.8468 273.986C137.992 294.624 193.532 308.022 256 308.022C318.468 308.022 374.008 294.624 413.153 273.986C453.936 252.484 469.975 227.392 469.975 207.548C469.975 187.703 453.936 162.612 413.153 141.109C374.008 120.471 318.468 107.073 256 107.073C193.532 107.073 137.992 120.471 98.8468 141.109ZM256 176.529C191.033 176.529 132.457 196.762 91.7205 220.164C83.6487 224.801 73.7058 221.268 69.5222 212.154C65.3273 203.015 68.5683 191.835 76.6782 187.176C121.514 161.419 185.21 139.471 256 139.471C326.79 139.471 390.486 161.419 435.322 187.176C443.432 191.835 446.673 203.015 442.478 212.154C438.294 221.268 428.351 224.801 420.28 220.164C379.543 196.762 320.967 176.529 256 176.529ZM186.61 234.117C186.61 237.305 185.564 241.255 183.783 245.408C182.006 249.554 179.512 253.868 176.651 257.777C173.79 261.687 170.573 265.175 167.358 267.681C164.13 270.197 160.964 271.675 158.193 271.675C152.586 271.675 147.402 267.299 143.583 261.139C139.778 255.001 137.43 247.241 137.43 240.777C137.43 234.318 139.774 227.314 143.567 221.931C147.365 216.54 152.548 212.862 158.193 212.862C163.96 212.862 171.094 214.984 176.78 218.731C182.47 222.48 186.61 227.785 186.61 234.117ZM262.304 272.175C262.304 278.801 258.341 284.857 252.86 289.278C247.38 293.699 240.471 296.413 234.738 296.413C229.01 296.413 222.531 293.704 217.48 289.291C212.432 284.88 208.872 278.822 208.872 272.175C208.872 265.501 212.456 257.761 217.535 251.673C222.626 245.57 229.09 241.277 234.738 241.277C237.553 241.277 240.687 242.343 243.823 244.192C246.954 246.039 250.057 248.648 252.803 251.688C258.311 257.783 262.304 265.518 262.304 272.175ZM363.513 234.117C363.513 240.472 359.559 246.288 354.077 250.541C348.598 254.792 341.686 257.403 335.947 257.403C330.198 257.403 322.426 254.785 316.088 250.521C309.735 246.248 304.978 240.432 304.978 234.117C304.978 231.04 306.78 228.165 309.799 225.553C312.812 222.947 316.972 220.663 321.529 218.775C330.647 214.997 341.212 212.862 347.003 212.862C352.737 212.862 356.83 214.957 359.509 218.609C362.206 222.286 363.513 227.603 363.513 234.117ZM469.975 290.738V289.646L469.148 290.36C457.706 300.24 444.632 308.85 430.816 316.134C385.003 340.288 323.174 354.595 256 354.595C188.826 354.595 126.997 340.288 81.1842 316.134C67.3678 308.85 54.2937 300.24 42.8517 290.36L42.0249 289.646V290.738V304.452C42.0249 324.297 58.0639 349.388 98.8468 370.891C137.992 391.529 193.532 404.927 256 404.927C318.468 404.927 374.008 391.529 413.153 370.891C453.936 349.388 469.975 324.297 469.975 304.452V290.738ZM511.5 207.548V304.452C511.5 352.956 474.996 389.746 430.816 413.039C385.003 437.193 323.174 451.5 256 451.5C188.826 451.5 126.997 437.193 81.1842 413.039C37.0036 389.746 0.5 352.956 0.5 304.452V207.548C0.5 159.044 37.0036 122.254 81.1842 98.9609C126.997 74.807 188.826 60.5 256 60.5C323.174 60.5 385.003 74.807 430.816 98.9609C474.996 122.254 511.5 159.044 511.5 207.548ZM98.8468 141.109C58.0639 162.612 42.0249 187.703 42.0249 207.548C42.0249 227.392 58.0639 252.484 98.8468 273.986C137.992 294.624 193.532 308.022 256 308.022C318.468 308.022 374.008 294.624 413.153 273.986C453.936 252.484 469.975 227.392 469.975 207.548C469.975 187.703 453.936 162.612 413.153 141.109C374.008 120.471 318.468 107.073 256 107.073C193.532 107.073 137.992 120.471 98.8468 141.109ZM256 176.529C191.033 176.529 132.457 196.762 91.7205 220.164C83.6487 224.801 73.7058 221.268 69.5222 212.154C65.3273 203.015 68.5683 191.835 76.6782 187.176C121.514 161.419 185.21 139.471 256 139.471C326.79 139.471 390.486 161.419 435.322 187.176C443.432 191.835 446.673 203.015 442.478 212.154C438.294 221.268 428.351 224.801 420.28 220.164C379.543 196.762 320.967 176.529 256 176.529ZM186.61 234.117C186.61 237.305 185.564 241.255 183.783 245.408C182.006 249.554 179.512 253.868 176.651 257.777C173.79 261.687 170.573 265.175 167.358 267.681C164.13 270.197 160.964 271.675 158.193 271.675C152.586 271.675 147.402 267.299 143.583 261.139C139.778 255.001 137.43 247.241 137.43 240.777C137.43 234.318 139.774 227.314 143.567 221.931C147.365 216.54 152.548 212.862 158.193 212.862C163.96 212.862 171.094 214.984 176.78 218.731C182.47 222.48 186.61 227.785 186.61 234.117ZM262.304 272.175C262.304 278.801 258.341 284.857 252.86 289.278C247.38 293.699 240.471 296.413 234.738 296.413C229.01 296.413 222.531 293.704 217.48 289.291C212.432 284.88 208.872 278.822 208.872 272.175C208.872 265.501 212.456 257.761 217.535 251.673C222.626 245.57 229.09 241.277 234.738 241.277C237.553 241.277 240.687 242.343 243.823 244.192C246.954 246.039 250.057 248.648 252.803 251.688C258.311 257.783 262.304 265.518 262.304 272.175ZM363.513 234.117C363.513 240.472 359.559 246.288 354.077 250.541C348.598 254.792 341.686 257.403 335.947 257.403C330.198 257.403 322.426 254.785 316.088 250.521C309.735 246.248 304.978 240.432 304.978 234.117C304.978 231.04 306.78 228.165 309.799 225.553C312.812 222.947 316.972 220.663 321.529 218.775C330.647 214.997 341.212 212.862 347.003 212.862C352.737 212.862 356.83 214.957 359.509 218.609C362.206 222.286 363.513 227.603 363.513 234.117Z`,
+  ],
+};
+
+export const IconCellLine = createFontAwesomeIcon(dvCellLine);
