+++ src/icons/font-awesome/icon-chemical-structure.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvChemicalStructure: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'chemicalStructure' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M253.549 213.538L298.645 142.445C304.45 133.293 314.697 127.934 325.525 128.385L350.626 129.431L350.782 129.438L407.553 131.804L443.531 75.8179C443.153 74.9596 442.816 74.0922 442.521 73.2182L442.512 73.1936C437.697 58.9103 443.843 42.8479 457.622 35.6405C472.88 27.6596 491.719 33.5587 499.7 48.8166C507.681 64.0745 501.782 82.9133 486.524 90.8943C481.76 93.3861 476.647 94.5249 471.625 94.4482L434.864 151.652C429.039 160.715 418.845 166.01 408.08 165.561L326.065 162.143L279.642 235.328L311.104 295.476C313.337 299.744 313.122 304.879 310.542 308.947L265.56 379.863C262.98 383.93 258.426 386.313 253.613 386.112L181.818 383.119L147.026 437.967C152.377 452.455 146.262 469.011 132.213 476.36C116.955 484.341 98.1161 478.441 90.1351 463.184C82.1541 447.926 88.0533 429.087 103.311 421.106C108.645 418.316 114.417 417.222 120.01 417.632L153.116 365.441L123.717 309.237L66.2436 306.842C63.4104 311.308 59.4013 315.113 54.3776 317.741C39.1197 325.722 20.2809 319.823 12.3 304.565C4.319 289.307 10.2181 270.468 25.476 262.487C39.8473 254.97 57.3953 259.766 66.0552 273.099L127.475 275.658L164.942 216.591C167.522 212.523 172.076 210.141 176.888 210.342L253.549 213.538ZM247.725 247.03L187.15 244.505L154.676 295.701L182.776 349.423L243.351 351.949L275.825 300.752L247.725 247.03Z
+    M253.549 213.538L298.645 142.445C304.45 133.293 314.697 127.934 325.525 128.385L350.626 129.431L350.782 129.438L407.553 131.804L443.531 75.8179C443.153 74.9596 442.816 74.0922 442.521 73.2182L442.512 73.1936C437.697 58.9103 443.843 42.8479 457.622 35.6405C472.88 27.6596 491.719 33.5587 499.7 48.8166C507.681 64.0745 501.782 82.9133 486.524 90.8943C481.76 93.3861 476.647 94.5249 471.625 94.4482L434.864 151.652C429.039 160.715 418.845 166.01 408.08 165.561L326.065 162.143L279.642 235.328L311.104 295.476C313.337 299.744 313.122 304.879 310.542 308.947L265.56 379.863C262.98 383.93 258.426 386.313 253.613 386.112L181.818 383.119L147.026 437.967C152.377 452.455 146.262 469.011 132.213 476.36C116.955 484.341 98.1161 478.441 90.1351 463.184C82.1541 447.926 88.0533 429.087 103.311 421.106C108.645 418.316 114.417 417.222 120.01 417.632L153.116 365.441L123.717 309.237L66.2436 306.842C63.4104 311.308 59.4013 315.113 54.3776 317.741C39.1197 325.722 20.2809 319.823 12.3 304.565C4.319 289.307 10.2181 270.468 25.476 262.487C39.8473 254.97 57.3953 259.766 66.0552 273.099L127.475 275.658L164.942 216.591C167.522 212.523 172.076 210.141 176.888 210.342L253.549 213.538ZM247.725 247.03L187.15 244.505L154.676 295.701L182.776 349.423L243.351 351.949L275.825 300.752L247.725 247.03Z
+    M298.645 142.445L294.423 139.767L294.423 139.767L298.645 142.445ZM253.549 213.538L253.341 218.534L256.225 218.654L257.771 216.216L253.549 213.538ZM325.525 128.385L325.733 123.389L325.733 123.389L325.525 128.385ZM350.626 129.431L350.418 134.427L350.418 134.427L350.626 129.431ZM350.782 129.438L350.99 124.442L350.99 124.442L350.782 129.438ZM407.553 131.804L407.345 136.799L410.209 136.919L411.759 134.507L407.553 131.804ZM443.531 75.8179L447.738 78.521L449.191 76.26L448.106 73.8007L443.531 75.8179ZM442.521 73.2182L447.257 71.6168L447.257 71.6147L442.521 73.2182ZM442.512 73.1936L437.774 74.7911L437.777 74.7971L442.512 73.1936ZM457.622 35.6405L459.94 40.071L459.94 40.071L457.622 35.6405ZM499.7 48.8166L504.13 46.4992L504.13 46.4991L499.7 48.8166ZM486.524 90.8943L484.206 86.4638L484.206 86.4638L486.524 90.8943ZM471.625 94.4482L471.701 89.4488L468.921 89.4064L467.418 91.7451L471.625 94.4482ZM434.864 151.652L430.657 148.949L430.657 148.949L434.864 151.652ZM408.08 165.561L408.289 160.565L408.289 160.565L408.08 165.561ZM326.065 162.143L326.274 157.147L323.39 157.027L321.843 159.465L326.065 162.143ZM279.642 235.328L275.42 232.65L273.874 235.087L275.212 237.645L279.642 235.328ZM311.104 295.476L306.673 297.793L306.673 297.793L311.104 295.476ZM310.542 308.947L306.32 306.268L306.32 306.268L310.542 308.947ZM265.56 379.863L261.337 377.185L261.337 377.185L265.56 379.863ZM253.613 386.112L253.821 381.116L253.821 381.116L253.613 386.112ZM181.818 383.119L182.026 378.123L179.142 378.003L177.595 380.44L181.818 383.119ZM147.026 437.967L142.804 435.289L141.478 437.378L142.336 439.699L147.026 437.967ZM132.213 476.36L129.895 471.929L129.895 471.929L132.213 476.36ZM103.311 421.106L105.629 425.536L105.629 425.536L103.311 421.106ZM120.01 417.632L119.645 422.618L122.629 422.837L124.232 420.31L120.01 417.632ZM153.116 365.441L157.338 368.119L158.884 365.681L157.546 363.123L153.116 365.441ZM123.717 309.237L128.148 306.92L126.81 304.362L123.926 304.241L123.717 309.237ZM66.2436 306.842L66.4518 301.846L63.5676 301.726L62.0214 304.164L66.2436 306.842ZM54.3776 317.741L56.6951 322.171L56.6951 322.171L54.3776 317.741ZM12.3 304.565L7.86946 306.882L7.86946 306.882L12.3 304.565ZM25.476 262.487L27.7935 266.917L27.7935 266.917L25.476 262.487ZM66.0552 273.099L61.8621 275.822L63.2681 277.987L65.847 278.094L66.0552 273.099ZM127.475 275.658L127.266 280.654L130.151 280.774L131.697 278.337L127.475 275.658ZM164.942 216.591L169.164 219.269L169.164 219.269L164.942 216.591ZM176.888 210.342L176.68 215.338L176.68 215.338L176.888 210.342ZM187.15 244.505L187.358 239.509L184.474 239.389L182.928 241.827L187.15 244.505ZM247.725 247.03L252.155 244.713L250.817 242.155L247.933 242.035L247.725 247.03ZM154.676 295.701L150.454 293.023L148.907 295.461L150.245 298.019L154.676 295.701ZM182.776 349.423L178.346 351.741L179.684 354.299L182.568 354.419L182.776 349.423ZM243.351 351.949L243.143 356.944L246.027 357.065L247.573 354.627L243.351 351.949ZM275.825 300.752L280.047 303.43L281.594 300.993L280.256 298.435L275.825 300.752ZM294.423 139.767L249.327 210.86L257.771 216.216L302.867 145.123L294.423 139.767ZM325.733 123.389C313.12 122.864 301.185 129.107 294.423 139.767L302.867 145.123C307.716 137.48 316.273 133.004 325.317 133.381L325.733 123.389ZM350.834 124.435L325.733 123.389L325.317 133.381L350.418 134.427L350.834 124.435ZM350.99 124.442L350.834 124.435L350.418 134.427L350.574 134.433L350.99 124.442ZM407.761 126.808L350.99 124.442L350.574 134.433L407.345 136.799L407.761 126.808ZM439.325 73.1148L403.347 129.101L411.759 134.507L447.738 78.521L439.325 73.1148ZM437.784 74.8195C438.127 75.8341 438.518 76.8403 438.956 77.8351L448.106 73.8007C447.788 73.079 447.505 72.3504 447.257 71.6168L437.784 74.8195ZM437.777 74.7971L437.785 74.8216L447.257 71.6147L447.248 71.5902L437.777 74.7971ZM455.305 31.21C439.312 39.5754 432.185 58.2131 437.774 74.7911L447.25 71.5962C443.208 59.6074 448.375 46.1204 459.94 40.071L455.305 31.21ZM504.13 46.4991C494.869 28.7944 473.01 21.9492 455.305 31.21L459.94 40.071C472.751 33.37 488.568 38.3231 495.269 51.1341L504.13 46.4991ZM488.841 95.3248C506.546 86.0639 513.391 64.204 504.13 46.4992L495.269 51.1341C501.97 63.9451 497.017 79.7627 484.206 86.4638L488.841 95.3248ZM471.548 99.4476C477.383 99.5367 483.321 98.2122 488.841 95.3248L484.206 86.4638C480.199 88.56 475.911 89.513 471.701 89.4488L471.548 99.4476ZM439.07 154.355L475.831 97.1513L467.418 91.7451L430.657 148.949L439.07 154.355ZM407.872 170.557C420.411 171.079 432.286 164.912 439.07 154.355L430.657 148.949C425.793 156.518 417.279 160.94 408.289 160.565L407.872 170.557ZM325.857 167.138L407.872 170.557L408.289 160.565L326.274 157.147L325.857 167.138ZM283.864 238.006L330.288 164.821L321.843 159.465L275.42 232.65L283.864 238.006ZM315.534 293.159L284.073 233.01L275.212 237.645L306.673 297.793L315.534 293.159ZM314.765 311.625C318.302 306.049 318.595 299.01 315.534 293.158L306.673 297.793C308.078 300.479 307.943 303.709 306.32 306.268L314.765 311.625ZM269.782 382.541L314.765 311.625L306.32 306.268L261.337 377.185L269.782 382.541ZM253.405 391.108C260.002 391.383 266.245 388.117 269.782 382.541L261.337 377.185C259.714 379.744 256.849 381.242 253.821 381.116L253.405 391.108ZM181.609 388.114L253.405 391.108L253.821 381.116L182.026 378.123L181.609 388.114ZM151.248 440.645L186.04 385.797L177.595 380.44L142.804 435.289L151.248 440.645ZM134.53 480.79C150.837 472.261 157.927 453.051 151.716 436.235L142.336 439.699C146.827 451.859 141.688 465.761 129.895 471.929L134.53 480.79ZM85.7046 465.501C94.9654 483.206 116.825 490.051 134.53 480.79L129.895 471.929C117.084 478.63 101.267 473.677 94.5656 460.866L85.7046 465.501ZM100.994 416.675C83.2889 425.936 76.4437 447.796 85.7046 465.501L94.5656 460.866C87.8645 448.055 92.8176 432.237 105.629 425.536L100.994 416.675ZM120.375 412.645C113.881 412.169 107.176 413.441 100.994 416.675L105.629 425.536C110.114 423.19 114.954 422.275 119.645 422.618L120.375 412.645ZM148.894 362.763L115.788 414.953L124.232 420.31L157.338 368.119L148.894 362.763ZM119.287 311.555L148.685 367.758L157.546 363.123L128.148 306.92L119.287 311.555ZM66.0354 311.838L123.509 314.233L123.926 304.241L66.4518 301.846L66.0354 311.838ZM56.6951 322.171C62.517 319.126 67.175 314.708 70.4658 309.52L62.0214 304.164C59.6457 307.909 56.2855 311.1 52.0602 313.31L56.6951 322.171ZM7.86946 306.882C17.1303 324.587 38.9903 331.432 56.6951 322.171L52.0602 313.31C39.2492 320.011 23.4315 315.058 16.7305 302.247L7.86946 306.882ZM23.1586 258.056C5.45377 267.317 -1.3914 289.177 7.86946 306.882L16.7305 302.247C10.0294 289.436 14.9825 273.619 27.7935 266.917L23.1586 258.056ZM70.2483 270.375C60.1967 254.9 39.837 249.332 23.1586 258.056L27.7935 266.917C39.8575 260.607 54.5939 264.633 61.8621 275.822L70.2483 270.375ZM127.683 270.663L66.2634 268.103L65.847 278.094L127.266 280.654L127.683 270.663ZM160.719 213.913L123.252 272.98L131.697 278.337L169.164 219.269L160.719 213.913ZM177.097 205.346C170.499 205.071 164.256 208.337 160.719 213.913L169.164 219.269C170.787 216.71 173.652 215.211 176.68 215.338L177.097 205.346ZM253.757 208.542L177.097 205.346L176.68 215.338L253.341 218.534L253.757 208.542ZM186.942 249.501L247.517 252.026L247.933 242.035L187.358 239.509L186.942 249.501ZM158.898 298.38L191.372 247.183L182.928 241.827L150.454 293.023L158.898 298.38ZM187.207 347.106L159.106 293.384L150.245 298.019L178.346 351.741L187.207 347.106ZM243.559 346.953L182.984 344.428L182.568 354.419L243.143 356.944L243.559 346.953ZM271.603 298.074L239.129 349.27L247.573 354.627L280.047 303.43L271.603 298.074ZM243.294 249.348L271.395 303.07L280.256 298.435L252.155 244.713L243.294 249.348Z
+    `,
+  ],
+};
+export const IconChemicalStructure = createFontAwesomeIcon(dvChemicalStructure);
