+++ src/icons/font-awesome/icon-clear-selection.ts
@@ -1,0 +1,23 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvClearSelection: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'select' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `M199.875 0C184.407 3.52605 172.865 17.3658 172.865 33.9015C172.865 53.1049 188.432 68.6723 207.636 68.6723H303.495C322.699 68.6723 338.266 53.1049 338.266 33.9015C338.266 17.3658 326.723 3.52605 311.256 0H199.875Z
+      M413.715 0C398.248 3.52605 386.705 17.3658 386.705 33.9015C386.705 53.1049 402.273 68.6723 421.476 68.6723H443.328V90.524C443.328 109.727 458.895 125.295 478.098 125.295C494.634 125.295 508.474 113.752 512 98.2845V33.9023C508.543 16.8715 495.128 3.45674 478.098 0H413.715Z
+      M512 200.744C508.474 185.277 494.634 173.734 478.098 173.734C458.895 173.734 443.328 189.302 443.328 208.505V304.364C443.328 323.568 458.895 339.135 478.098 339.135C494.634 339.135 508.474 327.593 512 312.125V200.744Z
+      M314.382 512C328.253 507.431 338.266 494.369 338.266 478.968C338.266 459.764 322.699 444.197 303.495 444.197H207.636C188.432 444.197 172.865 459.764 172.865 478.968C172.865 494.369 182.877 507.431 196.748 512H314.382Z
+      M100.542 512C114.413 507.431 124.426 494.369 124.426 478.968C124.426 459.764 108.858 444.197 89.6547 444.197H67.8031V422.345C67.8031 403.142 52.2356 387.574 33.0323 387.574C17.6314 387.574 4.56916 397.587 0 411.458V482.485C4.14462 496.672 15.3281 507.855 29.5148 512H100.542Z
+      M0 315.252C4.56916 329.123 17.6314 339.135 33.0323 339.135C52.2356 339.135 67.8031 323.568 67.8031 304.364V208.505C67.8031 189.301 52.2356 173.734 33.0323 173.734C17.6314 173.734 4.56916 183.747 0 197.618V315.252Z
+      M472 512C482.237 512 492.474 508.095 500.284 500.284C515.905 484.663 515.905 459.337 500.284 443.716L69.2843 12.7157C53.6633 -2.90524 28.3367 -2.90524 12.7157 12.7157C-2.90524 28.3367 -2.90524 53.6633 12.7157 69.2843L443.716 500.284C451.526 508.095 461.763 512 472 512Z
+      `,
+  ],
+};
+export const IconClearSelection = createFontAwesomeIcon(dvClearSelection);
