+++ src/icons/font-awesome/icon-correlation-plot.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvCorrelationplot: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'correlationplot' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M61.6424 502.5H61.6512H474.861C490.112 502.5 502.5 490.489 502.5 475.643C502.5 460.796 490.112 448.786 474.861 448.786H75.6H75.5843L75.5689 448.787L75.5665 448.787L75.5555 448.787L75.5078 448.79C75.4648 448.792 75.3998 448.794 75.3149 448.796C75.1452 448.799 74.8963 448.8 74.5854 448.791C73.963 448.771 73.0959 448.709 72.1202 448.536C70.157 448.189 67.8143 447.408 66.1143 445.701C64.4153 443.994 63.6334 441.643 63.2837 439.672C63.1099 438.693 63.0455 437.822 63.0246 437.197C63.0142 436.885 63.0146 436.636 63.0176 436.465C63.0191 436.38 63.0213 436.315 63.023 436.271L63.0251 436.223L63.0257 436.212L63.0258 436.21L63.0258 436.21L63.0267 436.195V436.181V37.2357C63.0267 21.935 51.0613 9.5 36.2633 9.5C21.4654 9.5 9.49997 21.935 9.49997 37.2357L9.49996 449.286L9.49998 449.288L9.99997 449.286C9.49998 449.288 9.49998 449.289 9.49998 449.289L9.49999 449.29L9.50002 449.294L9.50015 449.311L9.50082 449.373C9.50149 449.428 9.50268 449.51 9.50475 449.616C9.50889 449.829 9.51653 450.141 9.53059 450.542C9.55869 451.344 9.61243 452.502 9.71499 453.929C9.92008 456.782 10.3206 460.714 11.103 465.027C12.6623 473.623 15.7577 483.851 21.9472 490.013C28.0137 496.053 37.9816 499.16 46.3465 500.766C50.5429 501.572 54.3637 502.005 57.1353 502.237C58.5215 502.353 59.6461 502.419 60.4252 502.455C60.8147 502.474 61.1179 502.485 61.3243 502.491C61.4274 502.495 61.5064 502.497 61.5599 502.498L61.6207 502.499L61.6365 502.5L61.6407 502.5L61.6418 502.5C61.6419 502.5 61.642 502.5 61.6421 502.494C61.6427 502.474 61.6444 502.386 61.6512 502L61.6424 502.5ZM153 166.5C186.413 166.5 213.5 139.413 213.5 106C213.5 72.5868 186.413 45.5 153 45.5C119.587 45.5 92.5 72.5868 92.5 106C92.5 139.413 119.587 166.5 153 166.5ZM153 277.5C178.129 277.5 198.5 257.129 198.5 232C198.5 206.871 178.129 186.5 153 186.5C127.871 186.5 107.5 206.871 107.5 232C107.5 257.129 127.871 277.5 153 277.5ZM153 390.5C169.845 390.5 183.5 376.845 183.5 360C183.5 343.155 169.845 329.5 153 329.5C136.155 329.5 122.5 343.155 122.5 360C122.5 376.845 136.155 390.5 153 390.5ZM279 151.5C304.129 151.5 324.5 131.129 324.5 106C324.5 80.871 304.129 60.5 279 60.5C253.871 60.5 233.5 80.871 233.5 106C233.5 131.129 253.871 151.5 279 151.5ZM279 292.5C312.413 292.5 339.5 265.413 339.5 232C339.5 198.587 312.413 171.5 279 171.5C245.587 171.5 218.5 198.587 218.5 232C218.5 265.413 245.587 292.5 279 292.5ZM279 405.5C304.129 405.5 324.5 385.129 324.5 360C324.5 334.871 304.129 314.5 279 314.5C253.871 314.5 233.5 334.871 233.5 360C233.5 385.129 253.871 405.5 279 405.5ZM407 136.5C423.845 136.5 437.5 122.845 437.5 106C437.5 89.1553 423.845 75.5 407 75.5C390.155 75.5 376.5 89.1553 376.5 106C376.5 122.845 390.155 136.5 407 136.5ZM407 277.5C432.129 277.5 452.5 257.129 452.5 232C452.5 206.871 432.129 186.5 407 186.5C381.871 186.5 361.5 206.871 361.5 232C361.5 257.129 381.871 277.5 407 277.5ZM407 420.5C440.413 420.5 467.5 393.413 467.5 360C467.5 326.587 440.413 299.5 407 299.5C373.587 299.5 346.5 326.587 346.5 360C346.5 393.413 373.587 420.5 407 420.5Z',
+  ],
+};
+export const IconCorrelationPlot = createFontAwesomeIcon(dvCorrelationplot);
