+++ src/icons/font-awesome/icon-disease.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvDisease: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'disease' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M107.055 28.9758C112.485 45.2667 103.661 62.8183 87.4667 68.2486L62.0606 76.6849V193.921C62.0606 245.315 103.758 287.012 155.152 287.012C206.545 287.012 248.242 245.315 248.242 193.921V76.6849L222.933 68.2486C206.642 62.8183 197.915 45.2667 203.345 28.9758C208.776 12.6849 226.327 3.95766 242.618 9.38796L267.927 17.8243C293.236 26.2607 310.303 49.9213 310.303 76.6849V193.921C310.303 268.782 257.358 331.23 186.861 345.873C192.97 400.079 238.933 442.164 294.788 442.164C354.812 442.164 403.394 393.582 403.394 333.558V265C375.952 253.073 356.848 225.727 356.848 193.921C356.848 151.061 391.564 116.346 434.424 116.346C477.285 116.346 512 151.061 512 193.921C512 225.727 492.897 253.073 465.455 265V333.558C465.455 427.812 389.042 504.224 294.788 504.224C204.703 504.224 131.006 434.503 124.606 346.067C53.5273 331.812 0 269.17 0 193.921V76.6849C0 50.0183 17.0667 26.2607 42.4727 17.8243L67.7818 9.38796C84.0727 3.95766 101.624 12.7819 107.055 28.9758ZM434.424 224.952C442.654 224.952 450.547 221.682 456.366 215.863C462.185 210.044 465.455 202.151 465.455 193.921C465.455 185.692 462.185 177.799 456.366 171.98C450.547 166.16 442.654 162.891 434.424 162.891C426.195 162.891 418.302 166.16 412.483 171.98C406.663 177.799 403.394 185.692 403.394 193.921C403.394 202.151 406.663 210.044 412.483 215.863C418.302 221.682 426.195 224.952 434.424 224.952Z
+    `,
+  ],
+};
+
+export const IconDisease = createFontAwesomeIcon(dvDisease);
