+++ src/icons/font-awesome/icon-download-visualization.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvDownloadVisualization: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'downloadVisualization' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M9.49999 449.5L9.50001 449.503L10 449.5C9.50001 449.503 9.50001 449.503 9.50001 449.503L9.50002 449.504L9.50005 449.509L9.50018 449.525L9.50084 449.587C9.50152 449.642 9.5027 449.723 9.50476 449.829C9.50888 450.041 9.5165 450.352 9.53049 450.751C9.55848 451.55 9.612 452.704 9.71416 454.125C9.91841 456.966 10.3173 460.882 11.0965 465.177C12.6495 473.739 15.7324 483.926 21.8972 490.063C27.9395 496.079 37.8675 499.173 46.1984 500.773C50.3778 501.576 54.1832 502.007 56.9436 502.238C58.3242 502.354 59.4442 502.419 60.2201 502.456C60.6081 502.474 60.9101 502.485 61.1156 502.491C61.2184 502.495 61.297 502.497 61.3503 502.498L61.4109 502.499L61.4266 502.5L61.4308 502.5L61.4319 502.5C61.4322 502.5 61.4324 502.5 61.4413 502L61.4324 502.5H61.4413H472.971C488.161 502.5 500.5 490.537 500.5 475.75C500.5 460.963 488.161 449 472.971 449H75.3333H75.3176L75.3023 449.001L75.2999 449.001L75.289 449.002L75.2415 449.004C75.1986 449.006 75.1339 449.008 75.0494 449.01C74.8804 449.013 74.6325 449.014 74.323 449.005C73.7032 448.986 72.8397 448.923 71.868 448.752C69.913 448.406 67.5803 447.628 65.8876 445.928C64.196 444.229 63.4174 441.888 63.0691 439.925C62.896 438.95 62.8319 438.083 62.8111 437.461C62.8007 437.15 62.8012 436.901 62.8042 436.732C62.8057 436.647 62.8078 436.582 62.8095 436.539L62.8116 436.491L62.8122 436.48L62.8123 436.478L62.8123 436.478L62.8132 436.463V436.448V39.125C62.8132 23.8854 50.8956 11.5 36.1566 11.5C21.4176 11.5 9.5 23.8854 9.5 39.125L9.49999 449.5ZM218.44 75.2262L218.808 74.5H217.994L130.072 74.5C114.861 74.5 102.5 86.3946 102.5 101.105C102.5 115.816 114.861 127.71 130.072 127.71L201.611 127.71H202.046L202.106 127.279C204.686 108.787 210.304 91.2636 218.44 75.2262ZM201.735 170.884L201.682 170.445H201.239H130.072C114.861 170.445 102.5 182.339 102.5 197.05C102.5 211.76 114.861 223.655 130.072 223.655L216.584 223.655H217.382L217.035 222.937C209.256 206.851 203.987 189.332 201.735 170.884ZM247.543 266.542L247.396 266.39H247.184H130.072C114.861 266.39 102.5 278.284 102.5 292.995C102.5 307.705 114.861 319.6 130.072 319.6H275.717C286.839 319.6 296.433 313.243 300.79 304.076L301.012 303.61L300.54 303.403C280.514 294.608 262.543 282.016 247.543 266.542ZM413.119 415.545C428.329 415.545 440.691 403.65 440.691 388.94C440.691 374.229 428.329 362.335 413.119 362.335H130.072C114.861 362.335 102.5 374.229 102.5 388.94C102.5 403.65 114.861 415.545 130.072 415.545H413.119ZM427.181 138.396L407.5 158.077V53C407.5 33.3939 391.606 17.5 372 17.5C352.394 17.5 336.5 33.3939 336.5 53V161.454L313.319 138.273C299.507 124.461 277.113 124.461 263.302 138.273C249.49 152.085 249.49 174.479 263.302 188.291L340.56 265.549C341.82 267.42 343.277 269.197 344.931 270.851C352.391 278.311 362.344 281.756 372.108 281.188C380.742 280.888 389.288 277.442 395.879 270.851C397.353 269.378 398.669 267.807 399.828 266.159L477.386 188.601C491.249 174.737 491.249 152.26 477.386 138.396C463.522 124.533 441.045 124.533 427.181 138.396Z',
+  ],
+};
+export const IconDownloadVisualization = createFontAwesomeIcon(dvDownloadVisualization);
