+++ src/icons/font-awesome/icon-drug.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvDrug: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'drug' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M303.388 440.368C369.935 478.789 454.954 456.008 493.375 389.461C505.935 367.706 511.917 343.958 512 320.519C512.044 308.098 499.777 300.313 487.903 303.495L265.143 363.183C253.226 366.44 246.613 379.284 252.787 389.975C264.578 410.233 281.632 427.807 303.388 440.368Z
+    M252.481 250.381C239.92 272.137 233.938 295.884 233.856 319.324C233.811 331.744 246.078 339.529 257.953 336.348L480.712 276.659C492.63 273.402 499.242 260.558 493.069 249.868C481.278 229.61 464.223 212.035 442.468 199.474C375.921 161.054 290.902 183.834 252.481 250.381Z
+    M0 356.716C0 347.781 1.15243 338.697 3.5652 329.652L57.2404 128.43C72.0729 72.8249 128.881 39.8786 184.237 54.7781C239.593 69.6775 272.391 126.742 257.559 182.348C242.726 237.953 203.884 383.569 203.884 383.569C189.051 439.175 132.243 472.121 76.8868 457.222C30.5356 444.746 0 402.707 0 356.716ZM168.901 112.27C145.203 105.891 120.824 120.03 114.474 143.835L87.6366 244.446L173.487 267.554L200.325 166.943C206.675 143.137 192.6 118.649 168.901 112.27Z
+    `,
+  ],
+};
+export const IconDrug = createFontAwesomeIcon(dvDrug);
