+++ src/icons/font-awesome/icon-filter-clear.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvFilterClear: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'filterClear' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M182.47 320.171C182.964 320.782 183.234 321.545 183.234 322.332V403.913C184.447 415.915 187.528 421.126 196.799 428.028L279.143 476.062C279.32 476.193 279.515 476.31 279.715 476.403C298.697 485.278 313.738 480.748 321.536 469.188C272.975 449.051 238.819 401.186 238.819 345.342C238.819 271.329 298.819 211.329 372.832 211.329C385.375 211.329 397.514 213.052 409.028 216.274L504.611 93.6265C522.166 68.7071 507.005 32.5339 474.29 30.9262L45.0065 31.73C5.1103 30.1223 -9.25235 71.1187 5.90822 91.2149L182.47 320.171Z
+    M374.55 237.101C434.345 237.101 482.792 285.547 482.792 345.342C482.792 405.137 434.345 453.584 374.55 453.584C314.756 453.584 266.309 405.137 266.309 345.342C266.309 285.547 314.756 237.101 374.55 237.101ZM427.362 374.148L398.992 345.342L427.362 316.972C429.544 315.227 429.544 311.735 427.362 309.553L410.34 292.531C408.158 290.349 404.666 290.349 402.92 292.531L374.55 320.901L346.181 292.531C343.998 290.349 340.507 290.349 338.761 292.531L321.302 309.553C319.12 311.735 319.12 315.227 321.302 316.972L350.109 345.342L321.302 373.712C319.12 375.894 319.12 379.386 321.302 381.132L338.324 398.59C340.507 400.772 343.998 400.772 345.744 398.59L374.55 369.784L402.92 398.59C404.666 400.772 408.158 400.772 410.34 398.59L427.362 381.568C429.544 379.386 429.544 375.894 427.362 374.148Z
+    `,
+  ],
+};
+
+export const IconFilterClear = createFontAwesomeIcon(dvFilterClear);
