+++ src/icons/font-awesome/icon-filter-outline.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvFilterOutline: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'filterOutline' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M182.47 320.171C182.964 320.782 183.234 321.545 183.234 322.332V403.913C184.447 415.915 187.528 421.126 196.799 428.028L279.143 476.062C279.32 476.193 279.515 476.31 279.715 476.403C306.683 489.012 325.695 474.564 327.285 451.34V322.342C327.285 321.549 327.559 320.781 328.061 320.167L504.611 93.6265C522.166 68.7071 507.005 32.5339 474.29 30.9262L45.0065 31.73C5.1103 30.1223 -9.25235 71.1187 5.90822 91.2149L182.47 320.171ZM215.869 293.162L47.7754 74.6779L465.263 73.8962L294.802 292.986C288.031 301.271 284.332 311.641 284.332 322.342V426.526L226.187 396.33V322.332C226.187 311.714 222.545 301.417 215.869 293.162ZM41.2861 74.7644C41.2282 74.7875 41.1954 74.7938 41.1954 74.7938C41.1954 74.7938 41.2231 74.7805 41.2861 74.7644Z
+    `,
+  ],
+};
+
+export const IconFilterOutline = createFontAwesomeIcon(dvFilterOutline);
