+++ src/icons/font-awesome/icon-gene.ts
@@ -1,0 +1,17 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvGene: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'gene' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M228.204 80.9721C228.213 80.9585 228.222 80.9445 228.229 80.9301C246.828 45.4255 282.306 34.3055 301.536 31.4207C312.186 29.8229 322.116 37.1615 323.713 47.8119C325.311 58.4622 317.972 68.3913 307.322 69.9891C293.296 72.0932 272.569 79.6979 262.45 99.6603C262.086 100.379 261.678 101.073 261.228 101.741C252.441 114.779 249.762 130.761 256.009 163.692C256.01 163.699 256.011 163.705 256.013 163.712L275.316 247.158L275.316 247.159L275.338 247.255L275.338 247.256L295.521 336.542C295.569 336.755 295.614 336.969 295.655 337.183C302.871 374.964 300.048 405.037 282.969 430.38C258.09 467.3 221.197 476.095 198.581 480.444C188.006 482.477 177.783 475.553 175.75 464.977C173.716 454.401 180.641 444.179 191.216 442.145C212.085 438.132 235.032 431.728 250.628 408.586C259.744 395.057 263.488 376.906 257.409 344.823C257.408 344.818 257.407 344.812 257.405 344.806L237.32 255.951L237.32 255.949L217.928 172.119L217.441 172.232L217.928 172.119C217.871 171.874 217.819 171.629 217.772 171.382C210.821 134.988 211.461 106.537 228.204 80.9721ZM432.267 232.301C432.275 232.293 432.282 232.285 432.289 232.277C441.328 221.871 446.911 208.226 450.041 194.802C453.171 181.376 453.858 168.131 453.075 158.487C452.205 147.752 460.201 138.345 470.935 137.474C481.669 136.603 491.077 144.6 491.948 155.334C493.916 179.607 489.579 226.317 461.206 258.453C460.998 258.688 460.785 258.919 460.566 259.144C428.043 292.577 381.291 297.775 318.702 298.334C307.933 298.43 299.125 289.778 299.029 279.008C298.933 268.239 307.585 259.431 318.354 259.335C381.941 258.768 412.369 252.471 432.267 232.301ZM84.3441 275.776C84.3255 275.795 84.3083 275.815 84.2928 275.836C67.0992 299.383 59.4295 320.525 58.6115 334.669C57.9897 345.421 48.7697 353.633 38.0181 353.011C27.2666 352.389 19.0547 343.169 19.6766 332.418C21.0938 307.912 33.3035 279.145 53.7779 251.502C54.2969 250.801 54.8623 250.136 55.4703 249.511C87.8997 216.173 131.814 210.793 194.725 210.232C205.494 210.136 214.302 218.788 214.398 229.557C214.494 240.326 205.842 249.134 195.073 249.23C132.411 249.789 104.273 255.967 84.3441 275.776ZM286.443 100.704C293.386 92.4707 305.688 91.4244 313.921 98.3668L422.956 190.308C431.189 197.25 432.235 209.552 425.293 217.786C418.35 226.019 406.048 227.065 397.815 220.123L288.78 128.182C280.547 121.239 279.501 108.937 286.443 100.704ZM280.095 173.91C287.037 165.677 299.34 164.631 307.573 171.573L354.309 210.983C362.543 217.925 363.589 230.227 356.647 238.461C349.704 246.694 337.402 247.74 329.169 240.798L282.432 201.388C274.199 194.446 273.152 182.143 280.095 173.91ZM156.781 270.105C163.723 261.872 176.025 260.825 184.259 267.768L230.995 307.177C239.228 314.12 240.275 326.422 233.332 334.655C226.39 342.889 214.088 343.935 205.854 336.993L159.118 297.583C150.885 290.641 149.838 278.338 156.781 270.105ZM88.1342 290.78C95.0767 282.547 107.379 281.5 115.612 288.443L224.647 380.384C232.88 387.326 233.926 399.628 226.984 407.861C220.041 416.095 207.739 417.141 199.506 410.199L90.4714 318.258C82.2381 311.315 81.1918 299.013 88.1342 290.78Z',
+  ],
+};
+
+export const IconGene = createFontAwesomeIcon(dvGene);
