+++ src/icons/font-awesome/icon-heatmap.ts
@@ -1,0 +1,17 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvHeatmap: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'heatmap' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M61.6424 502.5H61.6512H474.861C490.112 502.5 502.5 490.489 502.5 475.643C502.5 460.796 490.112 448.786 474.861 448.786H75.6H75.5843L75.5689 448.787L75.5665 448.787L75.5555 448.787L75.5078 448.79C75.4648 448.792 75.3998 448.794 75.3149 448.796C75.1452 448.799 74.8963 448.8 74.5854 448.791C73.963 448.771 73.0959 448.709 72.1202 448.536C70.157 448.189 67.8143 447.408 66.1143 445.701C64.4153 443.994 63.6334 441.643 63.2837 439.672C63.1099 438.693 63.0455 437.822 63.0246 437.197C63.0142 436.885 63.0146 436.636 63.0176 436.465C63.0191 436.38 63.0213 436.315 63.023 436.271L63.0251 436.223L63.0257 436.212L63.0258 436.21L63.0258 436.21L63.0267 436.195V436.181V37.2357C63.0267 21.935 51.0613 9.5 36.2633 9.5C21.4654 9.5 9.49997 21.935 9.49997 37.2357L9.49996 449.286L9.49998 449.288L9.99997 449.286C9.49998 449.288 9.49998 449.289 9.49998 449.289L9.49999 449.29L9.50002 449.294L9.50015 449.311L9.50082 449.373C9.50149 449.428 9.50268 449.51 9.50475 449.616C9.50889 449.829 9.51653 450.141 9.53059 450.542C9.55869 451.344 9.61243 452.502 9.71499 453.929C9.92008 456.782 10.3206 460.714 11.103 465.027C12.6623 473.623 15.7577 483.851 21.9472 490.013C28.0137 496.053 37.9816 499.16 46.3465 500.766C50.5429 501.572 54.3637 502.005 57.1353 502.237C58.5215 502.353 59.6461 502.419 60.4252 502.455C60.8147 502.474 61.1179 502.485 61.3243 502.491C61.4274 502.495 61.5064 502.497 61.5599 502.498L61.6207 502.499L61.6365 502.5L61.6407 502.5L61.6418 502.5C61.6419 502.5 61.642 502.5 61.6421 502.494C61.6427 502.474 61.6444 502.386 61.6512 502L61.6424 502.5ZM189 29.5H99C95.9624 29.5 93.5 31.9624 93.5 35V125C93.5 128.038 95.9624 130.5 99 130.5H189C192.038 130.5 194.5 128.038 194.5 125V35C194.5 31.9624 192.038 29.5 189 29.5ZM449 29.5H359C355.962 29.5 353.5 31.9624 353.5 35V125C353.5 128.038 355.962 130.5 359 130.5H449C452.038 130.5 454.5 128.038 454.5 125V35C454.5 31.9624 452.038 29.5 449 29.5ZM189 174.5H99C95.9624 174.5 93.5 176.962 93.5 180V270C93.5 273.038 95.9624 275.5 99 275.5H189C192.038 275.5 194.5 273.038 194.5 270V180C194.5 176.962 192.038 174.5 189 174.5ZM319 174.5H229C225.962 174.5 223.5 176.962 223.5 180V270C223.5 273.038 225.962 275.5 229 275.5H319C322.038 275.5 324.5 273.038 324.5 270V180C324.5 176.962 322.038 174.5 319 174.5ZM319 319.5H229C225.962 319.5 223.5 321.962 223.5 325V415C223.5 418.038 225.962 420.5 229 420.5H319C322.038 420.5 324.5 418.038 324.5 415V325C324.5 321.962 322.038 319.5 319 319.5ZM449 319.5H359C355.962 319.5 353.5 321.962 353.5 325V415C353.5 418.038 355.962 420.5 359 420.5H449C452.038 420.5 454.5 418.038 454.5 415V325C454.5 321.962 452.038 319.5 449 319.5Z',
+  ],
+};
+
+export const IconHeatmap = createFontAwesomeIcon(dvHeatmap);
