+++ src/icons/font-awesome/icon-hexbin-plot.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvHexbinplot: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'hexbinplot' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M61.6512 502V501H474.861C489.325 501 501 489.62 501 475.643C501 461.666 489.325 450.286 474.861 450.286H75.6274C75.613 450.286 75.5944 450.287 75.5717 450.288C75.5182 450.291 75.4421 450.293 75.3456 450.295C75.1528 450.299 74.878 450.3 74.5389 450.29C73.8619 450.269 72.921 450.201 71.8591 450.013C69.7586 449.642 67.0641 448.78 65.0514 446.759C63.0416 444.741 62.1804 442.04 61.8068 439.934C61.6179 438.87 61.5481 437.926 61.5255 437.247C61.5141 436.907 61.5145 436.632 61.5179 436.439C61.5196 436.342 61.522 436.266 61.5242 436.212C61.5251 436.189 61.526 436.169 61.5267 436.155V37.2357C61.5267 22.7116 50.1819 11 36.2633 11C22.3447 11 11 22.7116 11 37.2357V449.283L11 449.284L11.0001 449.297L11.0007 449.355C11.0014 449.406 11.0025 449.484 11.0045 449.587C11.0085 449.792 11.0159 450.096 11.0297 450.49C11.0572 451.276 11.1101 452.415 11.2112 453.822C11.4135 456.636 11.8086 460.512 12.5789 464.759C14.1302 473.311 17.1644 483.135 23.0055 488.95C28.7333 494.653 38.3109 497.695 46.6294 499.293C50.7608 500.086 54.5269 500.514 57.2604 500.742C58.6265 500.857 59.7329 500.921 60.4958 500.957C60.8773 500.975 61.1727 500.986 61.3718 500.992C61.4713 500.995 61.5467 500.997 61.5967 500.999L61.6524 501L61.6656 501L61.6685 501H61.6691C61.6691 501 61.669 501 61.6512 502ZM61.6512 502H474.861C489.85 502 502 490.199 502 475.643C502 461.086 489.85 449.286 474.861 449.286H75.6L61.6512 502Z
+    M153.169 243.901H244.832L290.664 323.336L244.832 402.771H153.169L107.337 323.336L153.169 243.901ZM319.648 148.697H411.311L457.143 228.132L411.311 307.567H319.648L273.816 228.132L319.648 148.697ZM153.169 56.5736L244.832 56.5736L290.664 136.009L244.832 215.444H153.169L107.337 136.009L153.169 56.5736Z
+    `,
+  ],
+};
+export const IconHexbinPlot = createFontAwesomeIcon(dvHexbinplot);
