+++ src/icons/font-awesome/icon-junction.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvJunction: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'junction' as IconName,
+  icon: [
+    24,
+    24,
+    [],
+    '',
+    `M9.90909 19.7647H0V24H9.90909V19.7647Z
+    M24 19.7647H14.0909V24H24V19.7647Z
+    M16.8868 4.85812C16.8868 6.69237 15.8643 8.2892 14.3549 9.11596C15.2396 10.7537 16.5039 12.1457 17.1972 12.7664C17.9371 13.3696 18.364 14.3022 18.6149 15.0594C18.8801 15.86 19.0126 16.6656 19.0552 17.1858C19.1038 17.78 18.6587 18.301 18.0609 18.3493C17.4631 18.3977 16.9391 17.9551 16.8904 17.3609C16.8599 16.9876 16.7565 16.3524 16.5519 15.7348C16.3358 15.0824 16.066 14.632 15.8128 14.4302L15.79 14.4121L15.7682 14.3928C14.926 13.6444 13.2884 11.871 12.2215 9.71133C12.1481 9.71459 12.0742 9.71623 12 9.71623C11.9258 9.71623 11.8519 9.71459 11.7785 9.71133C10.7116 11.871 9.074 13.6444 8.23175 14.3928L8.21 14.4121L8.18723 14.4302C7.93401 14.632 7.66419 15.0824 7.44808 15.7348C7.24346 16.3524 7.14014 16.9876 7.10959 17.3609C7.06094 17.9551 6.53691 18.3977 5.93913 18.3493C5.34134 18.301 4.89617 17.78 4.94481 17.1858C4.98739 16.6656 5.11991 15.86 5.38514 15.0594C5.63596 14.3022 6.06293 13.3696 6.80283 12.7664C7.49615 12.1457 8.76044 10.7537 9.64505 9.11596C8.13572 8.2892 7.11316 6.69237 7.11316 4.85812C7.11316 2.17505 9.30107 0 12 0C14.6989 0 16.8868 2.17505 16.8868 4.85812Z
+    `,
+  ],
+};
+
+export const IconJunction = createFontAwesomeIcon(dvJunction);
