+++ src/icons/font-awesome/icon-mouse.ts
@@ -1,0 +1,17 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvMouse: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'mouse' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `M433.2 93.5c-19.7 13.3-21.6 41.5-4.9 74.9 4.3 8.7 11.7 20.3 22.8 35.8 17 23.6 23.1 35.3 27.3 52.3 5.1 20.5 1.6 41.3-10 60.6-4.6 7.6-19.1 24.3-18 20.7 3.1-10.2 3.9-32.8 1.7-45.8-6-34.1-27.2-63.3-58.3-80-22.5-12-47.1-15.4-84.5-11.5-23.3 2.5-53.6 9.1-69.8 15.4-4.5 1.8-5 2.2-4.7 4.8.2 1.5.4 9.5.5 17.8.1 14.4 0 15.4-2.9 23.2-7.4 19.9-22 35.8-40.8 44.3-13.5 6.2-14.5 6.3-59.7 6.9l-41.7.6-9.3 10c-10.7 11.6-24.8 28.4-33.8 40.3-8.1 10.7-11.7 13.1-21.8 14.3-9.2 1.2-15.8 4.1-17.7 7.8-2 3.8-6.5 18.9-7.2 24.3-.8 5.7.9 10.6 4 11.3 1.1.3 91.9.4 201.7.3l199.6-.3 7.9-4c21-10.6 45.9-28.3 61.3-43.3 36.4-35.8 45.7-75.1 28.6-121-7.8-20.7-17.9-37.7-43.5-73.2-18-25.1-22.8-36.7-23.7-58-.5-11.7.8-24.2 3.2-30.3 1-2.5-.4-2.1-6.3 1.8zM109.4 343.9c9.9 6.1 10.2 22.4.5 28.8-10.8 7.1-25.3 1.1-26.6-11.1-.3-2.5-.3-6 .1-7.9.8-4.2 6.5-10.3 10.9-11.6 4.4-1.4 11.3-.5 15.1 1.8M154 189.1c-17.6 2.5-30.7 13.3-40.6 33.4-6.5 13.2-8.5 22.1-9.1 40.5-.4 10-.1 18.1.7 22.8l1.2 7.2h29.1c33.2 0 38.7-.8 50.6-6.6 8.8-4.3 19.1-14.1 23.4-22.2 9.1-16.9 8.9-37.4-.6-51.1-11.6-16.9-34.6-27-54.7-24z`,
+  ],
+};
+
+export const IconMouse = createFontAwesomeIcon(dvMouse);
