+++ src/icons/font-awesome/icon-number-array-column.ts
@@ -1,0 +1,21 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvNumberArrayCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'numberArrayCol' as IconName,
+  icon: [
+    512,
+    312,
+    [],
+    '',
+    `M0 312V0H83.4854V33.28H41.0103V278.72H83.4854V312H0Z
+    M80.0617 250.187V216.907H123.123V102.923L90.6072 115.403L76.5465 82.9547L138.648 58.8267H169.992V216.907H210.709V250.187H80.0617Z
+    M232.857 301.493L246.332 252.128C241.45 250.649 237.251 247.876 233.736 243.808C230.221 239.556 228.463 234.286 228.463 228C228.463 220.604 231.002 214.503 236.08 209.696C241.352 204.889 247.992 202.485 255.999 202.485C264.006 202.485 270.548 204.981 275.625 209.973C280.898 214.78 283.534 221.067 283.534 228.832C283.534 234.564 282.753 239.925 281.191 244.917C279.629 249.909 277.285 256.196 274.161 263.776L258.635 301.493H232.857Z
+    M293.002 85.1733C299.642 79.4418 306.867 74.3573 314.679 69.92C322.49 65.4827 330.888 61.9698 339.871 59.3813C348.854 56.7929 358.326 55.4987 368.285 55.4987C381.565 55.4987 393.477 58.0871 404.023 63.264C414.568 68.256 422.868 75.0969 428.922 83.7867C434.976 92.4764 438.003 102.368 438.003 113.461C438.003 122.151 436.343 130.471 433.023 138.421C429.898 146.187 424.821 154.137 417.791 162.272C410.76 170.222 401.582 179.004 390.255 188.619L363.013 211.36H440.932V250.187H296.224V221.067L361.841 160.331C368.481 154.229 373.851 148.775 377.952 143.968C382.053 139.161 385.08 134.724 387.033 130.656C388.986 126.588 389.962 122.706 389.962 119.008C389.962 114.756 388.888 110.965 386.74 107.637C384.787 104.124 381.858 101.351 377.952 99.3173C374.046 97.2836 369.066 96.2667 363.013 96.2667C357.74 96.2667 352.565 97.1911 347.487 99.04C342.41 100.704 337.43 103.108 332.548 106.251C327.861 109.394 323.174 113.092 318.487 117.344L293.002 85.1733Z
+    M428.515 312V278.72H470.99V33.28H428.515V0H512V312H428.515Z`,
+  ],
+};
+
+export const IconNumberArrayColumn = createFontAwesomeIcon(dvNumberArrayCol);
