+++ src/icons/font-awesome/icon-numeric-column.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvNumericCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'numericalCol' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M3.91283 329.338V368.452H149.34V329.338H104.016V143.548H69.1267L0 171.905L15.6513 210.041L51.845 195.24V329.338H3.91283Z
+    M190.623 156.586C181.928 161.801 173.885 167.777 166.494 174.513L194.862 212.323C200.079 207.325 205.296 202.979 210.513 199.285C215.948 195.591 221.491 192.766 227.143 190.81C232.795 188.637 238.555 187.551 244.425 187.551C251.163 187.551 256.706 188.746 261.054 191.136C265.402 193.527 268.662 196.786 270.836 200.915C273.227 204.826 274.423 209.281 274.423 214.279C274.423 218.625 273.336 223.188 271.162 227.968C268.988 232.749 265.619 237.964 261.054 243.614C256.489 249.264 250.511 255.674 243.12 262.845L170.081 334.228V368.452H331.159V322.819H244.398L274.749 296.092C287.357 284.792 297.574 274.47 305.399 265.127C313.225 255.565 318.877 246.222 322.355 237.095C326.051 227.751 327.898 217.973 327.898 207.76C327.898 194.722 324.529 183.096 317.79 172.883C311.051 162.67 301.813 154.63 290.074 148.763C278.336 142.679 265.076 139.636 250.294 139.636C239.207 139.636 228.665 141.157 218.665 144.2C208.666 147.242 199.318 151.37 190.623 156.586Z
+    M369.834 303.914L342.77 341.073C353.639 351.068 365.921 358.782 379.616 364.215C393.311 369.647 408.745 372.364 425.918 372.364C441.352 372.364 455.59 369.647 468.633 364.215C481.676 358.565 492.11 350.199 499.935 339.117C507.978 328.035 512 314.453 512 298.373C512 283.597 508.413 271.32 501.24 261.541C494.066 251.545 484.175 244.049 471.567 239.051C464.495 236.074 456.856 233.947 448.65 232.669L503.848 177.772V143.548H351.574V189.181H440.178L389.398 241.332V269.364H418.418C428.2 269.364 436.135 270.668 442.221 273.275C448.308 275.883 452.764 279.36 455.59 283.706C458.633 288.052 460.155 292.941 460.155 298.373C460.155 307.5 456.46 314.562 449.069 319.56C441.678 324.341 431.896 326.731 419.722 326.731C409.723 326.731 400.593 324.775 392.333 320.864C384.072 316.735 376.573 311.085 369.834 303.914Z    `,
+  ],
+};
+export const IconNumericColumn = createFontAwesomeIcon(dvNumericCol);
