+++ src/icons/font-awesome/icon-protein-ball-and-stick.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvProteinBallAndStick: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'proteinBallAndStick' as IconName,
+  icon: [
+    23,
+    24,
+    [],
+    '',
+    `M6.85714 2.28571C6.85714 2.64748 6.94119 2.98961 7.09081 3.29365L4.40099 4.84662C4.05931 4.01453 3.24094 3.42857 2.28571 3.42857C1.02335 3.42857 0 4.45192 0 5.71429C0 6.90101 0.904379 7.8765 2.06152 7.98914V12.5823C0.904379 12.6949 0 13.6704 0 14.8571C0 16.1195 1.02335 17.1429 2.28571 17.1429C3.24094 17.1429 4.05932 16.5569 4.40099 15.7248L8 17.8027V19.7344C7.3168 20.1296 6.85714 20.8683 6.85714 21.7143C6.85714 22.9767 7.88049 24 9.14286 24C10.4052 24 11.4286 22.9767 11.4286 21.7143C11.4286 20.8683 10.9689 20.1296 10.2857 19.7344V17.8027L13.8847 15.7248C14.2264 16.5569 15.0448 17.1429 16 17.1429C17.2624 17.1429 18.2857 16.1195 18.2857 14.8571C18.2857 13.6704 17.3813 12.6949 16.2242 12.5823V7.38756L19.5439 5.47088C19.8528 5.62659 20.2019 5.71429 20.5714 5.71429C21.8338 5.71429 22.8571 4.69094 22.8571 3.42857C22.8571 2.16621 21.8338 1.14286 20.5714 1.14286C19.3091 1.14286 18.2857 2.16621 18.2857 3.42857C18.2857 3.47133 18.2869 3.51381 18.2892 3.55598L14.9692 5.47278L11.1949 3.29367C11.3445 2.98962 11.4286 2.64748 11.4286 2.28571C11.4286 1.02335 10.4052 0 9.14286 0C7.88049 0 6.85714 1.02335 6.85714 2.28571ZM13.7143 13.1839L9.14284 15.8232L4.34724 13.0545V7.51697L9.14284 4.74823L13.7143 7.38756V13.1839Z`,
+  ],
+};
+export const IconProteinBallAndStick = createFontAwesomeIcon(dvProteinBallAndStick);
