+++ src/icons/font-awesome/icon-protein-cartoon.ts
@@ -1,0 +1,17 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvProteinCartoon: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'proteinCartoon' as IconName,
+  icon: [
+    20,
+    18,
+    [],
+    '',
+    `M18.1132 9.15125C18.3735 9.42009 18.5179 9.62296 18.5878 9.76977C18.6066 9.80933 18.6164 9.83736 18.6215 9.85566C18.6079 9.87081 18.5836 9.89448 18.5424 9.92614C18.4351 10.0086 18.2779 10.0958 18.0565 10.2019C17.9892 10.2341 17.9071 10.2721 17.819 10.3127C17.6639 10.3843 17.49 10.4646 17.3467 10.5372C17.1061 10.659 16.8254 10.8182 16.591 11.0342C16.3476 11.2586 16.1128 11.5835 16.0601 12.0276C15.9926 12.5962 16.1226 13.1341 16.2587 13.5726C16.3145 13.7524 16.3779 13.9348 16.4362 14.1024L16.4753 14.215C16.546 14.4197 16.6074 14.6042 16.6549 14.782C16.7508 15.1406 16.7712 15.406 16.7184 15.6304C16.6701 15.8361 16.5387 16.0978 16.1531 16.4064C15.8189 16.6739 15.76 17.1677 16.0216 17.5095C16.2831 17.8513 16.7661 17.9115 17.1003 17.644C17.7039 17.161 18.0682 16.6122 18.2127 15.998C18.3527 15.4025 18.2633 14.8385 18.1374 14.3676C18.074 14.1307 17.9965 13.9003 17.9243 13.6913L17.8806 13.5653C17.823 13.3996 17.7706 13.2488 17.7235 13.0972C17.6089 12.7278 17.5623 12.4598 17.5834 12.2396C17.5897 12.2316 17.6011 12.219 17.62 12.2015C17.6946 12.1327 17.8226 12.0498 18.0285 11.9456C18.157 11.8806 18.2732 11.8272 18.4016 11.7683C18.4945 11.7256 18.5937 11.68 18.7086 11.625C18.9466 11.511 19.2244 11.3678 19.4652 11.1829C19.7063 10.9978 19.975 10.7221 20.0981 10.3196C20.2272 9.89762 20.1556 9.47255 19.9696 9.08183C19.7261 8.57046 19.2381 8.01647 18.4895 7.38197L17.9164 3.55115L17.8614 3.40956C17.825 3.31594 17.7914 3.22927 17.7572 3.14282C17.7128 3.03052 17.6705 2.91712 17.6281 2.80372C17.5317 2.54542 17.4353 2.2871 17.314 2.04172C17.0612 1.53088 16.6595 1.19629 16.0844 1.11798C15.62 1.05475 15.1557 0.990978 14.6914 0.927202L14.0882 0.844383C14.0175 0.85027 13.9469 0.856157 13.8387 0.853573C13.7789 0.841084 13.7566 0.837727 13.7342 0.835027C13.6634 0.841052 13.5927 0.847077 13.4814 0.843359C13.3878 0.826091 13.3348 0.818562 13.2818 0.811035C13.2108 0.817083 13.1397 0.823126 13.0476 0.824246C13.0265 0.819319 13.0265 0.819562 13.0265 0.819562L12.2869 1.00615C12.5836 0.931309 12.805 0.875354 13.0039 0.81197C12.8983 0.817061 12.7936 0.820225 12.6899 0.823358C12.3756 0.832857 12.0702 0.842086 11.7754 0.9039C11.6022 0.940207 11.449 1.07634 11.3041 1.20512C11.2679 1.23732 11.2322 1.26907 11.1968 1.29868C11.2325 1.35098 11.2681 1.4019 11.3032 1.452C11.4254 1.62645 11.5407 1.79113 11.6263 1.97063C11.6934 2.1112 11.7481 2.25854 11.8027 2.40566C11.8774 2.60678 11.9519 2.80754 12.0576 2.98996C12.1813 3.20353 12.3961 3.37901 12.6032 3.52122C12.9681 3.77182 13.3583 3.98397 13.7442 4.18947L13.7488 4.1736C13.824 3.91223 13.8931 3.67171 13.9565 3.40461C13.9099 3.27453 13.8602 3.14721 13.807 3.02267C13.8638 3.15473 13.917 3.29024 13.967 3.42926C13.9908 3.50422 14.0118 3.57366 14.0315 3.63918C14.0706 3.76865 14.1053 3.88366 14.1474 3.99475C14.4808 4.87365 14.8163 5.75173 15.1518 6.6298L15.1522 6.63106L15.1527 6.63232L15.4758 7.47811C15.5011 7.48345 15.5286 7.48616 15.5561 7.48888C15.6051 7.49371 15.6541 7.49855 15.6912 7.51825C15.6334 7.48783 15.5479 7.49616 15.4887 7.50572C15.8564 8.35808 16.3978 8.94294 17.331 9.03114L17.5312 9.06345C17.6354 9.08023 17.6914 9.08924 17.7473 9.09859C17.8121 9.10942 17.8769 9.1207 18.0165 9.14502C18.0489 9.15064 18.0816 9.15263 18.1132 9.15125ZM17.5761 12.2499C17.576 12.2502 17.5761 12.2499 17.5761 12.2499ZM13.7464 4.83233C13.7543 5.33418 13.7623 5.83603 13.7508 6.33742C13.7448 6.59547 13.6983 6.85255 13.6486 7.12676C13.6319 7.21891 13.6149 7.31299 13.599 7.40969C13.5655 7.48637 13.5478 7.5224 13.5302 7.55843C13.4604 7.63638 13.3907 7.71438 13.278 7.81121C13.2263 7.81217 13.1754 7.81505 13.126 7.81786C13.0322 7.82318 12.9435 7.8282 12.8643 7.81911C12.9297 7.82728 13.0007 7.82645 13.0741 7.8256C13.1348 7.82489 13.197 7.82417 13.2589 7.8285C13.1978 7.88074 13.1365 7.93499 13.0746 7.9898C12.8618 8.17828 12.6412 8.37356 12.3924 8.51743C12.0275 8.72838 11.6472 8.91156 11.267 9.09474C11.1057 9.17243 10.9445 9.25012 10.7844 9.32993C10.1533 9.64451 9.53403 9.71192 8.92139 9.2706C8.80896 9.1896 8.68345 9.12758 8.55154 9.06239C8.48282 9.02843 8.41237 8.99361 8.34112 8.9548C8.38045 8.92775 8.43115 8.893 8.48295 8.87672C8.45422 8.88494 8.42491 8.89798 8.3956 8.91103C8.37045 8.92222 8.3453 8.93341 8.32052 8.94155C8.26565 8.92021 8.211 8.89807 8.15636 8.87594C8.0151 8.81871 7.87372 8.76144 7.72851 8.71803C7.68494 8.70501 7.63178 8.72549 7.57864 8.74597C7.55431 8.75534 7.52999 8.76471 7.50658 8.77087C7.51349 8.79432 7.51823 8.81993 7.52296 8.84554C7.53315 8.90064 7.54334 8.95576 7.57521 8.98933C8.18832 9.6351 8.50498 10.4496 8.79717 11.2705C8.86189 11.4523 8.85422 11.661 8.84595 11.8863C8.84348 11.9537 8.84095 12.0225 8.84029 12.0926C8.45943 12.3915 8.16471 12.7213 7.96506 13.1178C8.16467 12.7222 8.45927 12.3966 8.83946 12.1163C8.8373 12.1512 8.83618 12.1879 8.83503 12.2256C8.83064 12.3693 8.8258 12.5276 8.76143 12.6542C8.26891 13.6228 7.75819 14.5819 7.24576 15.5398C7.08715 15.8364 6.82483 15.9446 6.47593 15.8917C7.10161 15.3386 7.43202 14.6029 7.69921 13.8293C7.43165 14.603 7.09859 15.3384 6.45196 15.8874C6.17569 15.8108 5.89919 15.7351 5.6227 15.6594L5.62165 15.6591C4.96991 15.4805 4.31818 15.302 3.66944 15.1128C3.06526 14.9365 2.46346 14.9743 1.80779 15.0722C1.71109 15.0958 1.6171 15.1175 1.52541 15.1386C1.25882 15.2002 1.0112 15.2574 0.773103 15.3415C0.178908 15.5512 0.103148 15.4956 0.197411 14.8509C0.199517 14.8365 0.20145 14.8221 0.203384 14.8076C0.209424 14.7624 0.215474 14.7171 0.226855 14.6733C0.39855 14.013 0.571134 13.3529 0.743718 12.6928L0.868742 12.2146C1.61314 11.4488 2.56716 11.3526 3.54337 11.3635C3.78834 11.3663 4.03283 11.4033 4.27746 11.4403C4.40554 11.4597 4.53368 11.4791 4.66192 11.4936C4.72815 11.5011 4.79536 11.4996 4.86763 11.4979C4.90224 11.4971 4.93865 11.4962 4.97604 11.4964C4.52884 10.8514 4.10642 10.2421 3.6782 9.58314C3.87524 9.30484 4.06433 9.06107 4.28527 8.85259C4.12095 9.00764 3.97427 9.18197 3.82094 9.36443C3.76814 9.42727 3.71446 9.49114 3.65909 9.55537C3.42706 9.29639 3.1938 9.02538 2.9937 8.73081C2.80533 8.45352 2.78233 8.14262 2.96719 7.84763C3.30568 7.3075 3.65152 6.77219 4.008 6.22041L4.11908 6.04846C4.06298 6.25478 3.99595 6.48374 3.9149 6.7074C3.97129 6.55178 4.02087 6.39384 4.07048 6.23565C4.09217 6.16649 4.11386 6.09733 4.13612 6.02839C4.22976 5.90468 4.32302 5.78066 4.41629 5.65665C4.63199 5.36984 4.84768 5.08302 5.06807 4.80003C5.32637 4.46834 5.69067 4.40032 6.06265 4.48279L6.06636 4.48362C6.98673 4.68766 7.90905 4.89213 8.81362 5.157C9.14834 5.25501 9.46681 5.41102 9.80005 5.57427C9.95483 5.65009 10.1128 5.72747 10.277 5.80133C10.2789 5.76349 10.2813 5.72549 10.2838 5.68737C10.2892 5.60418 10.2947 5.52001 10.2936 5.43639C10.2857 4.79755 10.2778 4.15867 10.2633 3.51996C10.2438 2.65627 10.6419 1.96944 11.1799 1.31973C11.3566 1.5335 11.5161 1.73957 11.6263 1.97063C11.6934 2.1112 11.7481 2.25854 11.8027 2.40566C11.8774 2.60678 11.9519 2.80754 12.0576 2.98996C12.1813 3.20353 12.3961 3.37901 12.6032 3.52122C12.8604 3.69787 13.1302 3.85584 13.4 4.01338C13.5129 4.07933 13.6259 4.14528 13.7379 4.21263C13.7398 4.41917 13.7431 4.62576 13.7464 4.83233Z`,
+  ],
+};
+
+export const IconProteinCartoon = createFontAwesomeIcon(dvProteinCartoon);
