+++ src/icons/font-awesome/icon-protein-stick.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvProteinStick: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'proteinStick' as IconName,
+  icon: [
+    22,
+    24,
+    [],
+    '',
+    `M8.64145 25C9.33999 25 9.90636 24.4344 9.90636 23.7368V18.1504L15.2141 15.0903L15.2198 15.0871L15.2373 15.0767L15.8466 14.7255C16.4791 14.1875 16.4791 13.8555 16.4791 13.3359V6.78208C16.4791 6.73489 16.4778 6.68771 16.4752 6.64114L22.1712 3.35714C22.7761 3.00835 22.9835 2.23584 22.6341 1.63171C22.2848 1.02759 21.5112 0.820658 20.9063 1.16944L15.0903 4.52253L9.90636 1.53396C9.12366 1.08279 8.15923 1.08279 7.37653 1.53396L2.06875 4.59407C1.28605 5.04555 0.803833 5.87943 0.803833 6.78208V12.9023C0.803833 13.805 1.28605 14.6388 2.06875 15.0903L7.37653 18.1504V23.7368C7.37653 24.4344 7.9429 25 8.64145 25ZM13.7011 13.0454L8.64145 15.9624L3.33366 12.9023V6.78208L8.64145 3.72196L13.7011 6.63898V13.0454Z`,
+  ],
+};
+export const IconProteinStick = createFontAwesomeIcon(dvProteinStick);
