+++ src/icons/font-awesome/icon-protein-surface.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvProteinSurface: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'proteinSurface' as IconName,
+  icon: [
+    24,
+    22,
+    [],
+    '',
+    `M16.8103 0.468166C14.917 -0.614518 12.4949 0.25773 11.7476 2.29133L11.6499 2.55724C11.1567 3.89944 9.91213 4.82923 8.47256 4.93099L6.04224 5.10277C4.12786 5.23809 2.64464 6.81442 2.64464 8.71365V9.4774C2.64464 10.5233 2.15666 11.5106 1.32237 12.1528C-0.577428 13.615 -0.400827 16.5091 1.66297 17.7345L2.19765 18.0519C2.86182 18.4463 3.35332 19.0718 3.57588 19.8059C4.23128 21.9678 6.93794 22.7185 8.63329 21.2086L8.87829 20.9904C10.0262 19.9681 11.7351 19.8647 13.0002 20.7411L13.2207 20.8938C15.0497 22.1607 17.589 21.5239 18.5874 19.5478L19.5841 17.5753C19.9823 16.7872 20.6706 16.1802 21.5074 15.879L21.8696 15.7487C24.0936 14.9484 24.7248 12.1332 23.0514 10.4773C22.0534 9.4897 21.8178 7.97576 22.4692 6.73673L22.8133 6.0821C24.17 3.50166 21.9811 0.487916 19.0712 0.899278C18.2916 1.0095 17.4922 0.858122 16.8103 0.468166Z`,
+  ],
+};
+export const IconProteinSurface = createFontAwesomeIcon(dvProteinSurface);
