+++ src/icons/font-awesome/icon-sankey.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvSankey: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'sankey' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M63.7633 473.613L63.7633 348.159C107.446 348.584 140.255 367.276 188.941 395.049C190.408 395.886 191.889 396.731 193.385 397.583C286.27 450.528 367.626 444.279 419.463 440.297L419.481 440.296C430.475 439.451 440.102 438.712 448.237 438.655V473.613C448.237 489.172 460.409 501.849 475.5 501.849C490.591 501.849 502.763 489.172 502.763 473.613V38.3871C502.763 22.8275 490.591 10.1514 475.5 10.1514C460.409 10.1514 448.237 22.8275 448.237 38.3871V63.3311C440.102 63.2738 430.475 62.5349 419.481 61.6905L419.463 61.6891C367.626 57.7073 286.27 51.4581 193.385 104.403C191.889 105.255 190.408 106.1 188.941 106.937C140.255 134.71 107.446 153.402 63.7633 153.827L63.7633 38.3871C63.7633 22.8275 51.591 10.1514 36.4999 10.1514C21.4089 10.1514 9.23659 22.8275 9.23659 38.3871L9.23657 473.613C9.23657 489.172 21.4088 501.849 36.4999 501.849C51.591 501.849 63.7633 489.172 63.7633 473.613ZM63.7633 277.827L63.7633 215.829C124.375 215.392 170.413 189.013 215.511 163.173L215.588 163.129C218.423 161.504 221.255 159.881 224.088 158.267C298.865 115.644 358.136 119.814 409.687 123.445L409.707 123.447C423.006 124.384 435.817 125.286 448.237 125.333V187.331C440.102 187.274 430.475 186.535 419.481 185.69L419.463 185.689C367.626 181.707 286.27 175.458 193.385 228.403C191.889 229.255 190.408 230.1 188.941 230.937C140.255 258.71 107.446 277.402 63.7633 277.827ZM215.511 338.813C199.972 329.909 184.321 320.941 168 312.993C184.321 305.045 199.972 296.077 215.511 287.173L215.588 287.129L215.628 287.105C218.451 285.488 221.269 283.873 224.088 282.267C298.865 239.644 358.136 243.814 409.687 247.445L409.709 247.447C423.007 248.384 435.818 249.286 448.237 249.333L448.237 376.653C435.818 376.7 423.007 377.602 409.709 378.539L409.687 378.541C358.136 382.172 298.865 386.342 224.088 343.719C221.269 342.112 218.451 340.498 215.628 338.881L215.588 338.857L215.511 338.813Z',
+  ],
+};
+export const IconSankey = createFontAwesomeIcon(dvSankey);
