+++ src/icons/font-awesome/icon-scatter-plot.ts
@@ -1,0 +1,20 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvScatterplot: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'scatterplot' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M480.18 88C480.18 110.443 462.048 128.5 439.68 128.5C417.312 128.5 399.18 110.443 399.18 88C399.18 65.5573 417.312 47.5 439.68 47.5C462.048 47.5 480.18 65.5573 480.18 88ZM333 129C333 151.445 314.866 169.637 292.5 169.637C270.134 169.637 252 151.445 252 129C252 106.555 270.134 88.3629 292.5 88.3629C314.866 88.3629 333 106.555 333 129ZM206.3 362.134C206.3 384.579 188.166 402.771 165.8 402.771C143.434 402.771 125.3 384.579 125.3 362.134C125.3 339.689 143.434 321.497 165.8 321.497C188.166 321.497 206.3 339.689 206.3 362.134ZM232 212.137C232 234.582 213.866 252.774 191.5 252.774C169.134 252.774 151 234.582 151 212.137C151 189.692 169.134 171.5 191.5 171.5C213.866 171.5 232 189.692 232 212.137ZM351.5 320.997C351.5 343.442 333.366 361.634 311 361.634C288.634 361.634 270.5 343.442 270.5 320.997C270.5 298.552 288.634 280.36 311 280.36C333.366 280.36 351.5 298.552 351.5 320.997ZM439.18 212.137C439.18 234.582 421.046 252.774 398.68 252.774C376.314 252.774 358.18 234.582 358.18 212.137C358.18 189.692 376.314 171.5 398.68 171.5C421.046 171.5 439.18 189.692 439.18 212.137Z
+    M75.5717 450.288C75.5943 450.287 75.613 450.286 75.6274 450.286H474.861C489.325 450.286 501 461.666 501 475.643C501 489.62 489.325 501 474.861 501H61.6512V502C61.6689 501 61.669 501 61.669 501H61.6685L61.6656 501L61.6524 501L61.5966 500.999C61.5467 500.997 61.4713 500.995 61.3717 500.992C61.1727 500.986 60.8772 500.975 60.4958 500.957C59.7328 500.921 58.6265 500.857 57.2604 500.742C54.5269 500.514 50.7607 500.086 46.6294 499.293C38.3109 497.695 28.7333 494.653 23.0055 488.95C17.1644 483.135 14.1301 473.311 12.5789 464.759C11.8086 460.512 11.4135 456.636 11.2111 453.822C11.11 452.415 11.0572 451.276 11.0297 450.49C11.0159 450.096 11.0085 449.792 11.0045 449.587C11.0025 449.484 11.0013 449.406 11.0007 449.355L11.0001 449.297L11 449.284L11 449.283V37.2357C11 22.7116 22.3447 11 36.2633 11C50.1819 11 61.5267 22.7116 61.5267 37.2357V436.155C61.5259 436.169 61.5251 436.189 61.5241 436.212C61.522 436.266 61.5196 436.342 61.5179 436.439C61.5145 436.632 61.5141 436.907 61.5254 437.247C61.5481 437.926 61.6178 438.87 61.8068 439.934C62.1804 442.04 63.0416 444.741 65.0514 446.759C67.064 448.78 69.7586 449.642 71.8591 450.013C72.921 450.201 73.8619 450.269 74.5388 450.29C74.878 450.3 75.1528 450.299 75.3456 450.295C75.4421 450.293 75.5182 450.291 75.5717 450.288Z
+    `,
+  ],
+};
+
+export const IconScatterPlot = createFontAwesomeIcon(dvScatterplot);
