+++ src/icons/font-awesome/icon-sort-asc.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvSortAsc: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'sortAsc' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M250.622 110.314C261.557 121.248 261.557 138.977 250.622 149.912C239.688 160.846 221.959 160.846 211.024 149.912L157.421 96.308V443.711C157.421 459.175 144.886 471.711 129.421 471.711C113.957 471.711 101.421 459.175 101.421 443.711V96.2895L47.7988 149.912C36.8643 160.846 19.1357 160.846 8.20117 149.912C-2.7334 138.977 -2.7334 121.248 8.20117 110.314L105.577 12.9375C106.61 11.3647 107.818 9.87315 109.201 8.49028C114.773 2.91771 122.111 0.185044 129.414 0.292283C136.716 0.186143 144.051 2.91881 149.622 8.49034C151.005 9.87322 152.213 11.3647 153.246 12.9374L250.622 110.314ZM482 451C498.568 451 512 437.569 512 421C512 404.431 498.568 391 482 391H222C205.432 391 192 404.431 192 421C192 437.569 205.432 451 222 451H482ZM482 265.586C498.568 265.586 512 252.154 512 235.585C512 219.017 498.568 205.585 482 205.585H222C205.432 205.585 192 219.017 192 235.585C192 252.154 205.432 265.586 222 265.586H482ZM482 81C498.568 81 512 67.5685 512 51C512 34.4314 498.568 21 482 21H282C265.432 21 252 34.4314 252 51C252 67.5685 265.432 81 282 81H482Z',
+  ],
+};
+export const IconSortAsc = createFontAwesomeIcon(dvSortAsc);
