+++ src/icons/font-awesome/icon-sort-desc.ts
@@ -1,0 +1,16 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvSortDesc: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'sortDesc' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    'M157.421 28.2893C157.421 12.8253 144.885 0.289307 129.421 0.289307C113.957 0.289307 101.421 12.8253 101.421 28.2893V375.711L47.7988 322.088C36.8641 311.154 19.1355 311.154 8.20087 322.088C-2.73383 333.023 -2.73383 350.752 8.20087 361.686L105.576 459.062C106.61 460.635 107.818 462.127 109.201 463.51C114.773 469.082 122.111 471.815 129.414 471.708C136.715 471.814 144.051 469.081 149.622 463.51C151.005 462.127 152.213 460.635 153.246 459.062L250.622 361.686C261.557 350.752 261.557 333.023 250.622 322.088C239.688 311.154 221.959 311.154 211.024 322.088L157.421 375.691V28.2893ZM482 21C498.569 21 512 34.4315 512 51C512 67.5685 498.569 81 482 81H222C205.431 81 192 67.5685 192 51C192 34.4315 205.431 21 222 21H482ZM482 206.415C498.569 206.415 512 219.846 512 236.414C512 252.983 498.569 266.414 482 266.414H222C205.431 266.414 192 252.983 192 236.414C192 219.846 205.431 206.415 222 206.415H482ZM482 391C498.569 391 512 404.431 512 421C512 437.569 498.569 451 482 451H282C265.431 451 252 437.569 252 421C252 404.431 265.431 391 282 391H482Z',
+  ],
+};
+export const IconSortDesc = createFontAwesomeIcon(dvSortDesc);
