+++ src/icons/font-awesome/icon-string-array-column.ts
@@ -1,0 +1,21 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvStringArrayCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'stringArrayCol' as IconName,
+  icon: [
+    512,
+    307,
+    [],
+    '',
+    `M0 307.5V0.5H78.4862V33.2467H38.5546V274.753H78.4862V307.5H0Z
+    M51.0333 246.998L120.156 58.7044H162.566L231.689 246.998H184.873L168.9 199.515L192.309 209.885H89.3125L111.895 201.698L96.7481 246.998H51.0333ZM114.373 193.784L100.879 175.228H180.742L167.248 193.784L134.477 96.636H147.144L114.373 193.784Z
+    M234.512 297.482L247.18 248.908C242.59 247.453 238.643 244.724 235.338 240.721C232.034 236.537 230.381 231.352 230.381 225.167C230.381 217.89 232.768 211.886 237.541 207.156C242.498 202.426 248.741 200.061 256.268 200.061C263.795 200.061 269.946 202.517 274.719 207.429C279.676 212.159 282.155 218.344 282.155 225.985C282.155 231.625 281.42 236.901 279.951 241.813C278.483 246.725 276.28 252.91 273.342 260.369L258.746 297.482H234.512Z
+    M304.274 246.998V58.7044H375.049C387.166 58.7044 397.999 60.7966 407.545 64.9809C417.276 68.9833 424.895 74.7139 430.403 82.1729C436.094 89.6319 438.94 98.6372 438.94 109.189C438.94 118.103 436.92 125.835 432.881 132.384C428.842 138.752 423.334 143.937 416.358 147.939C426.088 152.305 433.616 158.582 438.94 166.768C444.448 174.773 447.202 184.233 447.202 195.149C447.202 205.883 444.264 215.161 438.389 222.984C432.698 230.806 424.895 236.81 414.981 240.994C405.25 244.997 394.235 246.998 381.934 246.998H304.274ZM348.336 210.431H377.803C385.147 210.431 391.022 208.702 395.428 205.246C400.018 201.607 402.313 196.331 402.313 189.418C402.313 182.505 400.018 177.32 395.428 173.864C391.022 170.225 385.147 168.406 377.803 168.406H348.336V210.431ZM348.336 134.568H372.295C379.272 134.568 384.872 132.93 389.094 129.656C393.317 126.199 395.428 121.287 395.428 114.92C395.428 108.552 393.317 103.731 389.094 100.456C384.872 96.9998 379.272 95.2715 372.295 95.2715H348.336V134.568Z
+    M433.514 307.5V274.753H473.445V33.2467H433.514V0.5H512V307.5H433.514Z`,
+  ],
+};
+
+export const IconStringArrayColumn = createFontAwesomeIcon(dvStringArrayCol);
