+++ src/icons/font-awesome/icon-string-column.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvStringCol: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'stringCol' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+M403.178 293.855L403.182 293.863C407.692 303.971 413.72 311.585 421.28 316.669C428.829 321.746 437.139 324.286 446.199 324.286C453.925 324.286 461.095 322.633 467.7 319.323C474.297 316.108 480.444 311.85 486.141 306.553L511.353 346.295C501.963 354.857 492.219 361.388 482.12 365.899C471.924 370.235 460.513 372.407 447.879 372.407C435.071 372.407 422.822 369.803 411.127 364.593C399.621 359.166 389.41 351.459 380.494 341.467C371.767 331.479 364.797 319.31 359.589 304.952C354.572 290.606 352.058 274.291 352.058 256C352.058 237.708 354.572 221.394 359.589 207.048C364.797 192.69 371.767 180.522 380.493 170.533C389.41 160.539 399.528 152.943 410.847 147.734L410.852 147.732C422.359 142.306 434.512 139.593 447.319 139.593C459.398 139.593 470.442 141.549 480.458 145.453L480.458 145.453L480.47 145.458C490.56 149.105 500.386 154.664 509.949 162.141L486.097 203.183C480.404 198.736 474.265 195.122 467.681 192.34C460.895 189.255 453.639 187.714 445.919 187.714C436.86 187.714 428.548 190.254 421 195.331L420.995 195.334C413.623 200.419 407.69 208.032 403.182 218.137L403.179 218.144C398.851 228.268 396.698 240.893 396.698 256C396.698 271.107 398.851 283.732 403.178 293.855ZM113.033 143.523L183.004 368.477H136.46L123.941 324.305L123.838 323.941H123.46H59.1023H58.7244L58.6213 324.305L46.1025 368.477H0.679148L70.6494 143.523H113.033ZM91.7623 210.767L91.2812 209.07L90.8002 210.767L70.4088 282.716L70.2284 283.353H70.8898H111.673H112.334L112.154 282.716L91.7623 210.767ZM194.052 368.477V143.523H265.513C277.576 143.523 288.51 146.019 298.328 151.001L298.336 151.004C308.149 155.768 315.823 162.582 321.371 171.451L321.375 171.457C327.1 180.313 329.974 191.02 329.974 203.605C329.974 216.19 327.1 226.786 321.378 235.42L321.377 235.422C317.606 241.155 312.853 245.939 307.114 249.774L306.42 250.238L307.157 250.631C316.891 255.822 324.516 263.258 330.04 272.948L330.042 272.953C335.59 282.469 338.374 293.731 338.374 306.758C338.374 319.778 335.406 330.925 329.492 340.219L329.491 340.22C323.756 349.308 315.894 356.343 305.893 361.324C296.077 366.089 284.954 368.477 272.514 368.477H194.052ZM307.86 249.873C307.783 249.831 307.705 249.79 307.628 249.748L307.392 250.189L307.67 250.605C307.743 250.556 307.817 250.507 307.89 250.458C317.47 255.681 324.998 263.095 330.474 272.701C336.074 282.306 338.874 293.659 338.874 306.758C338.874 319.856 335.888 331.1 329.914 340.487L307.86 249.873ZM238.353 186.404H237.853V186.904V234.059V234.559H238.353H262.713C269.903 234.559 275.726 232.565 280.131 228.534L280.141 228.525C284.558 224.257 286.734 218.218 286.734 210.481C286.734 202.749 284.56 196.813 280.136 192.762C275.733 188.51 269.908 186.404 262.713 186.404H238.353ZM238.353 274.166H237.853V274.666V325.096V325.596H238.353H268.313C275.877 325.596 281.982 323.493 286.573 319.241L286.575 319.239C291.371 314.753 293.734 308.271 293.734 299.881C293.734 291.496 291.375 285.12 286.574 280.849C281.984 276.38 275.879 274.166 268.313 274.166H238.353Z
+    `,
+  ],
+};
+
+export const IconStringColumn = createFontAwesomeIcon(dvStringCol);
