+++ src/icons/font-awesome/icon-tissue.ts
@@ -1,0 +1,19 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvTissue: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'tissue' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M236.3 61.4L226.9 75.5C205.6 107.5 167.5 124 129.6 117.6L70.0002 107.7C33.3002 101.6 -0.099776 129.9 0.000223981 167.1C0.000223981 183 6.40022 198.3 17.6002 209.6L46.8002 238.8C57.8002 249.8 64.0002 264.7 64.0002 280.3C64.0002 296.1 57.6002 311.2 46.3002 322.3L33.2002 335.1C22.1002 345.9 15.9002 360.7 15.9002 376.2C15.9002 413 50.0002 440.4 86.0002 432.4L148.3 418.6C156 416.9 164 416 171.9 416H181.9C209.1 416 235.6 425.3 256.9 442.3L287.7 467C298.2 475.4 311.3 480 324.7 480C357.4 480 384 453.5 384 420.7V395.5C384 360.6 405.4 329.3 437.9 316.7L474.8 302.4C497.2 293.7 512 272.1 512 248.1C512 220 491.9 195.8 464.2 190.8L436.2 185.7C399.7 179 370.8 151.2 362.6 115L355.5 83.5C348.8 53.4 322 32 291.2 32C269.2 32 248.6 43 236.3 61.4ZM159.9 192C168.387 192 176.526 195.371 182.528 201.373C188.529 207.374 191.9 215.513 191.9 224C191.9 232.487 188.529 240.626 182.528 246.627C176.526 252.628 168.387 256 159.9 256C151.413 256 143.274 252.628 137.273 246.627C131.272 240.626 127.9 232.487 127.9 224C127.9 215.513 131.272 207.374 137.273 201.373C143.274 195.371 151.413 192 159.9 192ZM287.9 208C287.9 203.756 289.586 199.687 292.586 196.686C295.587 193.686 299.657 192 303.9 192C308.144 192 312.213 193.686 315.214 196.686C318.214 199.687 319.9 203.756 319.9 208C319.9 212.243 318.214 216.313 315.214 219.314C312.213 222.314 308.144 224 303.9 224C299.657 224 295.587 222.314 292.586 219.314C289.586 216.313 287.9 212.243 287.9 208ZM287.9 288C296.387 288 304.526 291.371 310.528 297.372C316.529 303.374 319.9 311.513 319.9 320C319.9 328.487 316.529 336.626 310.528 342.627C304.526 348.628 296.387 352 287.9 352C279.413 352 271.274 348.628 265.273 342.627C259.272 336.626 255.9 328.487 255.9 320C255.9 311.513 259.272 303.374 265.273 297.372C271.274 291.371 279.413 288 287.9 288Z
+    `,
+  ],
+};
+
+export const IconTissue = createFontAwesomeIcon(dvTissue);
