+++ src/icons/font-awesome/icon-violin.ts
@@ -1,0 +1,21 @@
+import { IconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-common-types';
+
+import { createFontAwesomeIcon } from '../factories/create-font-awesome-icon';
+
+export const dvViolin: IconDefinition = {
+  prefix: 'dv' as IconPrefix,
+  iconName: 'violin' as IconName,
+  icon: [
+    512,
+    512,
+    [],
+    '',
+    `
+    M61.6512 502V501H474.861C489.325 501 501 489.62 501 475.643C501 461.666 489.325 450.286 474.861 450.286H75.6274C75.613 450.286 75.5944 450.287 75.5717 450.288C75.5182 450.291 75.4421 450.293 75.3456 450.295C75.1528 450.299 74.878 450.3 74.5389 450.29C73.8619 450.269 72.921 450.201 71.8591 450.013C69.7586 449.642 67.0641 448.78 65.0514 446.759C63.0416 444.741 62.1804 442.04 61.8068 439.934C61.6179 438.87 61.5481 437.926 61.5255 437.247C61.5141 436.907 61.5145 436.632 61.5179 436.439C61.5196 436.342 61.522 436.266 61.5242 436.212C61.5251 436.189 61.526 436.169 61.5267 436.155V37.2357C61.5267 22.7116 50.1819 11 36.2633 11C22.3447 11 11 22.7116 11 37.2357V449.283L11 449.284L11.0001 449.297L11.0007 449.355C11.0014 449.406 11.0025 449.484 11.0045 449.587C11.0085 449.792 11.0159 450.096 11.0297 450.49C11.0572 451.276 11.1101 452.415 11.2112 453.822C11.4135 456.636 11.8086 460.512 12.5789 464.759C14.1302 473.311 17.1644 483.135 23.0055 488.95C28.7333 494.653 38.3109 497.695 46.6294 499.293C50.7608 500.086 54.5269 500.514 57.2604 500.742C58.6265 500.857 59.7329 500.921 60.4958 500.957C60.8773 500.975 61.1727 500.986 61.3718 500.992C61.4713 500.995 61.5467 500.997 61.5967 500.999L61.6524 501L61.6656 501L61.6685 501H61.6691C61.6691 501 61.669 501 61.6512 502ZM61.6512 502H474.861C489.85 502 502 490.199 502 475.643C502 461.086 489.85 449.286 474.861 449.286H75.6L61.6512 502Z
+    M389 24L385.539 30.3361L381.565 38.3509C377.703 46.7222 363.025 67.0526 335.214 81.4035C318.476 90.9708 285 114.411 285 131.632C285 140.003 288.863 159.257 304.313 169.303C321.051 181.86 354.527 212.355 354.527 233.882V262.583C353.883 267.367 351.823 279.446 348.733 289.491C346.158 299.656 341.008 321.781 341.008 328.956C341.652 340.317 345.257 365.551 354.527 375.596C363.54 385.762 381.565 409.321 381.565 422.237L384.481 427.401L389 433L393.523 427.401L396.439 422.237C396.439 409.321 414.464 385.762 423.476 375.596C432.746 365.551 436.351 340.317 436.995 328.956C436.995 321.781 431.845 299.656 429.27 289.491C426.18 279.446 424.12 267.367 423.476 262.583V233.882C423.476 212.355 456.951 181.86 473.688 169.303C489.138 159.257 493 140.003 493 131.632C493 114.411 459.526 90.9708 442.788 81.4035C414.979 67.0526 400.302 46.7222 396.439 38.3509L392.465 30.3361L389 24Z
+    M174.2 77H175.572H176.428H177.8C177.8 77 178.139 81.6353 181.916 86.6102C185.692 91.585 191.829 111.603 199.529 116.965C210.652 123.589 240.599 142.426 240.599 161.211V175.485C240.599 182.668 244.877 191.23 248.299 198.322C250.866 204.507 256 218.875 256 226.868C256 236.383 251.926 256.896 240.599 268.26C229.476 279.419 206.969 298.255 204.663 311.079C202.096 325.352 189.262 366.752 187.551 373.601C185.84 380.451 184.128 382.164 181.916 385.898C178.717 391.296 176.428 401 176.428 401H175.572C175.572 401 173.283 391.296 170.084 385.898C167.872 382.164 166.16 380.451 164.449 373.601C162.738 366.752 149.904 325.352 147.337 311.079C145.031 298.255 122.524 279.419 111.401 268.26C100.074 256.896 96 236.383 96 226.868C96 218.875 101.134 204.507 103.701 198.322C107.123 191.23 111.401 182.668 111.401 175.485V161.211C111.401 142.426 141.348 123.589 152.471 116.965C160.171 111.603 166.308 91.585 170.084 86.6102C173.861 81.6353 174.2 77 174.2 77Z
+    `,
+  ],
+};
+
+export const IconViolin = createFontAwesomeIcon(dvViolin);
