+++ src/icons/index.ts
@@ -1,2 +1,86 @@
-export * from './EntityIcons';
-export * from './Icons';
+// Lucide re-exports
+export { IconLasso } from './lucide/icon-lasso';
+
+// Font Awesome re-exports
+
+// Our own icons
+export { IconAddVisualization } from './font-awesome/icon-add-visualization';
+export { IconBooleanArrayColumn } from './font-awesome/icon-boolean-array-column';
+export { IconBoxplot } from './font-awesome/icon-boxplot';
+export { IconBrushSelection } from './font-awesome/icon-brush-selection';
+export { IconCategoricalArrayColumn } from './font-awesome/icon-categorical-array-column';
+export { IconChemicalStructure } from './font-awesome/icon-chemical-structure';
+export { IconClearSelection } from './font-awesome/icon-clear-selection';
+export { IconCorrelationPlot } from './font-awesome/icon-correlation-plot';
+export { IconDownloadVisualization } from './font-awesome/icon-download-visualization';
+export { IconDrug } from './font-awesome/icon-drug';
+export { IconFilterClear } from './font-awesome/icon-filter-clear';
+export { IconFilterOutline } from './font-awesome/icon-filter-outline';
+export { IconHeatmap } from './font-awesome/icon-heatmap';
+export { IconHexbinPlot as IconHexbinplot } from './font-awesome/icon-hexbin-plot';
+export { IconJunction } from './font-awesome/icon-junction';
+export { IconMouse } from './font-awesome/icon-mouse';
+export { IconNumberArrayColumn } from './font-awesome/icon-number-array-column';
+export { IconNumericColumn } from './font-awesome/icon-numeric-column';
+export { IconOncoprint } from './font-awesome/icon-oncoprint';
+export { IconProteinBackbone } from './font-awesome/icon-protein-backbone';
+export { IconProteinBallAndStick } from './font-awesome/icon-protein-ball-and-stick';
+export { IconProteinCartoon } from './font-awesome/icon-protein-cartoon';
+export { IconProteinStick } from './font-awesome/icon-protein-stick';
+export { IconProteinSurface } from './font-awesome/icon-protein-surface';
+export { IconSankey } from './font-awesome/icon-sankey';
+export { IconScatterPlot as IconScatterplot } from './font-awesome/icon-scatter-plot';
+export { IconSortAsc } from './font-awesome/icon-sort-asc';
+export { IconSortDesc } from './font-awesome/icon-sort-desc';
+export { IconSort } from './font-awesome/icon-sort';
+export { IconStringArrayColumn } from './font-awesome/icon-string-array-column';
+export { IconStringColumn } from './font-awesome/icon-string-column';
+export { IconViolin } from './font-awesome/icon-violin';
+
+// Entity icons
+export { IconCellLine } from './font-awesome/icon-cell-line';
+export { IconDisease } from './font-awesome/icon-disease';
+export { IconGene } from './font-awesome/icon-gene';
+export { IconTissue } from './font-awesome/icon-tissue';
+
+// Backwards compatibility
+// Re-exporting the IconDefinitions for legacy usage
+// @TODO: Remove as soon as we migrated these items to the new icon system
+export { dvAddVisualization } from './font-awesome/icon-add-visualization';
+export { dvBooleanArrayCol } from './font-awesome/icon-boolean-array-column';
+export { dvBoxplot } from './font-awesome/icon-boxplot';
+export { dvBrushSelection } from './font-awesome/icon-brush-selection';
+export { dvCategoricalArrayCol } from './font-awesome/icon-categorical-array-column';
+export { dvChemicalStructure } from './font-awesome/icon-chemical-structure';
+export { dvClearSelection } from './font-awesome/icon-clear-selection';
+export { dvCorrelationplot } from './font-awesome/icon-correlation-plot';
+export { dvDownloadVisualization } from './font-awesome/icon-download-visualization';
+export { dvDrug } from './font-awesome/icon-drug';
+export { dvFilterClear } from './font-awesome/icon-filter-clear';
+export { dvFilterOutline } from './font-awesome/icon-filter-outline';
+export { dvHeatmap } from './font-awesome/icon-heatmap';
+export { dvHexbinplot } from './font-awesome/icon-hexbin-plot';
+export { dvJunction } from './font-awesome/icon-junction';
+export { dvMouse } from './font-awesome/icon-mouse';
+export { dvNumberArrayCol } from './font-awesome/icon-number-array-column';
+export { dvNumericCol } from './font-awesome/icon-numeric-column';
+export { dvOncoprint } from './font-awesome/icon-oncoprint';
+export { dvProteinBackbone } from './font-awesome/icon-protein-backbone';
+export { dvProteinBallAndStick } from './font-awesome/icon-protein-ball-and-stick';
+export { dvProteinCartoon } from './font-awesome/icon-protein-cartoon';
+export { dvProteinStick } from './font-awesome/icon-protein-stick';
+export { dvProteinSurface } from './font-awesome/icon-protein-surface';
+export { dvSankey } from './font-awesome/icon-sankey';
+export { dvScatterplot } from './font-awesome/icon-scatter-plot';
+export { dvSortAsc } from './font-awesome/icon-sort-asc';
+export { dvSortDesc } from './font-awesome/icon-sort-desc';
+export { dvSort } from './font-awesome/icon-sort';
+export { dvStringArrayCol } from './font-awesome/icon-string-array-column';
+export { dvStringCol } from './font-awesome/icon-string-column';
+export { dvViolin } from './font-awesome/icon-violin';
+
+// Backwards compatibility for entity icons
+export { dvCellLine } from './font-awesome/icon-cell-line';
+export { dvDisease } from './font-awesome/icon-disease';
+export { dvGene } from './font-awesome/icon-gene';
+export { dvTissue } from './font-awesome/icon-tissue';
