+++ src/icons/lucide/icon-lasso.ts
@@ -1,0 +1,5 @@
+import { Lasso } from 'lucide-react';
+
+import { createLucideIcon } from '../factories/create-lucide-icon';
+
+export const IconLasso = createLucideIcon(Lasso);
