+++ src/icons/types.ts
@@ -1,0 +1,6 @@
+import { MantineSize } from '@mantine/core';
+
+export interface VisynIconProps {
+  size?: MantineSize | number;
+  // className?: string;
+}
